% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_fuel.R
\name{absolute_fuel_constraints}
\alias{absolute_fuel_constraints}
\title{Create either a minimum or maximum fuel constraint}
\usage{
absolute_fuel_constraints(data, decision_variables, constraint_type)
}
\arguments{
\item{data}{list of tables generated by preprocessing the raw inputs from the
input spreadsheet using \code{comit_preprocess_inputs()}}

\item{decision_variables}{dataframe for the decision variables. This is the
first element of the list produced by \code{comit_objective_function}.}

\item{constraint_type}{character, either 'max' or 'min' depending on whether
the limit to be set is on maximum or minimum fuel use.}
}
\value{
list of constraints. One constraint for each fuel specified to be
constrained in the input, in each time point. Each element contains a
nested list with 4 elements:
\itemize{
\item column indices
\item values
\item direction
\item rhs
}
}
\description{
Perform all manipulations required to the decision variables to get the data
required to set up the constraints. The final step then formulates the constraint
itself.
}
