% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_capacities.R
\name{add_alternative_technologies_data}
\alias{add_alternative_technologies_data}
\title{Get all alternative technologies that can be deployed instead/as well as the
given technologies}
\usage{
add_alternative_technologies_data(
  capacity_variables,
  data,
  decision_variables,
  capacity_tab
)
}
\arguments{
\item{capacity_variables}{dataframe for decision variables that are relevant
to the capacity constraints, as produced by \code{get_capacity_variables_data}.}

\item{data}{list of tables generated by preprocessing the raw inputs from the
input spreadsheet using \code{comit_preprocess_inputs()}}

\item{decision_variables}{dataframe for the decision variables. This is the
first element of the list produced by \code{comit_objective_function}.}

\item{capacity_tab}{dataframe, constaining data on the technologies to constrain
as produced by \code{get_technology_capacity_tab}.}
}
\value{
dataframe for capacity variable, now with one row per feasible
combination of decision variables with alternative technology decision
variables.
}
\description{
We need to constraint the \% of a commodity that a technology produces in each
year, therefore we need to join on the alternative technologies so that they
can be used to formulate the constraint.
}
