% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_cost_tables.R
\name{add_capex_adjustment}
\alias{add_capex_adjustment}
\title{Renames the PV_term label: capex for new_capacity, opex for
available_capacity variables, capex_lump for capex_lump}
\usage{
add_capex_adjustment(costs_capex, model_data, capex_type)
}
\arguments{
\item{costs_capex}{a table containing the capex cost data}

\item{model_data}{list of data tables read in from excel data template}

\item{capex_type}{string indicating if this is a technology or pipe cost dataset}
}
\value{
table of capex cost data with adjusted (spread) capex costs incorporated
}
\description{
Renames the PV_term label: capex for new_capacity, opex for
available_capacity variables, capex_lump for capex_lump
}
