% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_production.R
\name{add_demand_to_production_data}
\alias{add_demand_to_production_data}
\title{Add rows for demand data to the production data}
\usage{
add_demand_to_production_data(production_df, data)
}
\arguments{
\item{production_df}{dataframe of decision variables relevant to the production
constraints as produced by \code{initiate_production_constraint_data()}.}

\item{data}{list of tables generated by preprocessing the raw inputs from the
input spreadsheet using \code{comit_preprocess_inputs()}}
}
\value{
dataframe as input as production_df, but with rows added for the
demand of each commodity that each site is required to produce in each
year. An additional column has also been introduced, flagging whether the
output_commodity is a final (end demand) commodity or not.
}
\description{
Add a row to the production data for the demand of each commodity at each
site and in each year, taken from the site_demand data frame.
}
\details{
Note that demand rows are distinguishable from decision variables by not
having an variable_index number (the value will be NA instead).
}
