% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_DB.R
\name{add_extra_metadata}
\alias{add_extra_metadata}
\title{Insert location and sequencing information}
\usage{
add_extra_metadata(DB, metadata)
}
\arguments{
\item{DB}{filepath of CoMIT database}

\item{metadata}{filepath of metadata}
}
\description{
Insert location and sequencing information
}
