% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_emissions_tables.R
\name{add_fuel_emissions_for_H2}
\alias{add_fuel_emissions_for_H2}
\title{Calculates the fuel emissions for hydrogen use in the model, according to
the mix of blue/green/grey hydrogen that are deployed in the optimal solution}
\usage{
add_fuel_emissions_for_H2(output_data, model_data)
}
\arguments{
\item{output_data}{a table containing a subset of the model solution outputs}

\item{model_data}{list of data tables read in from excel data template}
}
\value{
table of fuel emissions factors with rows for hydrogen use in each
year of the model added
}
\description{
Calculates the fuel emissions for hydrogen use in the model, according to
the mix of blue/green/grey hydrogen that are deployed in the optimal solution
}
