% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_emissions_tables.R
\name{add_ghg_splits}
\alias{add_ghg_splits}
\title{Adds the Greenhouse gas splits to the table of direct emissions
using the information in the ghg_splits tab of the Excel input workbook
Only the splits for CO2, CH4 and N20 are used}
\usage{
add_ghg_splits(emissions_data, model_data)
}
\arguments{
\item{emissions_data}{a table containing the (direct) emissions data}

\item{model_data}{list of data tables read in from excel data template}
}
\value{
table of (direct) emissions with the following columns added:
\itemize{
\item ghg_type (either CO2, CH4, N20 or total_CO2e)
\item CO2_NonCO2 (either CO2 or NonCO2)
\item years - one column for each timestep in the model
}
}
\description{
Adds the Greenhouse gas splits to the table of direct emissions
using the information in the ghg_splits tab of the Excel input workbook
Only the splits for CO2, CH4 and N20 are used
}
