% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_hydrogen_modelling_adjustments.R
\name{add_H2_plants}
\alias{add_H2_plants}
\title{Adds H2 production plants at clusters if required}
\usage{
add_H2_plants(data, model_H2_production)
}
\arguments{
\item{data}{data read in from excel template}

\item{model_H2_production}{logical TRUE/FALSE depending on whether H2 production is modeled explicitly}
}
\value{
list of data
}
\description{
Adds H2 production plants at clusters if required
}
