% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_cost_tables.R
\name{add_pipe_capex_lump}
\alias{add_pipe_capex_lump}
\title{Adds rows for capex_lump costs ('lumpy' yearly investment cost figures  without
borrowing, spreading or  discounting) for pipes to a table of pipe costs}
\usage{
add_pipe_capex_lump(capacity_variable, cost_data, model_data, pipes_lifetime)
}
\arguments{
\item{capacity_variable}{variable type to filter for (H2 or CO2 pipes)}

\item{cost_data}{a table containing the pipe cost data}

\item{model_data}{list of data tables read in from excel data template}

\item{pipes_lifetime}{pipe lifetime}
}
\value{
table of cost data with capex_lump rows added
}
\description{
Adds rows for capex_lump costs ('lumpy' yearly investment cost figures  without
borrowing, spreading or  discounting) for pipes to a table of pipe costs
}
