% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_cost_tables.R
\name{add_PVterm_labels}
\alias{add_PVterm_labels}
\title{Renames the PV_term label: capex for new_capacity, opex for
available_capacity variables, capex_lump for capex_lump}
\usage{
add_PVterm_labels(cost_data, capacity_type)
}
\arguments{
\item{cost_data}{a table containing the pipe cost data}
}
\value{
table of cost data with updated PV_term labels
}
\description{
Renames the PV_term label: capex for new_capacity, opex for
available_capacity variables, capex_lump for capex_lump
}
