% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_sites.R
\name{add_site_ID}
\alias{add_site_ID}
\title{Add a unique id number for each site in the site data table}
\usage{
add_site_ID(data)
}
\arguments{
\item{data}{list of data tables as read in from excel template and having
been passed to \code{process_sites()}}
}
\value{
list of data tables as input but with site_ID column added to the
NAEI_clean data frame.
}
\description{
Add a unique id number for each site in the site data table
}
