% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_capacity_transfer.R
\name{add_techs_to_capacity_transfer}
\alias{add_techs_to_capacity_transfer}
\title{Add technology and site demand data to get the required data to set up
capacity transfer constraint}
\usage{
add_techs_to_capacity_transfer(
  capacity_transfer_constraint,
  temp_techs,
  site_demand,
  data
)
}
\arguments{
\item{capacity_transfer_constraint}{\itemize{
\item data.table created by \code{initiate_capacity_transfer_dt}
}}

\item{temp_techs}{\itemize{
\item data.table for technology and pipe lifetime data
}}

\item{site_demand}{\itemize{
\item data.table for site_demand
}}

\item{data}{list of tables generated by preprocessing the raw inputs from the
input spreadsheet using \code{comit_preprocess_inputs()}}
}
\value{
data.table containing four columns:
\itemize{
\item variable_index.available - index of decision variables for available capacity
\item variable_index.new - index of decision variables for new capacity
\item residual_capacity.available - remaining initial capacity
\item int - multiplier to adjust retrofit technologies
}
}
\description{
Joins and processes data from technology and site_demand data.tables and
selects the variables required to set up the constraint.
}
