% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F-2-classify_helpers.R
\name{addVariant2DB}
\alias{addVariant2DB}
\title{Insert new Variant into a CoMIT database and add to cache$primerVars}
\usage{
addVariant2DB(variant, currPrimerID)
}
\arguments{
\item{variant}{String of the variant DNA sequence}

\item{currPrimerID}{Database Id for the primer string is a variant of}
}
\value{
VarID number recieved from the database
}
\description{
Insert new Variant into a CoMIT database and add to cache$primerVars
}
