% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_starting_values.R
\name{adjust_capacity}
\alias{adjust_capacity}
\title{Amends the total capacity for technologies that produce a given commodity
when there isn't enough existing capacity}
\usage{
adjust_capacity(rows, key)
}
\arguments{
\item{rows}{subset of the original dataframe which belong to a given group.}

\item{key}{a single row of the dataframe belonging to the group}
}
\value{
rows as input but with the capacity updated if so required. If an
update is required a message will also be sent to the console to inform the
user.
}
\description{
Updates the total capacity for technologies when there isn't enough demand
for a given commodity. If there is already enough capacity, the original
values are returned. Otherwise the current capacities for the relevant
technologies are increased proportionally to the amount which means that
total demand is met.
}
\details{
Used inside \code{group_modify()} of \code{adjust_existing_capacity()}.
}
