% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_starting_values.R
\name{adjust_existing_capacity}
\alias{adjust_existing_capacity}
\title{Amend the capacity in the start year for existing technologies to ensure that
demand is met}
\usage{
adjust_existing_capacity(data)
}
\arguments{
\item{data}{list of dataframes, the input data produced from
\code{read_excel_data_template()} after some initial tidying, including by
\code{interpolate_data()}, \code{tidy()} and \code{round_years()}.}
}
\value{
list of dataframes as provided in the input, but with the Technologies
table updated with new \code{existing_capacity_2020()} values.
}
\description{
Finds the required amount of each technology already being deployed in 2020
that is needed to meet the demand of each commodity, then updates the
starting capacities accordingly. The necessary adjustments are made to
the Technologies dataframe and a message is sent to the console to notify
the user of any changes.
}
\details{
The calculation is done using linear algebra to solve a system of linear
equations.

An error is raised if there are TRUE values for objective function parameters
set in the 'objective_function' tab.
}
