% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_read_data.R
\name{adjust_for_optimism}
\alias{adjust_for_optimism}
\title{Apply a factor to all capex prices to account for underestimation in prices}
\usage{
adjust_for_optimism(data)
}
\arguments{
\item{data}{list of dataframes read in by \code{read_excel_data_template()}}
}
\value{
list of data tables as provided in inputs, with capex adjusted in the
relevant dataframes.
}
\description{
Accounts for optimism bias by multiplying capex prices by a set factor. The
factor is set in the input spreadsheet model parameter tabs. Both technology
and infrastructure prices are adjusted.
}
