% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_output_tables.R
\name{adjust_NAEI_data}
\alias{adjust_NAEI_data}
\title{Adds cluster radius to the NAEI data table, and creates the traded_site field}
\usage{
adjust_NAEI_data(NAEI_data)
}
\arguments{
\item{NAEI_data}{one of the data tables read in from excel template}
}
\value{
the NAEI_clean data as provided as in the input, with the following updates:
\itemize{
\item A column has been added to indicate if each site is within a 25km and 30km cluster radius
\item the traded_flag column is renamed to traded_site
}
}
\description{
Adds cluster radius to the NAEI data table, and creates the traded_site field
}
