% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_cost_tables.R
\name{aggregate_cost_table}
\alias{aggregate_cost_table}
\title{Calculates the technology costs incurred. Used to calculate the
non-infrastructure costs in the cost tabs of the Excel output workbook}
\usage{
aggregate_cost_table(cost_data, group_cols, Sector_infrastructure_label)
}
\arguments{
\item{cost_data}{a table containing the table of costs to aggregate over}

\item{group_cols}{list of variables to aggregate (summarise) over}

\item{Sector_infrastructure_label}{string to label the type of cost
("Sector" for technology costs or "H2_S2C","CO2_S2C", or "CO2_C2S" for infrastructure costs)}
}
\value{
table of aggregated costs
}
\description{
Calculates the technology costs incurred. Used to calculate the
non-infrastructure costs in the cost tabs of the Excel output workbook
}
