% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_fuel.R
\name{aggregate_fuel_constrained_variables}
\alias{aggregate_fuel_constrained_variables}
\title{Aggregate any fuel constraints which have more than one input from a commodity
group.}
\usage{
aggregate_fuel_constrained_variables(fuel_constrained_variables)
}
\arguments{
\item{fuel_constrained_variables}{dataframe containing all decision variables
to be constrained. Produced by \code{get_fuel_constrained_variables()} and
\code{apply_sector_filter_to_relevant_fuel_variables()}.}
}
\value{
dataframe with one row per variable/fuel constraint type combination.
There will often be no reduction in rows and often the input commodity
to variable/type combination ratio is 1:1 for all variables. Columns are:
\itemize{
\item year
\item variable_index
\item type
\item input
\item fuel_limit
}
}
\description{
Some commodity groups may mean that a technology has more than one input
from a commodity. This function ensures that there is only one row per
variable/type combination.
}
