% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_process_sites.R
\name{aggregate_small_point_sites}
\alias{aggregate_small_point_sites}
\title{Combines small point sites based on there characteristics and calculates
information for the sites at the aggregate level}
\usage{
aggregate_small_point_sites(df, input_data)
}
\arguments{
\item{df}{dataframe for small point sites produced by
\code{\link[=get_small_point_site_data]{get_small_point_site_data()}}}

\item{input_data}{data read in from excel template, same as what is passed to
\code{\link[=comit_solver]{comit_solver()}}.}
}
\value{
dataframe of the aggregated sites, with the following columns:
\itemize{
\item num_sites - the number of sites that were aggregated to form the new row
\item pipe_dist - the mean pipe distances from the aggregated sites
\item total_MtCO2 - the sum of the emissions from the aggregated sites
\item pipe_dist_category - the category for the pipe distance that the aggregated
site lies within.
}
}
\description{
The sites are grouped by \code{\link[=group_small_point_sites]{group_small_point_sites()}}. Calculations are then
made to combine the site level data into the aggregate level.
}
