% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_process_sites.R
\name{allocate_cluster_points}
\alias{allocate_cluster_points}
\title{Classify coordinates to closest cluster point}
\usage{
allocate_cluster_points(df, input_data)
}
\arguments{
\item{df}{dataframe containing the Longitude and Latitude for a site.}

\item{input_data}{data read in from excel template, same as what is passed to
\code{\link[=comit_solver]{comit_solver()}}.}
}
\value{
original dataframe with columns appended for cluster classification
and the coordinates of the cluster center.
}
\description{
For any set of coordinates find the closest cluster and it's respective
center point coordinates.
}
