% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_process_sites.R
\name{append_information_to_sites}
\alias{append_information_to_sites}
\title{Add required variables to site data}
\usage{
append_information_to_sites(input_data)
}
\arguments{
\item{input_data}{data read in from excel template, same as what is passed to
\code{\link[=comit_solver]{comit_solver()}}.}
}
\value{
dataframe of NAEI data with additional columns added including
\itemize{
\item H2 grid connection year
\item Longitude
\item Latitude
\item H2_point (cluster)
\item Lat_clust (cluster center coordinates)
\item Lon_clust (cluster center coordinates)
\item pipe_dist (pipe distance to cluster center in kilometers)
\item emissions_MtCO2 (site emissions in Megatons)
\item sector_share_by_traded (proportion of sector by traded_status emissions, see
\code{\link{calculate_share_of_sector_emissions}} for more information).
}
}
\description{
Appends columns to the NAEI data, including new sector mapping, longitude
and latitude, pipe distances and emissions variables.
}
