% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_starting_values.R
\name{apply_energy_efficiency}
\alias{apply_energy_efficiency}
\title{Applies resource efficiency and energy efficiency measures to demand figures}
\usage{
apply_energy_efficiency(data)
}
\arguments{
\item{data}{list of all dataframes with some initial tidying having been
completed.}
}
\value{
list of all dataframes as provided as inputs, but with adjustments
for resource and energy efficiency made to demand in the Demand_drivers
table.
}
\description{
Adjust the demand values by the estimates for resource efficiency and energy
efficiency in each year, for each commodity.
}
