% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_fuel.R
\name{apply_sector_filter_to_relevant_fuel_variables}
\alias{apply_sector_filter_to_relevant_fuel_variables}
\title{Filter out hydrogen and refinery variables when required}
\usage{
apply_sector_filter_to_relevant_fuel_variables(
  data,
  fuel_constrained_variables
)
}
\arguments{
\item{data}{list of tables generated by preprocessing the raw inputs from the
input spreadsheet using \code{comit_preprocess_inputs()}}

\item{fuel_constrained_variables}{dataframe for the variables that are to be
constrained, produced by \code{get_fuel_constrained_variables()}.}
}
\value{
dataframe as provided with filter applied.
}
\description{
Joins on the sector to the variables that are to be constrained, so that
when apply_to_industry only is TRUE, the hydrogen and refineries sector
variables can be removed. No filter is applied to rows where either the
sector is not hydrogen or refinery, or the apply_to_industry_only variable
is set to FALSE.
}
