% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_geography.R
\name{assign_site_region}
\alias{assign_site_region}
\title{Get site regions}
\usage{
assign_site_region(data)
}
\arguments{
\item{data}{dataframe of site of data, containing columns: site_ID, site_name,
Latitude and Longitude.}
}
\value{
The original dataframe with region column appended.
}
\description{
Assign region based on site long and lat. This allows for dis-aggregation of
regions for the sites in the outputs.
}
