% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_attribution.R
\name{attr_aggregations_by_type}
\alias{attr_aggregations_by_type}
\title{Get Sector Level Aggregate Energy and Emissions Data by Type}
\usage{
attr_aggregations_by_type(
  cf_energy_aggregated,
  sector_outputs,
  exclusion_type = c()
)
}
\arguments{
\item{cf_energy_aggregated}{A data frame containing aggregated emissions
and energy outputs. This is produced by \code{get_emissions_from_fuel_use()}.}

\item{sector_outputs}{A data frame containing all disctinct combinations of
sectors and primary outputs, produced by \code{get_sector_outputs()}.}

\item{exclusion_type}{A character vector of fuel categories to exclude from
the aggregation. Default is an empty character vector - i.e. no exclusion.}
}
\value{
A data frame with aggregated energy and emissions data, including
emissions intensity. There is one row per year and primary output.
}
\description{
This function aggregates energy and emissions data by sector, excluding
specified fuel categories.
}
