% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_availability.R
\name{availability}
\alias{availability}
\title{Constrains the amount of used capacity based on the available capacity}
\usage{
availability(data, decision_variables)
}
\arguments{
\item{data}{list of tables generated by preprocessing the raw inputs from the
input spreadsheet using \code{comit_preprocess_inputs()}}

\item{decision_variables}{dataframe for the decision variables. This is the
first element of the list produced by \code{comit_objective_function}.}
}
\value{
list of constraints. One constraint for each technology, at each site,
at each time point.
Each element contains a nested list with 4 elements:
\itemize{
\item column indices
\item values
\item direction
\item rhs
}
}
\description{
This constrains each technology, at each site, at each time point to only
use the capacity which is available to it. The available capacity is determined
by the available capacity variables multiplied by the availability factor of
the technology multiplied by the capacity to activity factor of the technology.
}
\details{
used_capacity(t,s,tech) <= available_capacity(t,s,tech) \* availability_factor(tech)
\* capacity_to_activity_factor(tech)
}
