% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_finance.R
\name{base_year_adjustment}
\alias{base_year_adjustment}
\title{Get prices in base year value.}
\usage{
base_year_adjustment(price_at_actual, parameter_data)
}
\arguments{
\item{price_at_actual}{numeric. The price to be inflated.}
}
\value{
numeric. The inflated price.
}
\description{
Inflates prices by required amount in order to set prices to the base year
specified in the input spreadsheet. This is used to convert the costs in the
outputs.
}
