% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_comit_counterfactual_solver.R
\name{baseline_constraint}
\alias{baseline_constraint}
\title{Create Constraint To Set Used Capacity to EEP rollout}
\usage{
baseline_constraint(data, decision_variables)
}
\arguments{
\item{data, }{list of dataframes containing input data, produced by
\code{comit_preprocess_inputs()}.}

\item{decision_variables}{dataframe for the decision variables. The first
element of the list produced by \code{comit_objective_function()}.}
}
\value{
list of lists, used to build constraints in the
\code{combine_matrix_constraints()} function.
}
\description{
Generate constraint for required used capacity of each technology for the
counterfactual run.
}
