% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_Primer_DB.R
\name{build_COVID_DB}
\alias{build_COVID_DB}
\title{Build COVID-19 Analysis Database}
\usage{
build_COVID_DB(
  DB_NAME,
  locationPrefix = "~",
  include_all_assays = TRUE,
  force = FALSE,
  silently = TRUE
)
}
\arguments{
\item{DB_NAME}{Desired database name, without filepath or .db extension}

\item{locationPrefix}{Folder location for database, defaults to home}

\item{include_all_assays}{Include primer information from all 7 assays or only A,D, and E assays?}

\item{force}{Force overwrite if file exists?}

\item{silently}{Turn off printed updates?}
}
\value{
full database file location name
}
\description{
Create a database given desired name and location, with primer information from the BioFire Defense Covid-19 Test assays
}
