% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_emissions.R
\name{calculate_CO2_direct_emissions}
\alias{calculate_CO2_direct_emissions}
\title{Calculate direct emissions of CO2 gasses}
\usage{
calculate_CO2_direct_emissions(emissions_df)
}
\arguments{
\item{emissions_df}{dataframe, created by \code{create_emissions_df()}}
}
\value{
dataframe, as input as emissions_df with additional columns:
\itemize{
\item CO2_process_released_direct
\item CO2_process_captured_direct
\item CO2_fuel_released_direct
\item CO2_fuel_captured_direct
}
}
\description{
Calculate direct emissions of CO2 gasses
}
