% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_process_sites.R
\name{calculate_coordinates}
\alias{calculate_coordinates}
\title{Convert easting and northing values into longitude and latitude}
\usage{
calculate_coordinates(df)
}
\arguments{
\item{df}{dataframe with columns 'Easting' and 'Northing' to be converted.}
}
\value{
dataframe as input with Longitude and Latitude columns output.
}
\description{
Convert easting and northing values into longitude and latitude
}
