% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_cost_tables.R
\name{calculate_costs_CO2_cluster2storage}
\alias{calculate_costs_CO2_cluster2storage}
\title{Produces the "CO2_C2S" rows in the cost tab of the Excel output workbook
(costs for transporting CO2 from clusters to terminals)}
\usage{
calculate_costs_CO2_cluster2storage(solved, model_data)
}
\arguments{
\item{solved}{a table containing the model solution, decision variables and
PV_coefficients}

\item{model_data}{list of data tables read in from excel data template}
}
\value{
table of cost data for transport of CO2 from clusters to storage sites
}
\description{
Produces the "CO2_C2S" rows in the cost tab of the Excel output workbook
(costs for transporting CO2 from clusters to terminals)
}
