% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_cost_tables.R
\name{calculate_costs_site2cluster}
\alias{calculate_costs_site2cluster}
\title{Generate table cost data for either H2 or CO2 pipes (and/or H2/CO2 trucking/grid transport)}
\usage{
calculate_costs_site2cluster(
  solved,
  model_data,
  site_cluster,
  capacity_types,
  pipe_PV_term,
  pipes_lifetime,
  cost_type_name
)
}
\arguments{
\item{solved}{a table containing the model solution, decision variables and
PV_coefficients}

\item{model_data}{list of data tables read in from excel data template}

\item{site_cluster}{variable indicating if site or cluster level data is to
be outputted.}

\item{capacity_types}{the variable types to filter for from the model solution}

\item{pipe_PV_term}{the PV term to filter for from the model output (e.g.PV_H2_pipe_cluster_to_site
for hydrogen pipes or PV_CO2_pipe_cluster_to_site for CO2 pipes)}

\item{pipes_lifetime}{the pipe lifetime (years)}

\item{cost_type_name}{string used to label the cost rows as site to cluster costs (e.g. H2_SCS or CO2_SCS)}
}
\value{
table of cost data for either H2 or CO2 pipes (and/or H2/CO2 trucking/grid transport)
}
\description{
Generate table cost data for either H2 or CO2 pipes (and/or H2/CO2 trucking/grid transport)
}
