% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_cost_tables.R
\name{calculate_costs_technology}
\alias{calculate_costs_technology}
\title{Calculates the technology costs incurred. Used to calculate the
non-infrastructure costs in the cost tabs of the Excel output workbook}
\usage{
calculate_costs_technology(solved_data, model_data, site_cluster)
}
\arguments{
\item{solved_data}{a table containing the model solution, decision variables and
PV_coefficients}

\item{model_data}{list of data tables read in from excel data template}

\item{site_cluster}{variable indicating if site or cluster level data is to
be outputted.}
}
\value{
table of technology costs with one row per cost type
}
\description{
Calculates the technology costs incurred. Used to calculate the
non-infrastructure costs in the cost tabs of the Excel output workbook
}
