% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_energy_tables.R
\name{calculate_fuel_use}
\alias{calculate_fuel_use}
\title{Calculate fuel used in the optimal model solution}
\usage{
calculate_fuel_use(output_data, model_data, group_cols)
}
\arguments{
\item{output_data}{a table containing the used technology capacities}

\item{model_data}{list of data tables read in from excel data template}

\item{group_cols}{a vector of the column names to be aggregated over}
}
\value{
a table with columns for Twh and ktCO2e added for each technology,
commodity, site/cluster-type and year combination
}
\description{
Calculate fuel used in the optimal model solution
}
