% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_cost_tables.R
\name{calculate_H2_extra_cost}
\alias{calculate_H2_extra_cost}
\title{Adds rows for capex_lump costs ('lumpy' yearly investment cost figures  without
borrowing, spreading or  discounting) for pipes to a table of pipe costs}
\usage{
calculate_H2_extra_cost(H2_extra_unit_cost, total_H2_energy, site_cluster)
}
\arguments{
\item{H2_extra_unit_cost}{a table containing the hydrogen_conversion rows
from the cost table data (i.e. output from \link{create_cost_tables})}

\item{total_H2_energy}{list of data tables read in from excel data template}

\item{site_cluster}{pipe lifetime}
}
\value{
table of cost data with H2 extra cost for each combination of
for each site, year, cost_type and technology

year
TWh
sector

H2_extra_unit_cost
Cost_type
2xxx
}
\description{
Adds rows for capex_lump costs ('lumpy' yearly investment cost figures  without
borrowing, spreading or  discounting) for pipes to a table of pipe costs
}
