% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_emissions_tables.R
\name{calculate_indirect_emissions}
\alias{calculate_indirect_emissions}
\title{Calculates the indirect emissions for the model solution}
\usage{
calculate_indirect_emissions(fuel_emissions_with_H2, energy_data)
}
\arguments{
\item{energy_data}{a table of energy (fuel) use through time}

\item{output_data}{a table containing a subset of the model solution outputs}
}
\value{
table of indirect emissions through time
}
\description{
Calculates the indirect emissions for the model solution
}
