% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_emissions_tables.R
\name{calculate_negative_emissions}
\alias{calculate_negative_emissions}
\title{Calculates the negative emissions for the model solution}
\usage{
calculate_negative_emissions(output_data, model_data, site_cluster)
}
\arguments{
\item{output_data}{a table containing a subset of the model solution outputs}

\item{model_data}{list of data tables read in from excel data template}

\item{site_cluster}{variable indicating if site or cluster level data is to
be outputted.}
}
\value{
table of negative emissions through time
}
\description{
Calculates the negative emissions for the model solution
}
