% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_process_sites.R
\name{calculate_non_point_site_emissions}
\alias{calculate_non_point_site_emissions}
\title{Helper function to calculate non_point_site_emissions using traded_share
only when parameter allows.}
\usage{
calculate_non_point_site_emissions(df, input_data)
}
\description{
Helper function to calculate non_point_site_emissions using traded_share
only when parameter allows.
}
