% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_emissions.R
\name{calculate_nonCO2_indirect_emissions}
\alias{calculate_nonCO2_indirect_emissions}
\title{Calculate indirect emissions of nonCO2 gasses}
\usage{
calculate_nonCO2_indirect_emissions(emissions_df)
}
\arguments{
\item{emissions_df}{dataframe, created by \code{create_emissions_df()}}
}
\value{
dataframe, as input as emissions_df with additional columns:
\itemize{
\item nonCO2_process_released_indirect
\item nonCO2_process_captured_indirect
\item nonCO2_fuel_released_indirect
\item nonCO2_fuel_captured_indirect
}
}
\description{
Calculate indirect emissions of nonCO2 gasses
}
