% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_process_sites.R
\name{calculate_pipe_distances}
\alias{calculate_pipe_distances}
\title{Get the pipe distance from site to cluster center}
\usage{
calculate_pipe_distances(df)
}
\arguments{
\item{df}{dataframe containing coordinates of point (Longitude and Latidude)
as well as coordinates of closest cluster center (Lon_clust and Lat_clust).}
}
\value{
input dataframe with pipe_dist column appended which contains the
pipe distances in kilometers.
}
\description{
Calculates the havershine distance between two set of coordinates - the sites
coordinates and the coordinates for the center of the nearest cluster.
}
