% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_process_sites.R
\name{calculate_share_of_sector_emissions}
\alias{calculate_share_of_sector_emissions}
\title{Get the proportion of emissions contributed by a site to its sector and
traded status combination}
\usage{
calculate_share_of_sector_emissions(df, input_data)
}
\arguments{
\item{df}{dataframe of the site (NAEI) data, containing sector, traded_status
emissions variables.}

\item{input_data}{data read in from excel template, same as what is passed to
\code{\link[=comit_solver]{comit_solver()}}.}
}
\value{
input data with sector_share_by_traded (as well as some columns used
to calculate this) appended, which is the proportion of emissions from the
site to the sum of emissions from the sites sector split and traded status.
}
\description{
Get the proportion of emissions contributed by a site to its sector and
traded status combination
}
