% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_emissions_tables.R
\name{calculate_total_emissions}
\alias{calculate_total_emissions}
\title{Calculates total emissions across both CO2 and nonCO2 and
pivots the table from wide to long format for years}
\usage{
calculate_total_emissions(emissions_data)
}
\arguments{
\item{emissions_data}{a table containing the (direct) emissions data}
}
\value{
table of (direct) emissions with a total emissions column added
}
\description{
Calculates total emissions across both CO2 and nonCO2 and
pivots the table from wide to long format for years
}
