% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_capacities.R
\name{capacity_constraints}
\alias{capacity_constraints}
\title{Create capacity limiting constraints (minimum or maximum capacities)}
\usage{
capacity_constraints(data, decision_variables, capacity_type)
}
\arguments{
\item{data}{list of tables generated by preprocessing the raw inputs from the
input spreadsheet using \code{comit_preprocess_inputs()}}

\item{decision_variables}{dataframe for the decision variables. This is the
first element of the list produced by \code{comit_objective_function}.}

\item{capacity_type}{character, either 'max' or 'min' depending on whether
the limit to be set is on maximum or minimum capacity.}
}
\value{
list of constraints. One constraint for each relevant technology, in
each time point.
Each element contains a nested list with 4 elements:
\itemize{
\item column indices
\item values
\item direction
\item rhs
}
}
\description{
This code contains all the steps required to build the minimum and maximum
capacity constraints, using the following functions:
\itemize{
\item \code{get_technology_capacity_tab} - gets the relevant input table containing
capacity limits.
\item \code{get_capacity_variables_data} - returns the data from the relevant decision
variables.
\item \code{add_alternative_technologies_data} - adds data for each variable of the
alternative technologies avaialble.
\item \code{formulate_capacity_constraint} - formally sets the constraint.
}
}
