% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_capacity_transfer.R
\name{capacity_transfer}
\alias{capacity_transfer}
\title{Create the capacity transfer constraint}
\usage{
capacity_transfer(data, decision_variables)
}
\arguments{
\item{data}{list of tables generated by preprocessing the raw inputs from the
input spreadsheet using \code{comit_preprocess_inputs()}}

\item{decision_variables}{dataframe for the decision variables. This is the
first element of the list produced by \code{comit_objective_function}.}
}
\value{
list of constraints. One constraint per variable for avaialable capacity.
Each element contains a nested list with 4 elements:
\itemize{
\item column indices
\item values
\item direction
\item rhs
}
}
\description{
This constrains the amount of availability of a given technology at a given
site, based on the new technologies added in previous years, plus the initial
capacity. In plain terms, it means that only the capacity that exists is available.
}
\details{
available_capacity(t,s,tech) = (∑ new_capacity(t,s,tech)) for all periods
t’ preceding t by the lifetime(tech) + residual_capacity(t,s,tech))
}
