% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_cost_tables.R
\name{capex_adjustment_pipe}
\alias{capex_adjustment_pipe}
\title{CAPEX ADJUSTMENT - undiscounted yearly spread cost}
\usage{
capex_adjustment_pipe(capacity_variable, cost_data, model_data, pipes_lifetime)
}
\arguments{
\item{capacity_variable}{variable type to filter for (H2 or CO2 pipes)}

\item{cost_data}{a table containing the pipe costs}

\item{model_data}{list of data tables read in from excel data template}

\item{pipes_lifetime}{lifetime for either H2 or CO2 pipes.}
}
\value{
table of costs with capex costs adjusted
}
\description{
CAPEX ADJUSTMENT - undiscounted yearly spread cost
}
