% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_emissions.R
\name{check_emissions_param}
\alias{check_emissions_param}
\title{Check that parameters provided to get_emissions() are allowed values}
\usage{
check_emissions_param(emissions_param, emissions_arg, allowed_args)
}
\arguments{
\item{emissions_param}{string, name of the parameter to check.}

\item{emissions_arg}{string, the value passed as a parameter to \code{get_emissions()}.}

\item{allowed_args}{character vector, the values allowed in this parameter.}
}
\value{
the argument as provided, unless the emissions_arg parameter is NULL
in which case the vector of allowed values is returned. Will raise an error
if an invalid value is provided as emissions_arg.
}
\description{
Check that parameters provided to get_emissions() are allowed values
}
