% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F-7-Align_Classify.R
\name{classifyFromAlignment}
\alias{classifyFromAlignment}
\title{Classify Variant from a masked alignment}
\usage{
classifyFromAlignment(
  primerAlignment,
  currPrimerID,
  accID,
  DATABASE_FILE,
  cospotList = NULL,
  saveIndelAligns = FALSE
)
}
\arguments{
\item{primerAlignment}{masked alignment}

\item{currPrimerID}{ID for the primer to classify}

\item{accID}{Accession Number for the sequence}

\item{DATABASE_FILE}{Database to add to}

\item{cospotList}{List of CoSpot Primers}

\item{saveIndelAligns}{Boolean that indicates whether the aligments should be saved to the server for later viewing}
}
\description{
Classify Variant from a masked alignment
}
