% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F-2-classify_helpers.R
\name{classifyVariant}
\alias{classifyVariant}
\title{Classify an identified variant sequence and insert into Variant Info and Variant Type tables}
\usage{
classifyVariant(origPrimer, variant, currPrimerID, isCospot = FALSE)
}
\arguments{
\item{origPrimer}{String of the DNA sequence for the primer}

\item{variant}{String of the DNA sequence for the variant}

\item{currPrimerID}{Database Id for the primer}

\item{isCospot}{Boolean, that determines whether variant info is pulled from the regular or cospot dataframe}
}
\value{
VarID number recieved from the database
}
\description{
Classify an identified variant sequence and insert into Variant Info and Variant Type tables
}
