% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_CO2.R
\name{CO2_cluster_to_storage_transport}
\alias{CO2_cluster_to_storage_transport}
\title{Constrain the capacity of CO2 transportation variables to meet the amount
of CO2 captured within a cluster}
\usage{
CO2_cluster_to_storage_transport(data, decision_variables)
}
\arguments{
\item{data}{list of tables generated by preprocessing the raw inputs from the
input spreadsheet using \code{comit_preprocess_inputs()}}

\item{decision_variables}{dataframe for the decision variables. This is the
first element of the list produced by \code{comit_objective_function}.}
}
\value{
list of constraints. One constraint for each cluster with the potential
to use CCS, in each time period. Each element contains a nested list
with 4 elements:
\itemize{
\item column indices
\item values
\item direction
\item rhs
}
}
\description{
Sets the amount of CO2 transportation required to remove the captured CO2
from each cluster to a storage site, often through means of another cluster
(referred to as a terminal).
}
