% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_CO2.R
\name{CO2_infrastructure_capacity}
\alias{CO2_infrastructure_capacity}
\title{Constrain CO2 infrastructure and CCS technology deployment, to ensure enough
CO2 can be transported from each site}
\usage{
CO2_infrastructure_capacity(data, decision_variables)
}
\arguments{
\item{data}{list of tables generated by preprocessing the raw inputs from the
input spreadsheet using \code{comit_preprocess_inputs()}}

\item{decision_variables}{dataframe for the decision variables. This is the
first element of the list produced by \code{comit_objective_function}.}
}
\value{
list of constraints. One constraint for each site with the potential
to use CCS technologies, in each time period. Each element contains a nested
list with 4 elements:
\itemize{
\item column indices
\item values
\item direction
\item rhs
}
}
\description{
Ensure that there is at least as much infrastructure capacity to accommodate
for the amount of carbon captured by CCS technologies. This constraint ensures
there is enough CO2 infrastructure implemented to transport the CO2 that is
captured.
}
\details{
Infrastructure here includes both pipes and trucking options. Each site has
only one CO2 infrastructure option available to it (pipes or trucking) which
is identified pre-model in the input template tab 'CO2_transport_cost'.
}
