% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_CO2.R
\name{CO2_storage_injection}
\alias{CO2_storage_injection}
\title{Constrains the amount of CO2 that can be stored at the CO2 storage site}
\usage{
CO2_storage_injection(data, decision_variables)
}
\arguments{
\item{data}{list of tables generated by preprocessing the raw inputs from the
input spreadsheet using \code{comit_preprocess_inputs()}}

\item{decision_variables}{dataframe for the decision variables. This is the
first element of the list produced by \code{comit_objective_function}.}
}
\value{
list of constraints. One constraint for each storage site that is
to be constrained, in each year. Each element contains a nested list
with 4 elements:
\itemize{
\item column indices
\item values
\item direction
\item rhs
}
}
\description{
Ensures that the maximum yearly amount CO2 that can be injected at each of the
storage sites is not exceeded.
}
