% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_decision_variables.R
\name{combine_decision_variables}
\alias{combine_decision_variables}
\title{Append all decision variable tables to create a single final table}
\usage{
combine_decision_variables(existing_tables)
}
\arguments{
\item{existing_tables}{a list of all decision variable tables that exist in
given scenario. Tables that don't exist in this scenario should be included
as NULL values in the list.}
}
\value{
dataframe containing the final decision variables. See
\code{\link[=create_decision_variables]{create_decision_variables()}} for more information.
}
\description{
Append all decision variable tables to create a single final table
}
