% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_combining_constraints.R
\name{combine_matrix_constraints}
\alias{combine_matrix_constraints}
\title{Combine a list of constraint sets into a single large set of constraints}
\usage{
combine_matrix_constraints(matrix_constraints)
}
\arguments{
\item{matrix_constraints}{a list for each constraint set, which contains a
sub list for direction, rhs and a sparse matrix for coefficients - see
\code{\link[=get_constraint_matrix]{get_constraint_matrix()}} for more details.}
}
\value{
a list with 3 elements, representing the entire set of constraints:
\itemize{
\item directions, a vector for the direction for each constraint.
\item rhs, a vector of the right hand side values for each constraint
\item matr, a simple triplet matrix. This is a converted version of all the input
matrices binded together. The simple triplet matrix is slower to work with
but required by the model, so it is converted here at the last stage. 'i'
references a constraint (row), 'j' references a decision variable (column)
and 'v' contains the value for the coefficient.
}
}
\description{
Collapse a series of constraint sets from different constraint types (such
production or emissions) into a single list containing all the required
information for each constraint, ready for modelling.
}
