% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_seqs.R
\name{comit_classify}
\alias{comit_classify}
\title{COMIT Variant Classifier}
\usage{
comit_classify(
  fileName,
  pullDate,
  DATABASE_FILE,
  cladeDF = NA,
  refSeqFile = NULL,
  removeHighN = TRUE,
  N_PER_FILTER = 1,
  silently = FALSE,
  leftSearchDist = 400,
  rightSearchDist = 150
)
}
\arguments{
\item{fileName}{File path to a FASTA containing sequences to sort}

\item{pullDate}{Date as a character vector that the FASTA was pulled from
GISAID,FORMAT: "YEAR-MONTH-DAY" i.e. "2021-03-05"}

\item{DATABASE_FILE}{File path to a CoMIT database}

\item{cladeDF}{Dataframe that can contain submission dates, lineages, and/or clades}

\item{refSeqFile}{A reference sequence for alignments. Default is the COVID Wuhan Sequence.}

\item{removeHighN}{Defaults to TRUE, will remove any sequences with more
than 1 percent N's}

\item{N_PER_FILTER}{Percentage of Ns to filer out}

\item{silently}{Run with out printing progress}

\item{leftSearchDist}{Integer, Left distance for buffer window region search}

\item{rightSearchDist}{Integer, Right distance for buffer window region search}
}
\value{
A vector with run log information
}
\description{
COMIT Variant Classifier
}
