% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comit_solver.R
\name{comit_constraints}
\alias{comit_constraints}
\title{Produce the constraints for the comit linear programme}
\usage{
comit_constraints(data, decision_variables, in_app = FALSE)
}
\arguments{
\item{data}{list of tables generated by preprocessing the raw inputs from the
input spreadsheet using \code{comit_preprocess_inputs()}}

\item{decision_variables}{dataframe for the decision variables. This is the
first element of the list produced by \code{comit_objective_function}.}

\item{in_app}{boolean, set to TRUE if running the function from the Rshiny
app. This allows the progress bar to be updated as the code progresses.}
}
\value{
List of length 3. The first element contains the direction for each
constraint (==, <=, or >=). The second element contains the right hand side
values. The third element contains the large simple triplet matrix, which
contains the non-zero coeffecients for decision variables for each constraint.
}
\description{
Sets up and combines all the constraints required for the linear programming
model. Uses the following functions:
\itemize{
\item \code{get_constraint_functions()}
\item \code{run_constraint_function()}
\item \code{combine_matrix_constraints()}
}
}
