% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_comit_counterfactual_solver.R
\name{comit_counterfactual_solver}
\alias{comit_counterfactual_solver}
\title{Solve the comit Counterfactual}
\usage{
comit_counterfactual_solver(raw_data)
}
\arguments{
\item{raw_data, }{a list of dataframes produced from \code{read_excel_data_template()}
that contains all data for the modelling. Importantly this list should include
the counterfactual rollout data named as 'counterfactual_rollout'.}
}
\value{
list of length four. The first element is the solution as produced by
\code{\link[ROI:ROI_solve]{ROI::ROI_solve()}}.The second element is the list of preprocessed data used to
find the solution. The third and fourth elements are the decision variables
and coefficients respectively, as produced by \code{\link[=comit_objective_function]{comit_objective_function()}}.
}
\description{
Finds the solution to comit under counterfactual conditions. The counterfactual
conditions are typically where no non-agreed future policies or interventions
from the government take place (EEP) however users can create there own
scenarios using the counterfactual_rollout tab of the input spreadsheet.
}
\details{
This allows comparisons to be made to normal comit runs and is fundamental
to producing attribution data.
}
