% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comit_solver.R
\name{comit_objective_function}
\alias{comit_objective_function}
\title{Generate decision variables and coefficients for the comit linear programming
model}
\usage{
comit_objective_function(data, in_app = FALSE)
}
\arguments{
\item{data}{list of tables generated by preprocessing the raw inputs from the
input spreadsheet using \code{comit_preprocess_inputs()}}

\item{in_app}{boolean, set to TRUE if running the function from the Rshiny
app. This allows the progress bar to be updated as the code progresses.}
}
\value{
list of length two. The first element is a dataframe comprising
the decision variables, as produced by create_decision_variables(). The
second element is a dataframe for the coefficients for the objective function
produced by the present value functions.
}
\description{
Create the data required to set up the comit objective function by creating
the tables for decision variables and coefficients.
}
