% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comit_solver.R
\name{comit_preprocess_inputs}
\alias{comit_preprocess_inputs}
\title{Perform the required data manipulation on the raw data}
\usage{
comit_preprocess_inputs(raw_data, in_app = FALSE)
}
\arguments{
\item{raw_data}{list of all input data, produced by \code{\link[=read_excel_data_template]{read_excel_data_template()}}
which is used to formulate the model and create all decision variables and
constraints.}

\item{in_app}{boolean, set to TRUE if running the function from the Rshiny
app. This allows the progress bar to be updated as the code progresses.}
}
\value{
list of all data tables for the input parameters, as provided by input
parameter raw_data, now cleaned and ready to be used to specify model
parameters.
}
\description{
Preprocesses the data required for the comit model, calling a series of other
functions:
\itemize{
\item \code{process_sites()}
\item \code{interpolate_data()}
\item \code{tidy()}
\item \code{round_years()}
\item \code{adjust_for_optimism()}
\item \code{apply_energy_efficiency()}
\item \code{adjust_existing_capacity()}
\item \code{add_site_ID()}
\item \code{assign_site_region()}
\item \code{add_H2_plants()}
\item \code{creat_H2_conversion()}
\item \code{site_demand()}
\item \code{site_H2C02_transport()}
}
}
