% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comit_solver.R
\name{comit_problem_solver}
\alias{comit_problem_solver}
\title{Solve the comit optimization problem and return the results}
\usage{
comit_problem_solver(
  data,
  decision_variables,
  PV_coefficients,
  constraints,
  in_app = FALSE
)
}
\arguments{
\item{data}{list of tables generated by preprocessing the raw inputs from the
input spreadsheet using \code{comit_preprocess_inputs()}}

\item{decision_variables}{dataframe for the decision variables. This is the
first element of the list produced by \code{comit_objective_function}.}

\item{PV_coefficients}{dataframe for the objective function coefficients
produced by \code{comit_objective_function}.}

\item{constraints}{dataframe for the constraints produced by
\code{comit_objective_function}.}

\item{in_app}{boolean, set to TRUE if running the function from the Rshiny
app. This allows the progress bar to be updated as the code progresses.}
}
\value{
list of length four. The first element is the solution as produced by
\code{\link[ROI:ROI_solve]{ROI::ROI_solve()}}.The second element is the list of data as included in the input.
The third and fourth elements are the decision variables and coefficients
respectively, as produced by \code{\link[=comit_objective_function]{comit_objective_function()}}.
}
\description{
The problem is solved using ROI_solve or HiGHS (depending on the problem)
and results are returned along with the input values. If the minimum
available capacity of H2 plants is set to 0 (the default) in the input
parameters, the problem is solved using linear programming, otherwise the
problem is solved using mixed-integer linear programming (MILP).
}
