% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comit_solver.R
\name{comit_solver}
\alias{comit_solver}
\title{Find lowest cost industry pathway based on input parameters and assumptions}
\usage{
comit_solver(raw_data, in_app = FALSE)
}
\arguments{
\item{raw_data}{list of all input data, produced by \code{\link[=read_excel_data_template]{read_excel_data_template()}}
which is used to formulate the model and create all decision variables and
constraints.}

\item{in_app}{boolean, set to TRUE if running the function from the Rshiny
app. This allows the progress bar to be updated as the code progresses.}
}
\value{
list of length 4, which contains:
\itemize{
\item solution - a list containing the model solution, produced by \code{\link[ROI:ROI_solve]{ROI::ROI_solve()}}
\item data - a list of data tables that is used to create the model
\item decision_variables - a sparse matrix of the decision variables used by the model
\item PV_coefficients - list of model coefficients
}
}
\description{
Creates and runs a linear programming model that finds an optimal solution
(minimized cost) for UK industry's pathway to 2050.
}
