% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comit_staged_solver.R
\name{comit_staged_solver}
\alias{comit_staged_solver}
\title{Run comit solver to a given stage}
\usage{
comit_staged_solver(raw_data, stage = 9)
}
\arguments{
\item{raw_data}{list of all input data, produced by \code{\link[=read_excel_data_template]{read_excel_data_template()}}
which is used to formulate the model and create all decision variables and
constraints.}

\item{stage}{integer signifying which stage of the system to run up to.
Default is 9.}
}
\value{
Output of comit_solver at specified stage, as follows:
\itemize{
\item 1 - list of dataframes returned after \code{\link[=process_sites]{process_sites()}}.
\item 2 - list of dataframes returned after additional data processing.
\item 3 - list of dataframes returned after further data processing, including
calculation of site demand.
\item 4 - dataframe for decision variables
\item 5 - dataframe for coeffecients
\item 6 - list of matrices for constraints
\item 7 - large single matrix for constraints
\item 8 - variable_types
\item 9 - list for final solved problem, returns same as \code{\link[=comit_solver]{comit_solver()}}.
}
}
\description{
This code is a copy of \code{\link[=comit_solver]{comit_solver()}}, with checkpoints added to allow
for staged running of the code for testing. When \code{\link[=comit_solver]{comit_solver()}} is updated
this function should be updated in the exact same way.
}
