% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utilities.R
\name{comit_tic}
\alias{comit_tic}
\title{Use tic function from tictoc package to record times, but only when tictoc
is available.}
\usage{
comit_tic(msg)
}
\arguments{
\item{msg}{string describing stage of model.}
}
\description{
Use tic function from tictoc package to record times, but only when tictoc
is available.
}
