% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utilities.R
\name{comit_toc}
\alias{comit_toc}
\title{Use toc function from tictoc package to record times, but only when tictoc
is available.}
\usage{
comit_toc()
}
\description{
Use toc function from tictoc package to record times, but only when tictoc
is available.
}
