% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_combining_constraints.R
\name{constraint_checks}
\alias{constraint_checks}
\title{Run a series of checks on constraint sets and stop the process if issues are
found}
\usage{
constraint_checks(constraint_group)
}
\arguments{
\item{constraint_group}{list of lists, for a set of constraints for a given
type of constraints, such as production constraints.
Each sub-list contains four elements:
\itemize{
\item column_indices (int) indicating the decision variables effected, varying
in length
\item values (num) indicating the coefficients for the variables effected, same length
as column_indices
\item direction (chr) indicating the direction of the constraint '==', '<=' or
'>='. Length 1.
\item rhs (num) the value of the right hand side of the constraint, length 1.
}}
}
\value{
NULL, process stopped if checks fail
}
\description{
Check the constraint sets for the expected shape and that there is no NA
values in the coefficient matrix.
}
