% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_combining_constraints.R
\name{constraint_size_check}
\alias{constraint_size_check}
\title{Raise message if a constraint includes more than a specified \% of the total
decision variables for a given period}
\usage{
constraint_size_check(
  constraint_name,
  this_constraint_list,
  data,
  decision_variables,
  issue_limit = 0.05,
  in_app = FALSE
)
}
\arguments{
\item{constraint_name}{str name of function to be run, usually passed from
an item of \code{\link[=get_constraint_functions]{get_constraint_functions()}}}

\item{this_constraint_list, }{list of constraint data produced by individual
constraint functions. The list is a list of nested lists, where each nested
list is a single constraint containing elements for column_indices, values,
direction and rhs.}

\item{data}{list of dataframes - raw_data after initial processing from
\code{\link[=process_sites]{process_sites()}}}

\item{decision_variables}{dataframe of decision variables as produced by
\code{\link[=create_decision_variables]{create_decision_variables()}}}

\item{issue_limit, }{numeric (default = 0.5\%). This is the proportion that, if
exceeded, the function will raise a message.}
}
\description{
This tests whether more than a specified proportion of decision variables
within a modelled period are used for a single constraint list. A message
is raised if the proportion is exceeded.
}
\details{
Exceeding the limit isn't inherently a problem, but large constraints do
impact on solve times, so this function flags when large constraints are
found to ensure that users are aware.
}
