% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_outputs.R
\name{correct_legacy_emissions_outputs}
\alias{correct_legacy_emissions_outputs}
\title{Backward-compatibility for legacy emissions outputs}
\usage{
correct_legacy_emissions_outputs(data)
}
\arguments{
\item{data}{Data frame of emissions outputs}
}
\value{
Data frame with consistent \code{Emissions_category}
}
\description{
Normalises historical files that used \code{Emissions_total} by creating or
renaming to \code{Emissions_category} and aligning labels.
}
\keyword{internal}
