% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_starting_values.R
\name{create_blue_and_green_hydrogen_technologies}
\alias{create_blue_and_green_hydrogen_technologies}
\title{Creates blue and green hydrogen technologies if required}
\usage{
create_blue_and_green_hydrogen_technologies(data, model_H2_production)
}
\arguments{
\item{data}{data read in from excel template}

\item{model_H2_production}{logical TRUE/FALSE depending on whether H2 production is modeled explicitly}
}
\value{
list of data
}
\description{
Creates blue and green hydrogen technologies if required
}
