% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_cost_tables.R
\name{create_cost_tables}
\alias{create_cost_tables}
\title{Generate table of cost outputs}
\usage{
create_cost_tables(solved, model_data, site_cluster)
}
\arguments{
\item{solved}{a table containing the model solution, decision variables and
PV_coefficients}

\item{model_data}{list of data tables read in from excel data template}

\item{site_cluster}{variable indicating if site or cluster level data is to
be outputted.}
}
\value{
table of cost outputs with one row per cost type
}
\description{
Uses the model solution to construct a table of all costs incurred,
used to populate the 'Costs' tab of the Excel output workbook.
}
