% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_decision_variables.R
\name{create_decision_variables}
\alias{create_decision_variables}
\title{Generate table of decision variables}
\usage{
create_decision_variables(data)
}
\arguments{
\item{data}{list of data tables read in from excel data template, after some
initial processing from preceding functions in \code{\link[=comit_solver]{comit_solver()}}.}
}
\value{
Table of decision variables, with one row per decision variable.
Columns include:
\itemize{
\item variable_index
\item variable_name
\item year
\item site_ID
\item code
\item cluster
\item variable_type (new, used or available capacity)
}
}
\description{
Uses the data read in from excel spreadsheet to build a table of all decision
variables that form the objective function and are to be solved by the model.
}
