% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_emissions.R
\name{create_emissions_df}
\alias{create_emissions_df}
\title{Create a dataframe with required information to calculate emissions}
\usage{
create_emissions_df(.data, technology_code, year)
}
\arguments{
\item{.data}{list of tables generated by preprocessing the raw inputs from the
input spreadsheet using \code{comit_preprocess_inputs()}. Default is the object
named 'data' which is used throughout comit code as the list of input tables.}

\item{technology_code}{either a string or a character vector, of UKTM technology
codes.}

\item{year}{either an integer or a vector of integers for the year which
emissions of a technology are to be calculated. Emissions can vary by year
due to fuel emissions changing over time for some fuels.}
}
\value{
dataframe, with one row per relevant commodity for distinct
combinations of technology and year provided as parameters. Columns include:
\itemize{
\item code - character
\item year - numeric
\item commodity - character
\item output - amount of commodity produced/used by the technology to produce
a single unit of output.
\item commodity_produces_emissions - boolean
\item proportion_emissions_CO2 - the proportion of emissions that are CO2 (as
opposed to non-CO2 emissions).
\item process_emission - boolean
\item commodity_category - character
\item fuel_CO2e - numeric, the amount of emissions per unit of fuel used.
\item emissions_released - the proportion of emissions released by the technology,
i.e. the proportion not captured.
\item direct_commodity - boolean, whether the commodity has direct (TRUE) or
INDIRECT (FALSE) emissions.
}
}
\description{
Creates a table for the relevant technologies and years, with all required
data to calculate emissions using the function \code{get_emissions()}.
}
