% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_emissions_tables.R
\name{create_emissions_tables}
\alias{create_emissions_tables}
\title{Generate table of emissions outputs}
\usage{
create_emissions_tables(output_data, energy, model_data, site_cluster)
}
\arguments{
\item{model_data}{list of data tables read in from excel data template}

\item{site_cluster}{variable indicating if site or cluster level data is to
be outputted.}

\item{solved}{a table containing the model solution, decision variables and
PV_coefficients}
}
\value{
table of emissions outputs with 1 row per emissions type
Columns include:
\itemize{
\item Emissions_category (e.g. direct / indirect)
\item Technology
\item GHG type
\item Emissions_type (energy/process)
\item years - one column for each timestep in the model
}
}
\description{
Uses the model solution to construct a table of all emissions through time
that will be used to populate the 'Emissions' tab of the Excel output workbook
}
