% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_energy_tables.R
\name{create_energy_tables}
\alias{create_energy_tables}
\title{Generates table of fuel use through time}
\usage{
create_energy_tables(output_data, model_data, site_cluster)
}
\arguments{
\item{model_data}{list of data tables read in from excel data template}

\item{site_cluster}{variable indicating if site or cluster level data is to
be outputted.}

\item{solved}{a table containing the model solution, decision variables and
PV_coefficients}
}
\value{
list containing a table of energy use and a table of
#' H2 energy use, to use for the cost calculation H2 sector adjustment
}
\description{
Uses the model solution to construct a table showing fuel use through time
that will be used to populate the 'Energy' tabs of the Excel output workbook
}
