% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_energy_tables.R
\name{create_infrastructure_tables}
\alias{create_infrastructure_tables}
\title{Generate table of H2 and CO2 transport infrastructure outputs}
\usage{
create_infrastructure_tables(solved, model_data, site_cluster)
}
\arguments{
\item{solved}{a table containing the model solution, decision variables and
PV_coefficients}

\item{model_data}{list of data tables read in from excel data template}

\item{site_cluster}{variable indicating if site or cluster level data is to
be outputted.}
}
\value{
table of infrastructure outputs with 1 row per variable_type for
each site or cluster.
Columns include:
\itemize{
\item site_ID
\item variable_type (new/used/available H2/CO2 pipe/grid/truck capacity)
\item unit
\item terminal
\item storage_site
\item pipe_cluster_end
\item years - one column for each timestep in the model
}
}
\description{
Uses the model solution to construct a table of all H2 and CO2 transport
infrastructure built and used through time that will be used to populate the
'Infrastructure' and 'Infrastructure_sites' tabs of the Excel output workbook
}
