% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_output_tables.R
\name{create_output_data}
\alias{create_output_data}
\title{A  table of the used technology capacity in the optimal model solution}
\usage{
create_output_data(solved, model_data)
}
\arguments{
\item{solved}{table containing the model solution}

\item{model_data}{list of data tables read in from excel template}
}
\value{
a data frame listing the used technology capacity
}
\description{
A  table of the used technology capacity in the optimal model solution
}
