% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_energy_tables.R
\name{create_outputs_table}
\alias{create_outputs_table}
\title{Generates table of 'Outputs', i.e. the deployment of industrial technologies
in terms of used capacity through time.}
\usage{
create_outputs_table(output_data, model_data, site_cluster)
}
\arguments{
\item{model_data}{list of data tables read in from excel data template}

\item{site_cluster}{variable indicating if site or cluster level data is to
be outputted.}

\item{solved}{a table containing the model solution, decision variables and
PV_coefficients}
}
\value{
table of Outputs with 1 row per technology for
each site or cluster.
Columns include:
\itemize{
\item sector
\item cluster_rad
\item Primary_output
\item Output_description
\item Technology_code
\item Technology_category
\item variable_type (used capacity)
\item years - one column for each timestep in the model
}
}
\description{
Uses the model solution to construct a table showing the deployment of
industrial technologies through time that will be used to populate the
'Outputs' tabs of the Excel output workbook
}
