% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_output_tables.R
\name{create_sites_info_tables}
\alias{create_sites_info_tables}
\title{Creates the Sites_info tab in the Output data workbook}
\usage{
create_sites_info_tables(model_data)
}
\arguments{
\item{model_data}{list of data tables read in from excel template}
}
\value{
a data frame listing information about each site in the model: name,
latitude, longitude, region, traded_status, cluster and NAEI plantID
}
\description{
Creates the Sites_info tab in the Output data workbook
}
