% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_output_tables.R
\name{create_sites_tables}
\alias{create_sites_tables}
\title{Creates the Sites tab in the Output data workbook}
\usage{
create_sites_tables(model_data)
}
\arguments{
\item{model_data}{list of data tables read in from excel template}
}
\value{
a data frame listing the number of sites for each sector, cluster,
traded status and cluster radius combination
}
\description{
Creates the Sites tab in the Output data workbook
}
