% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_output_tables.R
\name{create_solution_table}
\alias{create_solution_table}
\title{Creates a table containing the model solution, decision variables and PV_coefficients}
\usage{
create_solution_table(ROI_solution, decision_variables, PV_coefficients)
}
\arguments{
\item{ROI_solution}{table containing the model solution}

\item{decision_variables}{a sparse matrix of the decision variables used by the model}

\item{PV_coefficients}{list of model coefficients}
}
\value{
a table containing the model solution, decision variables and PV_coefficients
}
\description{
Creates a table containing the model solution, decision variables and PV_coefficients
}
