% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_run_model_for_all_inputs.R
\name{create_tables}
\alias{create_tables}
\title{Produce the required tables from the solved model}
\usage{
create_tables(solved, data, in_app = TRUE)
}
\arguments{
\item{solved}{the solved model object, produced from \code{comit_problem_solver()}.}

\item{data}{list of input data produced by \code{read_excel_data_template()}.}

\item{in_app, }{default = TRUE. Boolean for whether or not the function is
being ran as part of the main app, if it is the progress bar will be updated.}
}
\value{
list of all tables to be written to excel workbook.
}
\description{
Produce the required tables from the solved model
}
