% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_decision_variables.R
\name{create_variable_name}
\alias{create_variable_name}
\title{Generate decision variable name}
\usage{
create_variable_name(df, location_var = "site_ID")
}
\arguments{
\item{df}{dataframe containing original variable_name, year and site_ID variables.}

\item{location_var}{optional parameter (string) that can be used to provide the
name of the variable to be used in place of site_ID. Defaul it site_ID.}
}
\value{
input df with variable_name updated to include all required info
}
\description{
Creates names for decision variables based on an original variable name indicating
new/used/available capacity (or technology specific), combined with year and site_ID.
}
