% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_read_data.R
\name{custom_warning_handler}
\alias{custom_warning_handler}
\title{This funciton allows for the specific warning raised when reading in
'min_fuel_constraints' or 'fuel_constraints' sheet through the
be ran as usual. This is to avoid the expected warning be raised, meaning
it's easier to spot other warnings.}
\usage{
custom_warning_handler(w, sheet)
}
\arguments{
\item{w}{condition passed from withCallingHandlers}

\item{sheet}{string, the name of the sheet to bread}
}
\value{
NULL, restart of initial function is invoked
}
\description{
This funciton allows for the specific warning raised when reading in
'min_fuel_constraints' or 'fuel_constraints' sheet through the
be ran as usual. This is to avoid the expected warning be raised, meaning
it's easier to spot other warnings.
}
