% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_process_sites.R
\name{do_if_using_traded_share_calc}
\alias{do_if_using_traded_share_calc}
\title{Perform a given function only when traded share calculation parameter is true}
\usage{
do_if_using_traded_share_calc(df, f, input_data, pass_input_data = FALSE)
}
\arguments{
\item{df}{dataframe to perform function upon}

\item{f}{function to be carries out if traded_share_calc is TRUE}

\item{input_data}{data read in from excel template, same as what is passed to
\code{\link[=comit_solver]{comit_solver()}}.}

\item{pass_input_data}{boolean, whether input_data has been passed.}
}
\value{
dataframe, after function applied when traded_share_calc is TRUE, or
identical to input otherwise.
}
\description{
This function takes a function as an argument and runs it on a df when
the Traded_share_calc is set to TRUE, otherwise it just returns the df with
no change. This allows for a more succint implementation of the conditions
into the workflow.
}
