% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_attribution.R
\name{emissions_attribution}
\alias{emissions_attribution}
\title{Attribute Emissions Savings From Technology Categories}
\usage{
emissions_attribution(
  scenario_emissions,
  scenario_energy,
  cf_emissions,
  cf_energy,
  data,
  filter_type
)
}
\arguments{
\item{scenario_emissions}{dataframe, emissions outputs from a scenario run.
Can be generated from an output file by using \code{read_outputs()}.}

\item{scenario_energy}{dataframe, same as scenario_emissions but for energy.}

\item{cf_emissions}{dataframe, same as scenario_emissions, but for the
counterfactual.}

\item{cf_energy}{dataframe, same as scenario_emissions, but for the counterfactual's
energy outputs.}

\item{data}{list of dataframes, generated from passing the filepath for an
input spreadsheet to \code{read_excel_data_template()} and then passing the output
to \code{tidy()}.}

\item{filter_type}{character, the type of emissions to attribute (either
'Direct' or 'Direct_and_Indirect").}
}
\value{
dataframe, containing one row for each combination of year, sector and
abatement type. Columns produced are:
\itemize{
\item year - numeric
\item sector - character
\item savings - numeric, total amount of emissions saved in the scenario.
\item type - character, abatement type.
\item scaling_factor - numeric, the factor used in the calculation of savings.
}
}
\description{
Calculates the amount of emissions abated from each technology category, by
comparing the amount of emissions from a scenario with a counterfactual run.
}
