% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_attribution.R
\name{emissions_attribution_table}
\alias{emissions_attribution_table}
\title{Process Emissions Attribution Table}
\usage{
emissions_attribution_table(emissions_total, this_sector)
}
\arguments{
\item{emissions_total}{Data frame containing total emissions data.}

\item{this_sector}{Character string specifying the sector to filter by.
Use 'All' for no filter.}
}
\value{
Data frame containing emissions savings, in each year
and for each attribution type. Filtered by sector (and therefore not aggregated)
if a specific sector is provided as \code{this_sector}. Otherwise the outputs will
be aggregated to provide a total for all sectors.
}
\description{
This function processes the emissions attribution table providing savings for
a specified sector or all sectors.
}
