% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_emissions_tables.R
\name{emissions_combinations}
\alias{emissions_combinations}
\title{Calculates  total emissions through time across all combinations of
traded and non-traded sites, fuel/process and CO2/nonCO2 emissions using
\code{\link[=get_emissions]{get_emissions()}}.}
\usage{
emissions_combinations(
  emissions_data,
  captured,
  model_data,
  site_cluster,
  location_type = NULL
)
}
\arguments{
\item{emissions_data}{a table containing the emissions data}

\item{captured}{boolean TRUE, FALSE. When TRUE, returns only emissions that are captured. When FALSE, returns
only emissions that are released. If NULL, returns both captured and emitted emissions}

\item{model_data}{list of data tables read in from excel data template}

\item{site_cluster}{variable indicating if site or cluster level data is to
be outputted.}

\item{location_type}{string of either "direct" or "indirect". Default NULL returns both direct and indirect emissions}
}
\value{
table of emissions through time for each site, technology and output combination
}
\description{
Calculates  total emissions through time across all combinations of
traded and non-traded sites, fuel/process and CO2/nonCO2 emissions using
\code{\link[=get_emissions]{get_emissions()}}.
}
