% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_emissions.R
\name{emissions_limit}
\alias{emissions_limit}
\title{Constrain the amount of emissions that can be released by industry in each
year}
\usage{
emissions_limit(data, decision_variables)
}
\arguments{
\item{data}{list of tables generated by preprocessing the raw inputs from the
input spreadsheet using \code{comit_preprocess_inputs()}}

\item{decision_variables}{dataframe for the decision variables. This is the
first element of the list produced by \code{comit_objective_function}.}
}
\value{
list of constraints. One constraint for each time point.
Each element contains a nested list with 4 elements:
\itemize{
\item column indices
\item values
\item direction
\item rhs
}
}
\description{
This constraint states that the total amount of emissions released into
the atmosphere by industry cannot exceed a certain amount in a certain year.
}
