% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_attribution.R
\name{energy_output_by_tech}
\alias{energy_output_by_tech}
\title{Filter and Aggregate Energy Output by Technology}
\usage{
energy_output_by_tech(
  df,
  this_fuel_category,
  tech,
  negate_tech = FALSE,
  new_col_name
)
}
\arguments{
\item{df}{A data frame containing energy data.}

\item{this_fuel_category}{A string specifying the fuel category to filter by.}

\item{tech}{A string specifying the technology code to filter by.}

\item{negate_tech}{A boolean indicating whether to exclude the specified technology code
(TRUE) or include only the named technology code (FALSE). Default is FALSE.}

\item{new_col_name}{A string specifying the new column name for the summarized energy output.}
}
\value{
A data frame with total energy output by year, sector, and primary output.
}
\description{
This function filters and summarizes energy output by technology, sector, and
primary output.
}
