% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_upload_utilities.R
\name{error_notification}
\alias{error_notification}
\title{Generate Notification of Errorsome Upload to comit App}
\usage{
error_notification(err)
}
\description{
Generate Notification of Errorsome Upload to comit App
}
