% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_decision_variables.R
\name{expand_df_by_model_years}
\alias{expand_df_by_model_years}
\title{Add additional rows to account for years in model}
\usage{
expand_df_by_model_years(df, data)
}
\arguments{
\item{data}{list of data read in from model input file which contains model
parameters.}

\item{dataframe}{to expand.}
}
\value{
Dataframe with additional year column appended. Number of rows will be
number of rows in input column * number of years to be included in the model.
}
\description{
Add additional rows to account for years in model
}
