% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_emissions_tables.R
\name{expand_output_data}
\alias{expand_output_data}
\title{Adds additional required columns to the output_data table from the model_data list
Done upfront to avoid repetition}
\usage{
expand_output_data(output_data, model_data)
}
\arguments{
\item{output_data}{a table containing a subset of the model solution outputs}

\item{model_data}{list of data tables read in from excel data template
be outputted.}
}
\value{
table of output_data with the following columns added:
*cluster_rad
*traded_site
*(commodity) description
}
\description{
Adds additional required columns to the output_data table from the model_data list
Done upfront to avoid repetition
}
