% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_interpolate_data.R
\name{extrapolate_all_columns}
\alias{extrapolate_all_columns}
\title{Extrapolate missing start and/or end values of time series data}
\usage{
extrapolate_all_columns(df, extrapolated_years, new_years_to_add)
}
\arguments{
\item{df}{dataframe for the table to be extrapolated, with one row per year
for every year between the start and end years. The first column is the
year column and all following columns are extrapolated.}

\item{extrapolated_years}{a list of the years included in the time series
data plus the additional years that need to be extrapolated}

\item{new_years_to_add}{either 1 or 2, depending on whether either start year
and/or end year need to extrapolated}
}
\value{
dataframe as provided as df argument, but with missing end values
filled by extrapolation.
}
\description{
Extrapolates the missing end values outside of known points of time series
data. If either the model start year and/or end year are missing from
the time series, this function extrapolates these values.
For example, input data is often specified only to 2050, and
this function extrapolates to 2051 values to allow the model to run
with a 2051 end year. It also covers the case where the model start year
is 2021 and the time series includes values for 2020 but not 2021,
}
\details{
All extrapolation is linear. Only missing start/end year values are
extrapolated. Separate interpolation functions are used to fill other
missing values in the time series data.
}
