% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_interpolate_data.R
\name{extrapolate_for_years}
\alias{extrapolate_for_years}
\title{Function to check if extrapolation of start and/or end points is needed for
a dataframe. If so, it calls the 'extrapolate_all_columns()' function to
perform the extrapolation}
\usage{
extrapolate_for_years(df, all_years)
}
\arguments{
\item{df}{dataframe for the table to be extrapolated, with one row per year
for every year between the start and end years. The first column is the
year column and all following columns are to be extrapolated (if needed).}

\item{all_years}{dataframe with a single column containing every year between
the start and end years.}
}
\value{
dataframe as provided as df argument, but with missing end values
filled by extrapolation if needed. Otherwise, returns the df unchanged.
}
\description{
Function to check if extrapolation of start and/or end points is needed for
a dataframe. If so, it calls the 'extrapolate_all_columns()' function to
perform the extrapolation
}
