% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_process_sites.R
\name{filter_on_cluster_boundary}
\alias{filter_on_cluster_boundary}
\title{Joins cluster radius parameter for given cluster and then filters the df
based on whether inside or outside of cluster is required.}
\usage{
filter_on_cluster_boundary(df, input_data, inside = TRUE)
}
\description{
Joins cluster radius parameter for given cluster and then filters the df
based on whether inside or outside of cluster is required.
}
