% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-assay_summary.R
\name{format_as_gt_tbl}
\alias{format_as_gt_tbl}
\title{Format Assay Summary GT Table}
\usage{
format_as_gt_tbl(in_tbl, colWidth1 = NULL, colWidth2 = NULL)
}
\arguments{
\item{in_tbl}{GT table object}

\item{colWidth1}{width for "Pangolin Lineage" and "WHO Label" columns}

\item{colWidth2}{width for all other columns}
}
\value{
formatted GT table
}
\description{
Format Assay Summary GT Table
}
