% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-assay_summary_condensed.R
\name{format_as_gt_tbl_2}
\alias{format_as_gt_tbl_2}
\title{Format Assay Summary GT Table V2}
\usage{
format_as_gt_tbl_2(
  in_tbl,
  extraNotes = NULL,
  colWidth1 = NULL,
  colWidth2 = NULL,
  colWidth3 = NULL,
  HR = FALSE
)
}
\arguments{
\item{in_tbl}{GT table object}

\item{extraNotes}{vector of strings that are extra source notes to be displayed at the bottom of the table}

\item{colWidth1}{column width for "5" column}

\item{colWidth2}{column width for ">= 6" column}

\item{colWidth3}{column width for "Sequences by Lineage" column}

\item{HR}{aka high risk, a boolean value indicating whether or not the input table has only data associated with high-risk mutations}
}
\value{
formatted GT table
}
\description{
Format Assay Summary GT Table V2
}
