% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-co_occurring.R
\name{format_co_gt_tbl}
\alias{format_co_gt_tbl}
\title{Format Co-occurring Mutations GT Table}
\usage{
format_co_gt_tbl(in_tbl, colWidth1 = NULL, colWidth2 = NULL)
}
\arguments{
\item{in_tbl}{GT table object}

\item{colWidth1}{width of "Pangolin Lineage" column in px}

\item{colWidth2}{width of all other columns in px}
}
\value{
formatted GT table
}
\description{
Format Co-occurring Mutations GT Table
}
