% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-co_occurring_condensed.R
\name{format_co_gt_tbl_2}
\alias{format_co_gt_tbl_2}
\title{Format Co-occurring Mutations GT Table V2}
\usage{
format_co_gt_tbl_2(
  in_tbl,
  colWidth1 = NULL,
  colWidth2 = NULL,
  colWidth3 = NULL,
  colWidth4 = NULL,
  colWidth5 = NULL,
  HR = FALSE
)
}
\arguments{
\item{in_tbl}{GT table object}

\item{colWidth1}{width of "Pangolin Lineage" column in px}

\item{colWidth2}{width of 0, 1, 2, and 3 columns in px}

\item{colWidth3}{width of "5" column in px}

\item{colWidth4}{width of ">=6" column in px}

\item{colWidth5}{width of "Sequences by Lineage" column in px}

\item{HR}{aka high risk, a boolean value indicating whether or not the input table has only data associated with high-risk mutations}
}
\value{
formatted GT table
}
\description{
Format Co-occurring Mutations GT Table V2
}
