% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-database_breakdown_condensed.R
\name{formatDF_2}
\alias{formatDF_2}
\title{Format Database Breakdown Dataframe V2}
\usage{
formatDF_2(df, dateColName, total_seqs, total_seqs_1m, oneMonth = FALSE)
}
\arguments{
\item{df}{the dataframe to be formatted}

\item{dateColName}{the user-inputted string describing the date analyzed, either
the supplied 3-month period or 1-month period, e.g. May - July 2022 or July 2022}

\item{total_seqs}{total number of sequences for the 3-month period}

\item{total_seqs_1m}{total number of sequences for the 1-month period}

\item{oneMonth}{a logical variable, whether the dataframe to be formatted describes data over a 3-month or 1-month period}
}
\value{
out_df, a dataframe with preliminary formatting applied (e.g. counts, percentages, column names) before it is combined with the other to create the gt table
}
\description{
Format Database Breakdown Dataframe V2
}
