% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_availability.R
\name{formulate_availability_constraint}
\alias{formulate_availability_constraint}
\title{Create the list of constraints for availability}
\usage{
formulate_availability_constraint(availability_constraint_data)
}
\arguments{
\item{availability_constraint_data, }{data.table containing the data to set up
the constraint. Requires the following columns:
\itemize{
\item variable_index.used_capacity - the index of the used capacity variable
\item variable_index.avaialable_capacity - the index of the available capacity variable
\item total_factor - the factor to times the available capacity by (available_factor \*
capacity_to_activity_factor)
}}
}
\value{
list of constraints. One constraint for each technology, at each site,
at each time point.
Each element contains a nested list with 4 elements:
\itemize{
\item column indices
\item values
\item direction
\item rhs
}
}
\description{
Create the list of constraints for availability
}
