% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_capacities.R
\name{formulate_capacity_constraint}
\alias{formulate_capacity_constraint}
\title{Formulate the constraint for minimum and maximum capacities of technologies}
\usage{
formulate_capacity_constraint(capacity_variables, capacity_type)
}
\arguments{
\item{capacity_variables}{dataframe for decision variables that are relevant
to the capacity constraints, as produced by \code{add_alternative_technologies_data}.}

\item{capacity_type}{character, either 'max' or 'min' used to determine whether
the constraint is a minimum or maximum constraint.}
}
\value{
list of constraints. One constraint for each relevant technology, in
each year.
Each element contains a nested list with 4 elements:
\itemize{
\item column indices
\item values
\item direction
\item rhs
}
}
\description{
Sets up the list of constraints to be used in the model to provide either an
upper or lower limit to the share of a technology used to create a commodity
in each year.
}
\details{
The formulation of this constraint is not immediately intuitive when looking
at the code, but it has been arrived at based on the following rearrangement
which will hopefully provide more understanding.

Lets say we want to limit a technology (this_tech) to be less than 10\% of a
commodities production, then:

this_tech <= 0.1(this_tech + all_other_techs)

this_tech <= 0.1(this_tech) + 0.1(all_other_techs)

this_tech - 0.1(this_tech) - 0.1(all_other_techs) <= 0

0.9(this_tech) - 0.1(all_other_techs) <= 0

The formulation is the same for minimum and maximum technologies, with the slight
change from ">=" to "<=" from minimum to maximum.
}
