% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_capacity_transfer.R
\name{formulate_capacity_transfer_constraint}
\alias{formulate_capacity_transfer_constraint}
\title{Create the list of constraints for capacity transfer}
\usage{
formulate_capacity_transfer_constraint(capacity_transfer_constraint)
}
\arguments{
\item{capacity_transfer_constraint}{data.table, containing the data required
to generate the capacity transfer constraint (requires the following columns:
variable_index.available, variable_index.new, int, residual_capacity.available).}
}
\value{
list of constraints. One constraint per variable for available capacity.
Each element contains a nested list with 4 elements:
\itemize{
\item column indices
\item values
\item direction
\item rhs
}
}
\description{
Create the list of constraints for capacity transfer
}
