% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_CO2.R
\name{formulate_CO2_cluster_to_storage_constraint}
\alias{formulate_CO2_cluster_to_storage_constraint}
\title{Formulates the CO2 cluster to storage transport constraints}
\usage{
formulate_CO2_cluster_to_storage_constraint(CO2_cluster_to_storage_variables)
}
\arguments{
\item{CO2_cluster_to_storage_variables}{dataframe, produced by
\code{get_CO2_cluster_to_storage_variables()}.}
}
\value{
list of constraints. One constraint for each cluster with the potential
to use CCS, in each time period. Each element contains a nested list
with 4 elements:
\itemize{
\item column indices
\item values
\item direction
\item rhs
}
}
\description{
Formulates the CO2 cluster to storage transport constraints
}
