% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_CO2.R
\name{formulate_CO2_infrastructure_constraint}
\alias{formulate_CO2_infrastructure_constraint}
\title{Formulates the CO2 infrastructure capacity constraints}
\usage{
formulate_CO2_infrastructure_constraint(CO2_infrastructure_constraint_data)
}
\arguments{
\item{CO2_infrastructure_constraint_data}{dataframe containing both data on
the CO2 infrastructure variables, and the CO2 capture technology variables for
each site and in each year.}
}
\value{
list of constraints. One constraint for each site with the potential
to use CCS technologies, in each time period. Each element contains a nested
list with 4 elements:
\itemize{
\item column indices
\item values
\item direction
\item rhs
}
}
\description{
The constraint is essentially: \emph{CO2 infrastructure capacity >= CO2 caputred},
at each site and in each year.
}
