% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_emissions.R
\name{formulate_emissions_constraint}
\alias{formulate_emissions_constraint}
\title{Formulate the emissions limit constraint}
\usage{
formulate_emissions_constraint(emissions_variables)
}
\arguments{
\item{emissions_variables}{}
}
\value{
list of constraints. One constraint for each time point.
Each element contains a nested list with 4 elements:
\itemize{
\item column indices
\item values
\item direction
\item rhs
}
}
\description{
Formulate the emissions limit constraint
}
