% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_production.R
\name{formulate_end_demand_production_constraint}
\alias{formulate_end_demand_production_constraint}
\title{Create the constraint which ensures end demand is met}
\usage{
formulate_end_demand_production_constraint(end_demand_vars)
}
\arguments{
\item{end_demand_vars}{dataframe, production_df filtered to include only the
technologies which produce end demand commodities directly.}
}
\value{
list of constraints. One constraint for each commodity, at each site,
at each time point.
Each element contains a nested list with 4 elements:
\itemize{
\item column indices
\item values
\item direction
\item rhs
}
}
\description{
This constraint ensures each site produces the correct amount of its end
demand commodity (e.g. cement) in each year.
}
\details{
The inequality states that in the case of an end demand process, the sum of
technologies has to equal the demand in that year. The demand is given in the
input_amount.input column, which is an incorrect name but used for consistency.
This column is actually taken from the site_demand data.
}
