% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_fuel.R
\name{formulate_fuel_constraints}
\alias{formulate_fuel_constraints}
\title{Create the list of constraints to limit the absolute amount of fuel to be used}
\usage{
formulate_fuel_constraints(fuel_constrained_variables, constraint_type)
}
\arguments{
\item{fuel_constrained_variables}{dataframe containing all decision variables
to be constrained, as produced by \code{aggregate_fuel_constrained_variables()}.}

\item{constraint_type}{character, either 'max' or 'min' depending on whether
the limit to be set is on maximum or minimum fuel use.}
}
\value{
list of constraints. One constraint for each fuel specified to be
constrained in the input, in each time point. Each element contains a
nested list with 4 elements:
\itemize{
\item column indices
\item values
\item direction
\item rhs
}
}
\description{
Generates the information required in the correct format to set constraints in
comit. This funciton can be used for both minimum and maximum fuel constraints,
switching the direction sign '>=' or '<=' based on the constraint type parameter.
}
