% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_fuel_share.R
\name{formulate_fuel_share_constraints}
\alias{formulate_fuel_share_constraints}
\title{Create the list of constraints to limit the proportional amount of fuel to be
used}
\usage{
formulate_fuel_share_constraints(
  fuel_consuming_variables,
  fuel_constraint_combinations
)
}
\arguments{
\item{fuel_consuming_variables}{dataframe of the fuel consuming decision variables.
Produced by \code{get_fuel_consuming_variables}.}

\item{fuel_constraint_combinations}{dataframe, produced by
\code{get_fuel_share_constraint_combinations}.}
}
\value{
list of constraint. One constraint for each fuel to be constrained,
for each output commodity to be constrained, in each time point. Each element
contains a nested list with 4 elements:
\itemize{
\item column indices
\item values
\item direction
\item rhs
}
}
\description{
Generates the information required in the correct format to set constraints
in comit, allowing fuel shares to be constrained in the production of
commodities.
}
