% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_hydrogen.R
\name{formulate_H2_availability_constraint}
\alias{formulate_H2_availability_constraint}
\title{Formulates the H2 availability constraints}
\usage{
formulate_H2_availability_constraint(H2_variables)
}
\arguments{
\item{H2_variables}{dataframe produced by \code{get_H2_variables_with_availability()},
containing one row per decision variable to be constrained and including
the column 'net_availability' which informs the limit in the constraint.}
}
\value{
list of constraints. One constraint for each cluster at each time
point. Each element contains a nested list with 4 elements:
\itemize{
\item column indices
\item values
\item direction
\item rhs
}
}
\description{
Used by \code{H2_availability} to produce hydrogen availability constraints.
}
