% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_hydrogen.R
\name{formulate_H2_conversion_constraint}
\alias{formulate_H2_conversion_constraint}
\title{Formulates the H2 conversion constraints}
\usage{
formulate_H2_conversion_constraint(hydrogen_variables)
}
\arguments{
\item{hydrogen_variables, }{dataframe produced by
\code{get_hydrogen_conversion_constraint_data()}. Contains data for the hydrogen
consuming and hydrogen producing (converting) variables.}
}
\value{
list of constraints. One constraint for each time point where
there are hydrogen used capacity decision variables in the model. Each
element contains a nested list with 4 elements:
\itemize{
\item column indices
\item values
\item direction
\item rhs
}
}
\description{
Used by \code{H2_conversion} to produce hydrogen conversion constraints.
}
