% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_hydrogen.R
\name{formulate_H2_infrastructure_capacity_constraint}
\alias{formulate_H2_infrastructure_capacity_constraint}
\title{Formulates the H2 infrastructure capacity constraints}
\usage{
formulate_H2_infrastructure_capacity_constraint(H2_consuming_variables)
}
\arguments{
\item{H2_consuming_variables, }{dataframe produced by
\code{get_H2_consuming_variables(data, decision_variables)} joined with infrastructure
decision variables. One row per combination of hydrogen consuming technology
variables and H2 infrastructure variables.}
}
\value{
list of constraints. One constraint for each site with the potential
to use hydrogen, in each time period. Each element contains a nested list
with 4 elements:
\itemize{
\item column indices
\item values
\item direction
\item rhs
}
}
\description{
Formulates the H2 infrastructure capacity constraints
}
