% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_CO2.R
\name{formulate_injection_constraints}
\alias{formulate_injection_constraints}
\title{Formulate the constraint for CO2 storage injection}
\usage{
formulate_injection_constraints(injection_constrained_variables)
}
\arguments{
\item{injection_constrained_variables}{dataframe, produced by
\code{CO2_storage_injection()}.}
}
\value{
list of constraints. One constraint for each storage site that is
to be constrained, in each year. Each element contains a nested list
with 4 elements:
\itemize{
\item column indices
\item values
\item direction
\item rhs
}
}
\description{
Formulate the constraint for CO2 storage injection
}
