% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_CO2.R
\name{formulate_max_CCS_constraint}
\alias{formulate_max_CCS_constraint}
\title{Formulate the max_CCS constraint}
\usage{
formulate_max_CCS_constraint(max_ccs_data)
}
\arguments{
\item{max_ccs_data}{dataframe, produced by \code{get_max_CCS_data()} containing
the data for the decision variables to be constrained.}
}
\value{
list of constraints. One constraint for each time period with a
constraint on the maximum amount of carbon that can be captured.
Each element contains a nested list with 4 elements:
\itemize{
\item column indices
\item values
\item direction
\item rhs
}
}
\description{
Formulate the max_CCS constraint
}
