% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_production.R
\name{formulate_process_demand_production_constraint}
\alias{formulate_process_demand_production_constraint}
\title{Create the constraint that ensures enough process commodities are produced}
\usage{
formulate_process_demand_production_constraint(process_demand_vars)
}
\arguments{
\item{process_demand_vars}{dataframe, production_df filtered to include only
technologies which DO NOT produce end demand commodities directly.}
}
\value{
list of constraints. One constraint for each commodity, at each site,
at each time point.
Each element contains a nested list with 4 elements:
\itemize{
\item column indices
\item values
\item direction
\item rhs
}
}
\description{
This constraint ensures that each process at a site produces the
correct amount of commodity for inputs to other processes, in each year.
This constraint requires that the output commodity produced by any process
at a site (other than the process producing the end demand commodity) is equal
to the amount that commodity is used by other technologies at a site.
}
\details{
The sum of the output technologies minus the sum of technologies using that
output multiplied by the amount they use has to equal 0.
}
