% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_process_sites.R
\name{get_aggregated_small_point_sites}
\alias{get_aggregated_small_point_sites}
\title{Create table for small point sites}
\usage{
get_aggregated_small_point_sites(df, input_data)
}
\arguments{
\item{input_data}{data read in from excel template, same as what is passed to
\code{\link[=comit_solver]{comit_solver()}}.}
}
\value{
input dataframe now manipulated and reduced to the following columns,
for small point sites only:
\itemize{
\item site_name
\item IPM_sector
\item H2_point
\item total_MtCO2 (emissions)
\item pipe_dist
\item num_sites
\item Latitude
\item Longitude
\item traded_flag
}
}
\description{
Aggregate all small point sites that are in the same cluster, distance from
the center of the cluster (in or out of range) and sector. Note that these
are point sites (from NAEI data source) but are non-traded and have
smaller amounts of emissions. Therefore, we group them into single points
to reduce the computational burden of the model. This function does such
grouping, including required imputation of coordinates and returns the table
of the artificial aggregate sites.
}
