% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_fuel_share.R
\name{get_all_fuel_commodities}
\alias{get_all_fuel_commodities}
\title{Return the fuel commodities from the commodity input data}
\usage{
get_all_fuel_commodities(commodities)
}
\arguments{
\item{commodities}{dataframe for the commodities data.}
}
\value{
dataframe for the fuel commodities. Will include the columns from
the input parameter \emph{commodities}.
}
\description{
Return the fuel commodities from the commodity input data
}
