% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_attribution.R
\name{get_attribution_parameters}
\alias{get_attribution_parameters}
\title{Get Attribution Parameters}
\usage{
get_attribution_parameters(data)
}
\arguments{
\item{data}{list of dataframes, generated from passing the filepath for an
input spreadsheet to \code{read_excel_data_template()} and then passing the output
to \code{tidy()}.}
}
\value{
A list containing two data frames: \code{attr_params} and \code{BECCS_share}.
}
\description{
This function loads the default assumptions data and extracts attribution-specific
parameters and BECCS share data.
}
