% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_attribution.R
\name{get_beccs_and_css_savings}
\alias{get_beccs_and_css_savings}
\title{Calculate Emission Savings From BECCS and CCS}
\usage{
get_beccs_and_css_savings(ccs_captured, scaled_emissions, BECCS_share)
}
\arguments{
\item{ccs_captured}{Data frame containing captured emissions data.}

\item{scaled_emissions}{Data frame containing emissions scaling factors.}

\item{BECCS_share}{Data frame containing the share of BECCS in each sector.}
}
\value{
Data frame containing the savings from BECCS and CCS.
}
\description{
This function calculates the savings from BECCS (Bioenergy with Carbon
Capture and Storage) and CCS (Carbon Capture and Storage) based on captured
emissions, scaled emissions, and BECCS share.
}
