% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_attribution.R
\name{get_biomass_savings}
\alias{get_biomass_savings}
\title{Calculate Emission Savings From Biomass}
\usage{
get_biomass_savings(
  scenario_emissions,
  sector_outputs,
  scenario_energy,
  cf_energy,
  REEE_df,
  cf_energy_aggregated,
  attr_params,
  data
)
}
\arguments{
\item{scenario_emissions}{Data frame containing scenario emissions data.}

\item{sector_outputs}{Data frame containing sector outputs data.}

\item{scenario_energy}{Data frame containing scenario energy data.}

\item{cf_energy}{Data frame containing counterfactual energy data.}

\item{REEE_df}{Data frame containing REEE data.}

\item{cf_energy_aggregated}{Data frame containing aggregated counterfactual energy data.}

\item{attr_params}{Dataframe of attribute parameters.}

\item{data}{list of dataframes, generated from passing the filepath for an
input spreadsheet to \code{read_excel_data_template()} and then passing the output
to \code{tidy()}.}
}
\value{
Data frame containing biomass savings for each year and sector.
}
\description{
This function calculates the biomass savings based on scenario energy and
emissions compared to counterfactual energy use.
}
