% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_capacity_transfer.R
\name{get_capacity_variable_types}
\alias{get_capacity_variable_types}
\title{Generate table of capacity variable types}
\usage{
get_capacity_variable_types()
}
\value{
data.table with the corresponding new and available variables for each
capacity variable type.
}
\description{
Generate table of capacity variable types
}
