% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_decision_variables.R
\name{get_capacity_variables}
\alias{get_capacity_variables}
\title{Produce table for capacity variables}
\usage{
get_capacity_variables(site_technologies)
}
\arguments{
\item{site_technologies}{data frame containing both site and technology data.}
}
\value{
data frame with \code{3 * nrow(site_technologies)} rows and 5 columns:
\itemize{
\item variable name
\item year
\item site ID
\item technology code
\item cluster
@export
}
}
\description{
Creates the table of all variables for capacities (new, available and used)
for each site and technology combination.
}
