% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_capacities.R
\name{get_capacity_variables_data}
\alias{get_capacity_variables_data}
\title{Get the relevant decision variables for the capacity constraint}
\usage{
get_capacity_variables_data(data, decision_variables, capacity_tab)
}
\arguments{
\item{data}{list of tables generated by preprocessing the raw inputs from the
input spreadsheet using \code{comit_preprocess_inputs()}}

\item{decision_variables}{dataframe for the decision variables. This is the
first element of the list produced by \code{comit_objective_function}.}

\item{capacity_tab}{dataframe, constaining data on the technologies to constrain
as produced by \code{get_technology_capacity_tab}.}
}
\value{
dataframe, with one row per decision variable for the relevant
technologies. Columns are:
\itemize{
\item variable_index
\item year
\item site_ID
\item code - technology code
\item output_commodity
}
}
\description{
This function filters the decision_variables dataset down to only used_capacity
variables for the technologies that are present in the capacity_tab. It also
adds output commodity information from the technologies data.
}
