% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_attribution.R
\name{get_ccs_captured}
\alias{get_ccs_captured}
\title{Get CCS Captured Emissions}
\usage{
get_ccs_captured(scenario_emissions)
}
\arguments{
\item{scenario_emissions}{Data frame containing scenario emissions.}
}
\value{
Data frame with total captured emissions by year and sector.
}
\description{
This function calculates the total captured emissions for Carbon Capture and
Storage (CCS) technology by year and sector.
}
