% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_decision_variables.R
\name{get_CCS_from_cluster_transport}
\alias{get_CCS_from_cluster_transport}
\title{Create the decision variables for transport of CO2 from cluster to storage site}
\usage{
get_CCS_from_cluster_transport(data, CCS_sites_transport)
}
\arguments{
\item{data}{list of data tables read in from excel data template, after some
initial processing from preceding functions in \code{\link[=comit_solver]{comit_solver()}}.}

\item{CCS_sites_transport}{dataframe created by \code{\link[=get_CCS_sites_transport]{get_CCS_sites_transport()}}}
}
\value{
dataframe for decision variables regarding transport of C02 out of
clusters to carbon storage sites.
}
\description{
Create the decision variables for transport of CO2 from cluster to storage site
}
