% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_decision_variables.R
\name{get_CCS_sites}
\alias{get_CCS_sites}
\title{Produce CCS data for relevant site}
\usage{
get_CCS_sites(data, site_technologies)
}
\arguments{
\item{data}{list of data tables read in from excel data template, after some
initial processing from preceding functions in \code{\link[=comit_solver]{comit_solver()}}.}

\item{site_technologies}{data frame containing both site and technology data.}
}
\value{
dataframe containing the following information for each site and year for
sites which use CCS technologies:
\itemize{
\item H2_point
\item CCS_start
\item lowest_cost_option.C02
}
}
\description{
Generates CCS related variables for each combination of site and year for
all sites that use CCS technologies. This table can then be used to create
the CCS related decision variables.
}
