% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_decision_variables.R
\name{get_CCS_sites_transport}
\alias{get_CCS_sites_transport}
\title{Create the decision variables for CCS transport types for sites}
\usage{
get_CCS_sites_transport(data, CCS_sites)
}
\arguments{
\item{data}{list of data tables read in from excel data template, after some
initial processing from preceding functions in \code{\link[=comit_solver]{comit_solver()}}.}

\item{CCS_sites}{dataframe, output from \code{\link[=get_CCS_sites]{get_CCS_sites()}}.}
}
\value{
dataframe for decision variables regarding transport of CCS
at specific sites.
}
\description{
Includes the variables for pipes and trucks and combines them into a
single data frame.
}
