% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_process_sites.R
\name{get_cluster_locations}
\alias{get_cluster_locations}
\title{Helper function to call and tidy cluster locations}
\usage{
get_cluster_locations(input_data)
}
\arguments{
\item{input_data}{data read in from excel template, same as what is passed to
\code{\link[=comit_solver]{comit_solver()}}.}
}
\value{
dataframe with one row per cluster, with variables 'Cluster',
'Lat_clust' (cluster latitude) and 'Lon_clust' (cluster longitude).
}
\description{
Helper function to call and tidy cluster locations
}
