% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_CO2.R
\name{get_CO2_capture_variables}
\alias{get_CO2_capture_variables}
\title{Create a dataframe of the decision variables that use CCS}
\usage{
get_CO2_capture_variables(data, decision_variables, biomass_setting = FALSE)
}
\arguments{
\item{data}{list of tables generated by preprocessing the raw inputs from the
input spreadsheet using \code{comit_preprocess_inputs()}}

\item{decision_variables}{dataframe for the decision variables. This is the
first element of the list produced by \code{comit_objective_function}.}

\item{biomass_setting, }{boolean, default = FALSE. The setting to pass as
zero_emissions_from_biomass in the \code{get_emissions()} function used to
calculat the amount of captured emisssion.}
}
\value{
dataframe with one row per decision variable that is a technology
that captures CO2. Columns are:
\itemize{
\item variable_index
\item year
\item site_ID
\item code
\item captured_CO2 - numeric, the amount of CO2 captured per unit of capacity.
}
}
\description{
Get the decision variables for technologies that use carbon capture. These
are then used constrain the amount of infrastructure needed through the
'formulate_CO2_infrastructure_constraint()' function.
}
