% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_CO2.R
\name{get_CO2_cluster_to_storage_variables}
\alias{get_CO2_cluster_to_storage_variables}
\title{Get the data required to formulate the CO2 cluster to storage transport
constraint.}
\usage{
get_CO2_cluster_to_storage_variables(data, decision_variables)
}
\arguments{
\item{data}{list of tables generated by preprocessing the raw inputs from the
input spreadsheet using \code{comit_preprocess_inputs()}}

\item{decision_variables}{dataframe for the decision variables. This is the
first element of the list produced by \code{comit_objective_function}.}
}
\value{
dataframe, with one row per feasible combination of decision variables
with CCS and CO2_transported variables. Columns include:
\itemize{
\item variable_index.technology - the variable index for technologies.
\item year
\item site_ID
\item cluster
\item captured_CO2 - amount of CO2 captured within the cluster
\item demand - amount of demand for non-industry CO2 to be transported
\item variable_index.transport - the variable index for CO2_transported variable.
}
}
\description{
Gets the decision variable data for both \emph{CO2_transported} variables and
technologies that capture CO2. Non industry demand for CO2 transport is also
joined here to include in the constraint calculation. The two sets of decision
variables are joined together to allow for the formulation of the cluster to
storage site transport constraint.
}
\details{
CO2_transported variables represent the possible connections of cluster to
storage site for transporting CO2.
}
