% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_combining_constraints.R
\name{get_constraint_functions}
\alias{get_constraint_functions}
\title{Generate names of the constraint functions to run}
\usage{
get_constraint_functions(data)
}
\value{
vector of strings for the names of the functions to be run to produce
the matrix.
}
\description{
Generate names of the constraint functions to run
}
