% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_combining_constraints.R
\name{get_constraint_matrix}
\alias{get_constraint_matrix}
\title{Convert a set of constraints from a list to a sparse matrix format}
\usage{
get_constraint_matrix(constraint_group, n_decision_variables)
}
\arguments{
\item{constraint_group}{list of lists, for a set of constraints for a given
type of constraints, such as production constraints.
Each sub-list contains four elements:
\itemize{
\item column_indices (int) indicating the decision variables effected, varying
in length
\item values (num) indicating the coefficients for the variables effected, same length
as column_indices
\item direction (chr) indicating the direction of the constraint '==', '<=' or
'>='. Length 1.
\item rhs (num) the value of the right hand side of the constraint, length 1.
}}

\item{n_decision_variables}{integer, the total number of decision variables
in the model.}
}
\value{
list of 3 for the set of constraints where:
\itemize{
\item directions (chr) is a vector of directions for each constraints ('==', '<='
or >='). There is one direction for every row (one per constraint).
\item right hand side (num) is a vector of the rhs side values for each constraint.
There is one rhs value for each row.
\item constraint_matrix (dgCMatrix) a sparse matrix where values 'i' provides a
reference to a constraint as a row in the matrix, 'p' provides the
pointers for each column (decision variable) which is a smart
way of stating how many non-zero elements there are in each column.
By using i and p in combination, all of the locations of the non-zero
elements in the matrix can be found. 'v' provides the value of the
coefficient itself at a given location.
}
}
\description{
Reformat a set of constraints from pulling the indices of decision variables,
values of coeffecients, directions and rhs values and converting them into
another list which importantly contains a sparse matrix for the coeffecient
values. This allows for many constraints to be combined later without
maxing out computer memory.
}
