% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_fuel.R
\name{get_consuming_variables}
\alias{get_consuming_variables}
\title{Get decision variables that use commodities}
\usage{
get_consuming_variables(data, decision_variables)
}
\arguments{
\item{data}{list of tables generated by preprocessing the raw inputs from the
input spreadsheet using \code{comit_preprocess_inputs()}}

\item{decision_variables}{dataframe for the decision variables. This is the
first element of the list produced by \code{comit_objective_function}.}
}
\value{
dataframe, containing one row for each input used by a given variable.
Columns are:
\itemize{
\item variable_index
\item year
\item site_ID
\item code - technology code.
\item commodity - this is the commodity that is consumed by the technology.
\item input - numeric, amount of consumption of commodity per unit of output.
}
}
\description{
Return all variables that are both of \emph{used_capacity} type, and which are
technologies that use inputs, along with the data on what commodities are
used and in what volume.
}
