% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_comit_counterfactual_solver.R
\name{get_counterfactual_outputs}
\alias{get_counterfactual_outputs}
\title{Get Workbook for Counterfactual Results}
\usage{
get_counterfactual_outputs(solved_baseline)
}
\arguments{
\item{solved_baseline, }{list provided by \code{comit_counterfactual_solver()}
containing the problem solution and inputs.}
}
\value{
workbook object of the model outputs. This can be written to excel
using 'saveWorkbook()'.
}
\description{
Produces a workbook object of the counterfactual model outputs. This function
is not used by the comit app, but can be used for scripting another pipeline
or for testing the counterfactual seperately from the app.
}
