% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_finance.R
\name{get_deflator}
\alias{get_deflator}
\title{Helper function to pull correct deflator and given year}
\usage{
get_deflator(year, parameter_data)
}
\description{
Helper function to pull correct deflator and given year
}
