% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_comit_counterfactual_solver.R
\name{get_eep_rollout}
\alias{get_eep_rollout}
\title{Get EEP Rollout}
\usage{
get_eep_rollout(data)
}
\arguments{
\item{data}{list of dataframes containing input data, produced by
\code{comit_preprocess_inputs()}.}
}
\value{
dataframe with the scaled rollout for each technology.
}
\description{
This function processes the counterfactual rollout data.
}
