% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_attribution.R
\name{get_efficiencies}
\alias{get_efficiencies}
\title{Produce table resource and energy efficency factors}
\usage{
get_efficiencies(data)
}
\arguments{
\item{data}{list of dataframes, generated from passing the filepath for an
input spreadsheet to \code{read_excel_data_template()} and then passing the output
to \code{tidy()}.}
}
\value{
dataframe, containing one row for each combination of sector and year,
with the following columns:
\itemize{
\item efficiency_RE
\item efficiency_EE
\item REEE - combined RE and EE
\item REEE_adjust (1 - REEE, the factor to multiply by to get REEE adjusted emissions)
}
}
\description{
Get the RE and EE factors for each sector in each year.
}
