% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_attribution.R
\name{get_electricity_savings}
\alias{get_electricity_savings}
\title{Calculate Emissions Savings From Electricity}
\usage{
get_electricity_savings(
  scenario_energy,
  cf_energy,
  REEE_df,
  attr_params,
  emissions_attr,
  cf_energy_aggregated,
  cf_energy_sector_outputs,
  sector_outputs
)
}
\arguments{
\item{scenario_energy}{Data frame containing scenario energy data.}

\item{cf_energy}{Data frame containing counterfactual energy data.}

\item{REEE_df}{Data frame containing REEE data.}

\item{attr_params}{Data frame containing attribute parameters values.}

\item{emissions_attr}{Data frame containing emissions attribution data.}

\item{cf_energy_aggregated}{Data frame containing aggregated counterfactual energy data.}

\item{cf_energy_sector_outputs}{Data frame containing counterfactual energy sector outputs.}

\item{sector_outputs}{Data frame containing sector outputs.}
}
\value{
Data frame with total electricity savings by year and sector.
}
\description{
This function calculates the electricity savings by comparing scenario energy data
with counterfactual energy data, adjusting for REEE factors, and summarizing the results.
}
