% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_emissions.R
\name{get_emissions}
\alias{get_emissions}
\title{Calculate the emissions per unit of production for technology variables}
\usage{
get_emissions(
  technology_code,
  year,
  gas = NULL,
  source = NULL,
  capture = NULL,
  zero_emissions_from_biomass = TRUE,
  location = NULL,
  .data = data
)
}
\arguments{
\item{technology_code}{either a string or a character vector, of UKTM technology
codes.}

\item{year}{either an integer or a vector of integers for the year which
emissions of a technology are to be calculated. Emissions can vary by year
due to fuel emissions changing over time for some fuels.}

\item{gas}{character: 'CO2', 'nonCO2' or NULL (default). Indicates what type
of emission gas should be included. When NULL, both CO2 and non-CO2 gasses
are counted.}

\item{source}{character: 'process', 'fuel' or NULL (default). Indicates what
source of emissions should be counted. When NULL, both process and fuel
sourced emissions are counted.}

\item{capture}{boolean, default is NULL. Indicates whether the emissions that
are counted are those that are captured or not captured (i.e. released). When
NULL, both captured and not-captured emissions are counted.}

\item{zero_emissions_from_biomass}{boolean, default is TRUE. Whether to treat
biomass fuels as having 0 emissions.}

\item{location}{character: 'direct', 'indirect' or NULL (default). Indicates
whether to count direct or indirect emissions. When NULL, both direct and
indirect emissions are counted.}

\item{.data}{list of tables generated by preprocessing the raw inputs from the
input spreadsheet using \code{comit_preprocess_inputs()}. Default is the object
named 'data' which is used throughout comit code as the list of input tables.}
}
\value{
numeric, emissions in kilotonnes of CO2e for the specified technologies
and years.
}
\description{
Calculate the emissions per one unit of production for a given technology in
a given year. Parameters are used to specify what is included or excluded in
the emissions calculation.
}
