% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_attribution.R
\name{get_emissions_attr_df}
\alias{get_emissions_attr_df}
\title{Get Emissions Attribution Data Frame}
\usage{
get_emissions_attr_df(
  scenario_emissions,
  cf_emissions,
  filter_type,
  data,
  REEE_df
)
}
\arguments{
\item{scenario_emissions}{Data frame containing scenario emissions.}

\item{cf_emissions}{Data frame containing counterfactual emissions.}

\item{filter_type}{character, the type of emissions to attribute (either
'Direct' or 'Direct_and_Indirect").}

\item{data}{list of dataframes, generated from passing the filepath for an
input spreadsheet to \code{read_excel_data_template()} and then passing the output
to \code{tidy()}.}

\item{REEE_df}{Data frame containing REEE data.}
}
\value{
Data frame with emissions attribution data, for each sector and year.
Columns include:
\itemize{
\item scenario_emissions - total emissions from scenario
\item cf_emissions - total emissions from emissions run
\item emissions_reductions - the amount of emissions reduced when comparing the
counterfactual to the scenario.
\item cf_emissions_re - total emissions reductions due to re
\item cf_emissions_ee - total emissions reductions due to ee
\item emissions_reductions_post_reee - total emissions reductions after
accounting for re and ee.
}
}
\description{
This function calculates the emissions attribution data frame by summarizing
scenario and counterfactual emissions, and then joining them with the REEE data frame.
}
