% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_attribution.R
\name{get_emissions_from_fuel_use}
\alias{get_emissions_from_fuel_use}
\title{Calculate Emissions from Fuel Use}
\usage{
get_emissions_from_fuel_use(cf_energy, attr_params)
}
\arguments{
\item{cf_energy}{A dataframe containing energy consumption data from a
counterfactual run.}

\item{attr_params}{A dataframe of parameters for the attribution anaylsis,
containing emission factors for different fuel categories.}
}
\value{
A data frame with aggregated energy consumption and calculated
emissions by year, primary output, and fuel category.
}
\description{
This function calculates emissions from fuel use based on energy consumption
and emission factors.
}
