% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_process_sites.R
\name{get_emissions_from_share}
\alias{get_emissions_from_share}
\title{Apportion emissions to traded and non-traded point sites}
\usage{
get_emissions_from_share(df)
}
\arguments{
\item{df}{dataframe containing columns sector_share_by_traded and
point_sites_emissions_from_share.}
}
\value{
The input dataframe with the column emissions_MtCO2 appended.
}
\description{
The emissions apportioned are based on the site's proportion contributed to the
sector and traded_status combination overall multiplied by the total emissions
for the given sector and traded status.
}
