% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_attribution.R
\name{get_emissions_total}
\alias{get_emissions_total}
\title{Calculate Total Emissions}
\usage{
get_emissions_total(
  residual_emissions,
  emissions_reductions,
  beccs_and_ccs_savings
)
}
\arguments{
\item{residual_emissions}{Data frame containing residual emissions data.}

\item{emissions_reductions}{Data frame containing emissions reductions data.}

\item{beccs_and_ccs_savings}{Data frame containing savings from BECCS and CCS.}
}
\value{
Data frame containing the total emissions savings for each sector, year
and attribution type.
}
\description{
This function calculates the total emissions by combining residual emissions,
emissions reductions, and savings from BECCS and CCS.
}
