% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_emissions.R
\name{get_emissions_variables}
\alias{get_emissions_variables}
\title{Get emissions data for used capacity variables to set up emissions constraint}
\usage{
get_emissions_variables(data, decision_variables)
}
\arguments{
\item{data}{list of tables generated by preprocessing the raw inputs from the
input spreadsheet using \code{comit_preprocess_inputs()}}

\item{decision_variables}{dataframe for the decision variables. This is the
first element of the list produced by \code{comit_objective_function}.}
}
\value{
dataframe, with one row per used_capacity decision variable.
Columns include:
\itemize{
\item emissions_per_used_capacity, numeric, kt of CO2e
\item emissions_limit, numeric, the annual maximum emissions/
}
}
\description{
Gets the amount of emissions per used capacity of the used_capacity
decision variables, as well at the emissions limit for the relevant year
from the input data.
}
