% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_starting_values.R
\name{get_existing_techs}
\alias{get_existing_techs}
\title{Create table of starting commodity data}
\usage{
get_existing_techs(data, rollout = NULL)
}
\arguments{
\item{data}{list of dataframes, the input data produced from
\code{read_excel_data_template()} after some initial tidying, including by
\code{interpolate_data()}, \code{tidy()} and \code{round_years()}.}
}
\value{
dataframe of the unique input and output commidity pairs. One row per
pair and columns for:
\itemize{
\item output_commodity
\item commodity
\item average_output (average output across all technologies that produce the given
commodity that is used to create the output_commodity).
\item existing (total output of all technologies that produce the given commodity
that is used to create the output_commidity).
}
}
\description{
Gets the technologies with existing capacitiy at the beginning of the modelled
period and then joins on commodity variables to then reduce down to unique
input commodity and output commodity pairs. The average and total capacity
of each input_commodity and output_commodity pair across the different
technologies is calculated.
}
