% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_starting_values.R
\name{get_existing_techs_matrix}
\alias{get_existing_techs_matrix}
\title{Produce the matrix of coefficients required to solve the existing capacity
system of linear equations}
\usage{
get_existing_techs_matrix(data, rollout = NULL)
}
\arguments{
\item{data}{list of dataframes, the input data produced from
\code{read_excel_data_template()} after some initial tidying, including by
\code{interpolate_data()}, \code{tidy()} and \code{round_years()}.}

\item{rollout, }{dataframe, default is NULL. Optional argument to provide a
dataframe for the technology rollout in a given year, used for generating
the counterfactual. The operations are very similar but change slightly
when this dataframe is provided, to focus on capacity in a given year rather
than just initial starting capacity. The rollout dataframe should be filtered
to a single period and should contain a column for the technology code (code)
and a columns for scaled_output.}
}
\value{
matrix, the coefficient values. Rows represent the input commodities used
to produce the final output commodities. Columns represent the final output
commodities. The coefficient values themselves are the average output of final
commodities produced by the individual commodities.
}
\description{
Produces the coefficients matrix (A) used in the 'Ax = b' equation that solves
a system of linear equations problem for finding the required output commodities
required to meet demand in the start year.
}
