% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_fuel.R
\name{get_fuel_constrained_variables}
\alias{get_fuel_constrained_variables}
\title{Get the variables that are to be constrained by the fuel constraint}
\usage{
get_fuel_constrained_variables(data, consuming_variables, constraint_type)
}
\arguments{
\item{data}{list of tables generated by preprocessing the raw inputs from the
input spreadsheet using \code{comit_preprocess_inputs()}}

\item{consuming_variables}{dataframe, all decision variables that consume
commodities, as produced by \code{get_consuming_variables()}.}

\item{constraint_type}{character, either 'max' or 'min' depending on whether
the limit to be set is on maximum or minimum fuel use.}
}
\value{
dataframe, with one row per variable and commodity combination for
all of the variables involved in the specified fuel_constraint. Multiple
occurances of the same variable index are permitted as some variables will
be relevant to multiple fuel constraints where they have multiple fuel inputs.
Note that 'fuel' and 'group' columns are maintained to later allow for grouping
at the level at which the constraint was specified, however 'commodity'
determines the fuel which is being constrained in the row itself.
}
\description{
Reduce the consuming variables down to only those which are to relevant to the
constraint. Pulls the input tab relating to fuel constraints and then joins
the relevant commodities for constraints specified at the fuel group rather
than commodity level. This data is then joined back to the original
consuming_variables data provided and reduced down to only the variables
required.
}
