% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_fuel_share.R
\name{get_fuel_consuming_variables}
\alias{get_fuel_consuming_variables}
\title{Get decision variables that use fuel commodities}
\usage{
get_fuel_consuming_variables(
  data,
  decision_variables,
  fuel_commodities,
  max_fuel_share_df
)
}
\arguments{
\item{data}{list of tables generated by preprocessing the raw inputs from the
input spreadsheet using \code{comit_preprocess_inputs()}}

\item{decision_variables}{dataframe for the decision variables. This is the
first element of the list produced by \code{comit_objective_function}.}

\item{fuel_commodities}{dataframe containing all fuel commodities. Contains
two columns: commodity (commodity code) and group (fuel category).}

\item{max_fuel_share_df}{dataframe for the fuel share constraints information,
produced by \code{get_raw_max_fuel_share_data()}.}
}
\value{
dataframe, with one row for each fuel input for each decision variable.
Columns include:
\itemize{
\item variable_index
\item year
\item site_ID
\item code - technology code
\item commodity - input commodity that is consumed by the technology
\item input - numeric, amount of consumption of commodity per unit of output
\item output_commodity - the commodity produced by the technology
\item sector
\item group - fuel category
\item share_limit - numeric, from the fuel share constraint inputs
}
}
\description{
Return data for all variables of \emph{used_capacity} type which are technologies
that use fuel commodities as inputs. Also joins on fuel share
constraint input data.
}
