% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_fuel.R
\name{get_fuel_contraint_tab}
\alias{get_fuel_contraint_tab}
\title{Fetch the fuel constraint input table from the data list}
\usage{
get_fuel_contraint_tab(data, constraint_type)
}
\arguments{
\item{data}{list of tables generated by preprocessing the raw inputs from the
input spreadsheet using \code{comit_preprocess_inputs()}}

\item{constraint_type}{character, either 'max' or 'min' depending on whether
the limit to be set is on maximum or minimum fuel use.}
}
\value{
dataframe with one row per fuel commodity/group for each year, for
all fuels/commodities specified to be constrained in the inputs. Columns are:
\itemize{
\item year
\item fuel - string, the commodity code of a fuel, will be NA if group is provided
instead.
\item group - string, the fuel category. Used instead of a commodity code (fuel
column), hence will be NA if commodity code provided.
\item apply_to_industry_only - boolean, whether fuel used by non-industry sectors
(e.g. refineries and hydrogen) should be included in the constraint.
\item fuel_limit - the value to limit the amount of fuel that can be used.
}
}
\description{
Fetch the fuel constraint input table from the data list
}
