% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_fuel_share.R
\name{get_fuel_share_constraint_combinations}
\alias{get_fuel_share_constraint_combinations}
\title{Get the distinct groups for setting fuel share constraints.}
\usage{
get_fuel_share_constraint_combinations(
  fuel_consuming_variables,
  max_fuel_share_df,
  data
)
}
\arguments{
\item{fuel_consuming_variables}{dataframe for the fuel consuming
decision variables.}

\item{max_fuel_share_df}{dataframe for the fuel share constraints information,
produced by \code{get_raw_max_fuel_share_data()}.}

\item{data}{list of tables generated by preprocessing the raw inputs from the
input spreadsheet using \code{comit_preprocess_inputs()}}
}
\value{
dataframe, with as many rows as there are to be fuel share constraints.
There is one constraint per possible combination of sector, fuel category,
output_commodity and year. Columns are:
\itemize{
\item sector
\item group (fuel category)
\item output_commodity
\item year
}
}
\description{
Create a dataframe of all distinct possible combinations of sector, group (fuel
category), output_commodity and year for the fuel share constraints.
}
\details{
The combinations are found by finding the sector, fuel categories and output
commodity combinations that are present in both the fuel consuming variables
data and the fuel share constraint input data.
}
