% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_hydrogen.R
\name{get_H2_availability_data}
\alias{get_H2_availability_data}
\title{Get the amount of hydrogen available to technologies in each cluster, in each
year}
\usage{
get_H2_availability_data(data, decision_variables)
}
\arguments{
\item{data}{list of tables generated by preprocessing the raw inputs from the
input spreadsheet using \code{comit_preprocess_inputs()}}

\item{decision_variables}{dataframe for the decision variables. This is the
first element of the list produced by \code{comit_objective_function}.}
}
\value{
dataframe, with one row per year and cluster combination. Contains
the following columns:
\itemize{
\item year
\item cluster
\item net_availability (the amount of hydrogen available for industry use)
}
}
\description{
Pulls the data on overall H2 availability and non-industry H2 demand from
the input data and then calculates the net amount of hydrogen available to
be used for industry (net_availability = overall_availability - non_industry_demand).
}
