% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_hydrogen.R
\name{get_H2_consuming_variables}
\alias{get_H2_consuming_variables}
\title{Return all hydrogen consuming decision variables}
\usage{
get_H2_consuming_variables(data, decision_variables)
}
\arguments{
\item{data}{list of tables generated by preprocessing the raw inputs from the
input spreadsheet using \code{comit_preprocess_inputs()}}

\item{decision_variables}{dataframe for the decision variables. This is the
first element of the list produced by \code{comit_objective_function}.}
}
\value{
dataframe, with one row per decision variable and the following
columns:
\itemize{
\item variable_index.technology - the variable index. '.technology' suffix is
used to distinguish the source of data when joins are used later.
\item year
\item site_ID
\item input
}
}
\description{
Get all technology decision variables that use hydrogen as an input fuel.
}
