% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_decision_variables.R
\name{get_H2_parameters}
\alias{get_H2_parameters}
\title{Set H2 parameters based on read data}
\usage{
get_H2_parameters(data)
}
\arguments{
\item{data}{list of data tables read in from excel data template, after some
initial processing from preceding functions in \code{\link[=comit_solver]{comit_solver()}}.}
}
\value{
list of Boolean values (length 2).
First element contains parameter for whether to
model H2 production, second contains model for whether to enforce H2 plant
size.
Prints warning when condition met.
}
\description{
Also warns user when H2 parameters conflict.
}
