% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_hydrogen.R
\name{get_H2_variables_with_availability}
\alias{get_H2_variables_with_availability}
\title{Join hydrogen availability data to the hydrogen decision variables}
\usage{
get_H2_variables_with_availability(
  decision_variables,
  H2_variables,
  H2_availability_data
)
}
\arguments{
\item{decision_variables}{dataframe for the decision variables. This is the
first element of the list produced by \code{comit_objective_function}.}

\item{H2_variables}{dataframe for the hydrogen decision variables, produced
by \code{get_H2_availability_data()}.}

\item{H2_availability_data}{dataframe for the hydrogen available in each cluster and
in each year. Produced by \code{get_H2_availability_data()}.}
}
\value{
dataframe, one row per hydrogen technology decision variable. Columns
include:
\itemize{
\item variable_index.technology
\item year
\item site_ID
\item cluster
\item input
\item net_availability
}
}
\description{
Joins on the data for the amounts of hydrogen available in each cluster and
in each year to the relevant decision variables, to allow setting up the
constraint later.
}
