% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_hydrogen.R
\name{get_hydrogen_conversion_constraint_data}
\alias{get_hydrogen_conversion_constraint_data}
\title{Get the data required to generate the hydrogen conversion constraint}
\usage{
get_hydrogen_conversion_constraint_data(
  data,
  decision_variables,
  hydrogen_conversion_variables
)
}
\arguments{
\item{data}{list of tables generated by preprocessing the raw inputs from the
input spreadsheet using \code{comit_preprocess_inputs()}}

\item{decision_variables}{dataframe for the decision variables. This is the
first element of the list produced by \code{comit_objective_function}.}

\item{hydrogen_conversion_variables}{dataframe of the hydrogen_conversion
decision variables, produced by \code{get_hydrogen_conversion_variables()},}
}
\value{
dataframe, with one row per combination of hydrogen consuming
decision variables and each hydrogen conversion variable, in each year.
Columns include:
\itemize{
\item variable_index.consumption - variable index for the hydrogen consumption variables
}
}
\description{
Joins data from hydrogen consumption and hydrogen conversion decision
variables to allow the hydrogen conversion constraint to be formulated.
}
