% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_hydrogen.R
\name{get_hydrogen_conversion_variables}
\alias{get_hydrogen_conversion_variables}
\title{Get decision variables for hydrogen conversion technologies}
\usage{
get_hydrogen_conversion_variables(data, decision_variables)
}
\arguments{
\item{data}{list of tables generated by preprocessing the raw inputs from the
input spreadsheet using \code{comit_preprocess_inputs()}}

\item{decision_variables}{dataframe for the decision variables. This is the
first element of the list produced by \code{comit_objective_function}.}
}
\value{
dataframe, with one row per hydrogen conversion variable. Columns
include:
\itemize{
\item variable_index.production - the variable index, with 'production' suffix
to distinguish from consumption variables in later join.
\item year
\item code - technology code
}
}
\description{
Get decision variables for hydrogen conversion technologies
}
