% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_attribution.R
\name{get_hydrogen_savings}
\alias{get_hydrogen_savings}
\title{Calculate Emissions Savings From Hydrogen}
\usage{
get_hydrogen_savings(scenario_energy, cf_energy_sector_outputs, REEE_df)
}
\arguments{
\item{scenario_energy}{Data frame containing scenario energy data.}

\item{cf_energy_sector_outputs}{Data frame containing counterfactual energy
sector outputs.}

\item{REEE_df}{Data frame containing REEE data.}
}
\value{
Data frame with hydrogen savings by year and sector.
}
\description{
This function calculates the hydrogen savings by summarizing the energy data
for hydrogen, joining it with counterfactual energy sector outputs, and then
calculating the savings based on emissions intensity.
}
