% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_decision_variables.R
\name{get_hydrogen_sites}
\alias{get_hydrogen_sites}
\title{Produce hydrogen data for each site}
\usage{
get_hydrogen_sites(data, site_technologies)
}
\arguments{
\item{data}{list of data tables read in from excel data template, after some
initial processing from preceding functions in \code{\link[=comit_solver]{comit_solver()}}.}

\item{site_technologies}{data frame containing both site and technology data.}
}
\value{
dataframe containing the following information for each site and year:
\itemize{
\item H2_point
\item hydrogen_start
\item lowest_cost_option.H2
}
}
\description{
Generates hydrogen related variables for each combination of site and year for
all sites that have at least 1 year present that is after the hydrogen start
year at that site. This table can then be used to create the hydrogen related
decision variables.
}
