% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_decision_variables.R
\name{get_hydrogen_sites_transport}
\alias{get_hydrogen_sites_transport}
\title{Create the decision variables for hydrogen transport types for sites}
\usage{
get_hydrogen_sites_transport(data, hydrogen_sites)
}
\arguments{
\item{data}{list of data tables read in from excel data template, after some
initial processing from preceding functions in \code{\link[=comit_solver]{comit_solver()}}.}

\item{hydrogen_sites}{dataframe, output from \code{\link[=get_hydrogen_sites]{get_hydrogen_sites()}}.}
}
\value{
dataframe for decision variables regarding transport of hydrogen
to specific sites.
}
\description{
Includes the variables for grid, pipes and trucks and combines them into a
single data frame.
}
