% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_CO2.R
\name{get_injection_constrained_variables}
\alias{get_injection_constrained_variables}
\title{Get data for the decision variables to be constrained by CO2 storage injection
rates}
\usage{
get_injection_constrained_variables(data, decision_variables)
}
\arguments{
\item{data}{list of tables generated by preprocessing the raw inputs from the
input spreadsheet using \code{comit_preprocess_inputs()}}

\item{decision_variables}{dataframe for the decision variables. This is the
first element of the list produced by \code{comit_objective_function}.}
}
\value{
dataframe, with one row per \emph{CO2_transported} variable, in each year
to be constrained. Columns include:
\itemize{
\item variable_index
\item year
\item storage_site
\item max_injection
}
}
\description{
Get the \emph{CO2_transported} variables which are the decision variables for
cluster to storage site CO2 transport, along with the associated injection
limits for the storage site recieving the CO2. Variables with limits of
1,000,000 or over are removed and will not be constrained.
}
