% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_process_sites.R
\name{get_inside_vs_outside_cluster_ratios}
\alias{get_inside_vs_outside_cluster_ratios}
\title{Produce a table for the proportion of sites inside the cluster boundary for
each sector in each cluster}
\usage{
get_inside_vs_outside_cluster_ratios(input_data)
}
\arguments{
\item{input_data}{data read in from excel template, same as what is passed to
\code{\link[=comit_solver]{comit_solver()}}.}
}
\value{
dataframe with columns for sector, cluster an in_ratio which gives
the proportion of sites inside the cluster boundary for the given sector
and cluster combination. There is one row for every cluster and sector
combination.
}
\description{
This table can then be used to impute how many sites for each sector should
be found within the cluster for non-point source sites.
}
