% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_decision_variables.R
\name{get_inter_cluster_hydrogen}
\alias{get_inter_cluster_hydrogen}
\title{Create the decision variables for hydrogen transport between clusters}
\usage{
get_inter_cluster_hydrogen(data, hydrogen_sites)
}
\arguments{
\item{data}{list of data tables read in from excel data template, after some
initial processing from preceding functions in \code{\link[=comit_solver]{comit_solver()}}.}

\item{hydrogen_sites}{dataframe, output from \code{\link[=get_hydrogen_sites]{get_hydrogen_sites()}}.}
}
\value{
dataframe for decision variables regarding transport of hydrogen
between clusters.
}
\description{
Includes the variables for both new and available capacity as well as H2
outflows.
}
