% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_process_sites.R
\name{get_large_point_sites}
\alias{get_large_point_sites}
\title{Create table for large point sites}
\usage{
get_large_point_sites(df, input_data)
}
\arguments{
\item{input_data}{data read in from excel template, same as what is passed to
\code{\link[=comit_solver]{comit_solver()}}.}

\item{dataframe}{containing the NAEI data after processed by
\code{\link[=append_information_to_sites]{append_information_to_sites()}}}
}
\value{
input dataframe now manipulated and reduced to the following columns,
for large point sites only:
\itemize{
\item site_name
\item IPM_sector
\item H2_point
\item total_MtCO2 (emissions)
\item pipe_dist
\item num_sites
\item Latitude
\item Longitude
\item traded_flag
}
}
\description{
Filter for the large point sites and then get emissions and drop columns that
are no longer required.
}
