% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_CO2.R
\name{get_max_CCS_data}
\alias{get_max_CCS_data}
\title{Get decision variables to be constrained by the max_CCS constraint}
\usage{
get_max_CCS_data(data, decision_variables)
}
\arguments{
\item{data}{list of tables generated by preprocessing the raw inputs from the
input spreadsheet using \code{comit_preprocess_inputs()}}

\item{decision_variables}{dataframe for the decision variables. This is the
first element of the list produced by \code{comit_objective_function}.}
}
\value{
dataframe, with one row per decision variable to be constrained.
Columns include:
\itemize{
\item variable_index
\item year
\item captured_CO2 - amount of CO2 captured per unit of capacity
\item max_CO2_captured - total amount of CO2 that can be captured in given year
}
}
\description{
Get the data required to formulate the max_CCS constraint, by finding the
relevant decision variables and joining the relevant CCS limit from the
input data.
}
