% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_decision_variables.R
\name{get_min_hydrogen_plant_size_variables}
\alias{get_min_hydrogen_plant_size_variables}
\title{Create binary variables to enforce minimum H2 plant size}
\usage{
get_min_hydrogen_plant_size_variables(data)
}
\arguments{
\item{data}{list of data tables read in from excel data template, after some
initial processing from preceding functions in \code{\link[=comit_solver]{comit_solver()}}.}
}
\value{
dataframe containing the decision variables for the minimum H2 plant
size.
}
\description{
Creates a binary variable (0, 1) for each hydrogen producing
}
