% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_utilities.R
\name{get_my_colours}
\alias{get_my_colours}
\title{Default colour palette for runs}
\usage{
get_my_colours(data)
}
\arguments{
\item{data}{Data frame that includes a \code{run} column}
}
\value{
Named character vector of colours
}
\description{
Returns a named vector mapping \code{run} to colours.
}
