% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_decision_variables.R
\name{get_non_industry_hydrogen}
\alias{get_non_industry_hydrogen}
\title{Create the decision variables for non industrial hydrogen demand}
\usage{
get_non_industry_hydrogen(data)
}
\arguments{
\item{data}{list of data tables read in from excel data template, after some
initial processing from preceding functions in \code{\link[=comit_solver]{comit_solver()}}.}
}
\value{
dataframe for decision variables regarding transport of hydrogen
between clusters.
}
\description{
Create the decision variables for non industrial hydrogen demand
}
