% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_process_sites.R
\name{get_non_point_site_values_for_ratios}
\alias{get_non_point_site_values_for_ratios}
\title{Produce a table containing imputed site data for non-point sites}
\usage{
get_non_point_site_values_for_ratios(df, input_data)
}
\arguments{
\item{df}{dataframe produced in \code{\link[=non_point_sites_from_ratios]{non_point_sites_from_ratios()}} after
running \code{\link[=get_sector_emission_totals]{get_sector_emission_totals()}} and filtering for non-traded sites.}

\item{input_data}{data read in from excel template, same as what is passed to
\code{\link[=comit_solver]{comit_solver()}}.}
}
\value{
dataframe with a row for each of the artificial sites that represent
the aggregated non-point source sites. There should be as many sites as there
are combinations of sectors and clusters (minus the ones not used) - multiplied
by two if the Two_nps_sites parameter is set to TRUE. Columns include:
\itemize{
\item total_MtCO2 (imputed emissions)
\item pipe_dist (imputed pipe distance)
\item Latitude and Longitude (imputed site coordinates)
}
}
\description{
Creates the artificial sites used to model the non-point sites data,
imputing the site characteristics from the data we have available from
point sites and IDBR.
}
