% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_process_sites.R
\name{get_non_point_sites}
\alias{get_non_point_sites}
\title{Produce the initial table for non point source sites}
\usage{
get_non_point_sites(input_data)
}
\arguments{
\item{input_data}{data read in from excel template, same as what is passed to
\code{\link[=comit_solver]{comit_solver()}}.}
}
\value{
dataframe containing 1 row per sector. Columns are sic code, sector
classification and then a column for each region. Values are the number
sites in each region of the given sic code.
}
\description{
Finds the number of non point source sites (i.e. sites not in NAEI data) that
are in each combination of region and sic codes.
}
