% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_process_sites.R
\name{get_proportion_of_emissions_from_sector}
\alias{get_proportion_of_emissions_from_sector}
\title{Produce a table for the proportion of sites in each cluster which belong to
the different sectors}
\usage{
get_proportion_of_emissions_from_sector(input_data)
}
\arguments{
\item{input_data}{data read in from excel template, same as what is passed to
\code{\link[=comit_solver]{comit_solver()}}.}
}
\value{
dataframe with columns for cluster, sector and sector_share which is
the proportion of emissions belonging to the sector in each cluster. Should
have one row per sector per cluster.
}
\description{
Produce a table for the proportion of sites in each cluster which belong to
the different sectors
}
