% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_pv_generic_functions.R
\name{get_pv_functions}
\alias{get_pv_functions}
\title{Generate names of the present value functions to run}
\usage{
get_pv_functions(data)
}
\value{
vector of strings for the names of the functions to be run to produce
the decision variable coeffecients in the objective function.
}
\description{
Generate names of the present value functions to run
}
