% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_fuel_share.R
\name{get_raw_max_fuel_share_data}
\alias{get_raw_max_fuel_share_data}
\title{Get the max fuel share limit values from the input data}
\usage{
get_raw_max_fuel_share_data(data, commodities)
}
\arguments{
\item{data}{list of tables generated by preprocessing the raw inputs from the
input spreadsheet using \code{comit_preprocess_inputs()}}

\item{commodities}{dataframe for the commodities data.}
}
\value{
dataframe, with one row per input for each possible input commodity,
output_commodity and fuel type combination that is to be constrained. Columns
include:
\itemize{
\item sector
\item output_commodity
\item group (fuel category)
\item share_limit
\item commodity (inputs)
}
}
\description{
Gets the max fuel share constraint data used to formulate the constraint.
Reads in the table from the input, then tidies into a longer format and joins
fuel category information from the commodities data.
}
