% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_starting_values.R
\name{get_required_capacities}
\alias{get_required_capacities}
\title{Calculate the required amount of each commodity}
\usage{
get_required_capacities(data, existing_techs_matrix)
}
\arguments{
\item{data}{list of dataframes, the input data produced from
\code{read_excel_data_template()} after some initial tidying, including by
\code{interpolate_data()}, \code{tidy()} and \code{round_years()}.}

\item{existing_techs_matrix}{matrix containing the average output values for
each commodity. These are the coeffecients for the system of linear
equations, produced by \code{get_existing_techs_matrix()}.}
}
\value{
dataframe with one row per commodity and two columns:
\itemize{
\item output_commodity (name of the commodity)
\item required_commodity (calculated required demand)
}
}
\description{
Calculate the required amount of each commodity
}
