% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_attribution.R
\name{get_residual_emissions}
\alias{get_residual_emissions}
\title{Calculate Residual Emissions}
\usage{
get_residual_emissions(emissions_attr)
}
\arguments{
\item{emissions_attr}{Data frame containing emissions attribution data.}
}
\value{
Data frame containing the residual emissions.
}
\description{
This function calculates the residual emissions based on attribute emissions
data. This is the amount of emissions remaining (i.e. emissions not saved).
}
