% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_utilities.R
\name{get_run_name}
\alias{get_run_name}
\title{Normalise a workbook file name into a run name}
\usage{
get_run_name(file_name)
}
\arguments{
\item{file_name}{Character file name}
}
\value{
Character run name
}
\description{
Strips common prefixes/suffixes to produce a clean \code{run} label.
}
\keyword{internal}
