% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_attribution.R
\name{get_scaled_emissions}
\alias{get_scaled_emissions}
\title{Calculate Scaled Emissions}
\usage{
get_scaled_emissions(emissions_attr, emissions_reductions, data)
}
\arguments{
\item{emissions_attr}{Data frame containing attribution data.}

\item{emissions_reductions}{Data frame containing emissions reductions data.}

\item{data}{list of dataframes, generated from passing the filepath for an
input spreadsheet to \code{read_excel_data_template()} and then passing the output
to \code{tidy()}.}
}
\value{
Data frame containing the scaling factor for emissions reductions in
each year and sector.
}
\description{
This function calculates the scaling factor for emissions reductions based on
attribute emissions and total emissions reductions.
}
