% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_attribution.R
\name{get_sector_commodity_lookup}
\alias{get_sector_commodity_lookup}
\title{Get a Commodity to Sector Lookup}
\usage{
get_sector_commodity_lookup(data)
}
\arguments{
\item{data}{list of dataframes, generated from passing the filepath for an
input spreadsheet to \code{read_excel_data_template()} and then passing the output
to \code{tidy()}.}
}
\value{
A dataframe with all distinct sector and commodity pairs.
}
\description{
This function retrieves a lookup table for sectors and commodities.
}
