% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_process_sites.R
\name{get_sector_emission_totals}
\alias{get_sector_emission_totals}
\title{Produce emissions figures by sector and traded status}
\usage{
get_sector_emission_totals(df, input_data)
}
\arguments{
\item{df}{dataframe containing columns for IPM_sector, traded_flag and
emissions_MtCO2.}

\item{input_data}{data read in from excel template, same as what is passed to
\code{\link[=comit_solver]{comit_solver()}}.}
}
\value{
dataframe, with two rows per sector (one for each traded_status).
The output columns are the same as that from \code{\link{sum_emissions_by_sector}}
with additional columns added, including point_sites_emissions_from_share
which gives the total emissions for point sites in the respective sector and
traded status combination.
}
\description{
Get total emissions by sector and traded/non-traded split, in order to allow
allocation of emissions later.
}
