% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_attribution.R
\name{get_sector_outputs}
\alias{get_sector_outputs}
\title{Get Sector Outputs Lookup}
\usage{
get_sector_outputs(data)
}
\arguments{
\item{data}{list of dataframes, generated from passing the filepath for an
input spreadsheet to \code{read_excel_data_template()} and then passing the output
to \code{tidy()}.}
}
\value{
Data frame of the unique sector and outputs combinations/
}
\description{
This function provides all unique combinations of sector and primary outputs
from the input technologies data.
}
