% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_decision_variables.R
\name{get_site_technologies}
\alias{get_site_technologies}
\title{Produce table for site/technology combinations}
\usage{
get_site_technologies(data)
}
\arguments{
\item{data}{list of data tables read in from excel data template, after some
initial processing from preceding functions in \code{\link[=comit_solver]{comit_solver()}}.}
}
\value{
data frame containing data for all possible combinations of sites and
technologies.
}
\description{
Get data for site/technology combinations as a prerequisite for creating
capacity variables.
}
