% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_decision_variables.R
\name{get_site_technologies_filter}
\alias{get_site_technologies_filter}
\title{Produce filter to remove non-required rows from site_technologies table}
\usage{
get_site_technologies_filter(site_technologies_full)
}
\arguments{
\item{site_technologies_full}{dataframe containing the full set of combinations
of all sites, technologies and years.}
}
\value{
Large logical - TRUE when row is to be kept.
}
\description{
This is a helper function used for filtering in \code{\link[=get_site_technologies]{get_site_technologies()}}.
}
