% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_process_sites.R
\name{get_small_point_site_data}
\alias{get_small_point_site_data}
\title{Retrieve small point sites from NAEI data and tidy}
\usage{
get_small_point_site_data(df, input_data)
}
\arguments{
\item{input_data}{data read in from excel template, same as what is passed to
\code{\link[=comit_solver]{comit_solver()}}.}
}
\value{
input dataframe filtered for small sites and with additional columns
for emission and cluster data.
}
\description{
Filters for the small point sites from the NAEI data and then does some
manipulation to get the required variables, inlcuding calculating emissions
from share and determining whether a site is within the cluster radius.
}
