% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_process_sites.R
\name{get_small_point_sites_filter}
\alias{get_small_point_sites_filter}
\title{Produce variable to flag small point sites}
\usage{
get_small_point_sites_filter(df)
}
\value{
list of logical, TRUE when sites should be considered as small. Should
be same length as input dataframe.
}
\description{
Sites are determined as small if they are both non-traded and their emissions
are below a threshold specified elsewhere. This is done so that small sites
can be aggregated to reduce the complexity of the model and increase its speed.
}
