% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_production.R
\name{get_technologlies_that_use_outputs}
\alias{get_technologlies_that_use_outputs}
\title{Get technologies that use commodities from other technologies}
\usage{
get_technologlies_that_use_outputs(data)
}
\arguments{
\item{data}{list of tables generated by preprocessing the raw inputs from the
input spreadsheet using \code{comit_preprocess_inputs()}}
}
\value{
dataframe with one row per technology/commodity combination for all
of the technologies that use commodities produced by other technologies.
There are three columns:
\itemize{
\item code - character, the technology code
\item input_commodity - character, the commodity used as an input
\item input_amount - numeric, the amount of input commodity used per unit of the
technology.
}
}
\description{
Gets the technology data for technologies which use commodities produced by
other technology processes. This is because we need to make sure that each
technology is producing enough commodity to satisfy the inputs to other
processes at the site.
}
