% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_capacities.R
\name{get_technology_capacity_tab}
\alias{get_technology_capacity_tab}
\title{Get the relevant input table containing capacity limits}
\usage{
get_technology_capacity_tab(data, capacity_type)
}
\arguments{
\item{data}{list of tables generated by preprocessing the raw inputs from the
input spreadsheet using \code{comit_preprocess_inputs()}}

\item{capacity_type}{character, either 'max' or 'min' depending on whether
the limit to be set is on maximum or minimum capacity.}
}
\value{
dataframe containing one row for each year for each of the
technologies to be restricted, and three columns:
\itemize{
\item year
\item code (technology code)
\item capacity_limit - either the minimum or the maximum capacity depending on
the capacity_type.
}
}
\description{
Get the relevant input table containing capacity limits
}
