% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_read_data.R
\name{get_template_sheet_names}
\alias{get_template_sheet_names}
\title{Get the names of relevant input template sheets}
\usage{
get_template_sheet_names(path, sections)
}
\arguments{
\item{path}{character for the file location of the workbook}

\item{sections}{vector of characters for the section headers}
}
\value{
a vector of the sheet names to be read in for comit.
}
\description{
Returns the names of all input workbook sheets that are required. Section
headers (suffixed with '==>') are removed as well as the sheets before the
first header and the sheets after the last header, as these are not required.
}
