% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_emissions.R
\name{get_total_emissions}
\alias{get_total_emissions}
\title{Calculate total emissions for each technology in each year for all relevant
types of emissions}
\usage{
get_total_emissions(
  aggregated_emissions,
  gas_included,
  source_included,
  capture_included,
  location_included
)
}
\arguments{
\item{aggregated_emissions}{dataframe as produced by \code{sum_emission_types()}.}

\item{gas_included}{string or character vector for the gas types to include.
Allowed values are 'CO2' or 'nonCO2'.}

\item{source_included}{string or character vector for the gas types to include.
Allowed values are 'fuel' or 'process'.}

\item{capture_included}{boolean or boolean vector for the gas types to include.}

\item{location_included}{string or character vector for the gas types to include.
Allowed values are 'direct' or 'indirect'.}
}
\value{
dataframe with one row for each technology in each year. Columns are:
\itemize{
\item code
\item year
\item CO2e - amount of CO2e emitted (kt) per unit of used capacity
}
}
\description{
Calculate total emissions for each technology in each year for all relevant
types of emissions
}
