% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_process_sites.R
\name{get_values_for_two_non_point_sites}
\alias{get_values_for_two_non_point_sites}
\title{Transform the dataframe to account for two non-point sites in each cluster
and sector combination}
\usage{
get_values_for_two_non_point_sites(df)
}
\arguments{
\item{df}{dataframe produce inside of \code{\link[=get_non_point_site_values_for_ratios]{get_non_point_site_values_for_ratios()}}
function, that contains artificial sites for the non-point sources. One row
per sector and cluster combination.}
}
\value{
dataframe with twice as many rows (sites) as the input dataframe.
Columns 'total_MtCO2' (emissions), 'cluster_in_out' and 'pipe_dist' added.
}
\description{
Doubles the number of rows so that there are now two sites per sector and
cluster combination - one for inside cluster sites and another for outside
of cluster sites. The emissions and pipe distances are also set at that point
based on whether the site is inside or outside of the cluster.
}
