% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_capacity_transfer.R
\name{get_variable_index_splits}
\alias{get_variable_index_splits}
\title{Create (almost) even groups of index numbers to set constraints in groups,
reducing the amount of RAM required}
\usage{
get_variable_index_splits(decision_variables, max_indexes_per_group = 6e+05)
}
\arguments{
\item{decision_variables}{dataframe for the decision variables. This is the
first element of the list produced by \code{comit_objective_function}.}

\item{max_indexes_per_group}{integer, the maximum number of indexes allowed
in a single group. The default is 600000 which is just what worked well in
development when splitting yearly data to reduce peak RAM. This can be
changed to other values if more or less splits are required.}
}
\value{
list of integer vectors that are the indexes. There will be as many
groups as required to ensure that max_indexes_per_group is not exceeded.
}
\description{
Allows for the setting of the constraint in n different splits in order to
reduce the load on RAM.
}
