% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_comit_counterfactual_solver.R
\name{get_yearly_site_used_capacity}
\alias{get_yearly_site_used_capacity}
\title{Calculate the Yearly Required Used Capacity for Each Technology at Each Site}
\usage{
get_yearly_site_used_capacity(data, decision_variables)
}
\arguments{
\item{data}{list of dataframes containing input data, produced by
\code{comit_preprocess_inputs()}.}

\item{decision_variables}{dataframe for the decision variables. The first
element of the list produced by \code{comit_objective_function()}.}
}
\value{
dataframe containing the yearly required capacities for each
technology at each site.
}
\description{
Gets the amount of used capacity that each site needs for each technology in
order to match EEP outputs. Used for setting the baseline constraint in the
counterfactual model.
}
