% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_interpolate_data.R
\name{get_years_for_interpolation}
\alias{get_years_for_interpolation}
\title{Get the year vectors required for interpolation}
\usage{
get_years_for_interpolation(data)
}
\arguments{
\item{data}{list of dataframes containing the input data for the model.}
}
\value{
list of two vectors, first element is all years in period,
second element is the years to actually be modelled.
}
\description{
Get the year vectors required for interpolation
}
