% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-assay_summary.R
\name{getAssaySummaryCounts}
\alias{getAssaySummaryCounts}
\title{Get Assay Summary Counts}
\usage{
getAssaySummaryCounts(DB, HR = FALSE, BASE_PAIR_LIMIT = 10)
}
\arguments{
\item{DB}{File path to subset database file. In the case of making summary tables, we use the
3M version from CoMIT, returned from \code{pull_subset_DB()}}

\item{HR}{if TRUE, returns 3 prime end, i.e. high risk, version}

\item{BASE_PAIR_LIMIT}{number of base pairs from the 3 prime end to be considered high risk}
}
\value{
single_assays, a dataframe with mutated sequence counts per assay
}
\description{
Get Assay Summary Counts
}
