% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-co_occurring.R
\name{getCoOccCounts}
\alias{getCoOccCounts}
\title{Get Co-Occurring Counts}
\usage{
getCoOccCounts(DB, HR = FALSE)
}
\arguments{
\item{DB}{File path to subset database file. In the case of making summary tables, we use the
3M version from CoMIT, returned from \code{pull_subset_DB()}}

\item{HR}{if TRUE, returns 3 prime end, i.e. high risk, version}
}
\value{
df, a dataframe with mutated sequence counts co-occurring on 0 to 7 assays
}
\description{
Get Co-Occurring Counts
}
