% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-assay_summary.R
\name{getCoOccCounts5to7}
\alias{getCoOccCounts5to7}
\title{Get Co-occurring Counts for 5 to 7 Assays}
\usage{
getCoOccCounts5to7(DB, HR = FALSE)
}
\arguments{
\item{DB}{File path to subset database file. In the case of making summary tables, we use the
3M version from CoMIT, returned from \code{pull_subset_DB()}}

\item{HR}{if TRUE, returns 3 prime end, i.e. high risk, version}
}
\value{
df_5to7, a dataframe with mutated sequence counts co-occurring on multiple assays,
only including rows for 5-7 assays affected and the totals
}
\description{
Get Co-occurring Counts for 5 to 7 Assays
}
