% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-co_occurring_condensed.R
\name{getCoOccCounts_2}
\alias{getCoOccCounts_2}
\title{Get Co-Occurring Counts V2}
\usage{
getCoOccCounts_2(
  DB,
  HR = FALSE,
  startDate = NULL,
  endDate = NULL,
  archiveDBFile = NULL
)
}
\arguments{
\item{DB}{File path to subset database file. In the case of making summary tables, we use the
3M version from CoMIT, returned from \code{pull_subset_DB()}}

\item{HR}{if TRUE, returns 3 prime end, i.e. high risk, version}

\item{startDate}{start date for subset DB, if NULL then the .1 percent analysis won't be run}

\item{endDate}{end date for subset DB, if NULL then the .1 percent analysis won't be run}

\item{archiveDBFile}{File path to archive db (used to retrieve older subset of data for comparison; older subset uses same length of time as input subset db, e.g. start_date is "2022-5" and end_date is "2022-07", then older subset start date is "2022-02 and "2022-04)}
}
\value{
df, a dataframe with mutated sequence counts co-occurring on 0 to 7 assays
}
\description{
Make a dataframe with counts of sequences containing co-occurring mutations, organized by WHO designations/Pango lineages
}
