% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F-7-Align_Classify.R
\name{getMaskedAlignmentList}
\alias{getMaskedAlignmentList}
\title{Get Masked Alignment List from supplied primers}
\usage{
getMaskedAlignmentList(seq, accID, refSeq, primersNotClassified)
}
\arguments{
\item{seq}{Character vector, The current sequence being analyzed}

\item{accID, }{Accession Number}

\item{refSeq}{Reference Sequence to align to}

\item{primersNotClassified}{Vector of Primers still needed}
}
\value{
List containing masked alignment sections for each primer
}
\description{
Get Masked Alignment List from supplied primers
}
