% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F-4-GISAID_parsing.R
\name{getMetaData}
\alias{getMetaData}
\title{Parse metadata from a Dataframe}
\usage{
getMetaData(accClade)
}
\arguments{
\item{accClade}{Dataframe containing only rows with relevant data. The first row will be used.}
}
\value{
List with keys: ("clade", "lineage", "subDate")
         if one of the keys is not present it will be NA
}
\description{
Parse metadata from a Dataframe
}
