% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-co_occurring_heatmap.R
\name{getThreeAssayProportion}
\alias{getThreeAssayProportion}
\title{Get Three-Assay Proportion}
\usage{
getThreeAssayProportion(coOccDF3Assay)
}
\arguments{
\item{coOccDF3Assay}{a dataframe returned by \code{makeDFcoOccHeatmap()}}
}
\value{
the number of sequences with mutations affecting all three assays divided by the total number of sequences in the dataset
}
\description{
calculate proportion of sequences with mutations affecting all three of a, d, and e assays for post-market surveillance of 3-assay test
}
