% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-co_occurring_heatmap.R
\name{group_block_anno}
\alias{group_block_anno}
\title{Add Group Block Annotation}
\usage{
group_block_anno(
  group,
  empty_anno,
  gp = grid::gpar(),
  label = NULL,
  label_gp = grid::gpar()
)
}
\arguments{
\item{group}{the column groups this annotation will cover}

\item{empty_anno}{name of the anno_empty object created from HeatmapAnnotation}

\item{gp}{vector of graphic parameters for the rectangle}

\item{label}{text label for this annotation}

\item{label_gp}{vector of graphic parameters for the text}
}
\description{
Add Group Block Annotation
}
