% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_process_sites.R
\name{group_small_point_sites}
\alias{group_small_point_sites}
\title{Groups the small point sites data conditionally on Two_nps_sites parameter}
\usage{
group_small_point_sites(df, input_data)
}
\arguments{
\item{df}{dataframe for small point sites produced by
\code{\link[=get_small_point_site_data]{get_small_point_site_data()}}}

\item{input_data}{data read in from excel template, same as what is passed to
\code{\link[=comit_solver]{comit_solver()}}.}
}
\value{
input dataframe now grouped as required
}
\description{
When Two_nps_sites parameter is FALSE, the data is grouped by sector, cluster
and pipe distance category (in or out of cluster). If the parameter is TRUE
then the data is also grouped by in_cluster_H2.
}
