% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_hydrogen.R
\name{H2_availability}
\alias{H2_availability}
\title{Constrain the amount of hydrogen technologies used in each cluster so that
the amount of available hydrogen is not exceeded}
\usage{
H2_availability(data, decision_variables)
}
\arguments{
\item{data}{list of tables generated by preprocessing the raw inputs from the
input spreadsheet using \code{comit_preprocess_inputs()}}

\item{decision_variables}{dataframe for the decision variables. This is the
first element of the list produced by \code{comit_objective_function}.}
}
\value{
list of constraints. One constraint for each cluster at each time
point. Each element contains a nested list with 4 elements:
\itemize{
\item column indices
\item values
\item direction
\item rhs
}
}
\description{
Limit the amount of used capacity for hydrogen consuming technologies, so
that the amount of hydrogen used in each cluster does not exceed the amount
of hydrogen each cluster has available.
}
