% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_hydrogen.R
\name{H2_conversion}
\alias{H2_conversion}
\title{Constrain hydrogen conversion, so that the correct amount of hydrogen is
produced overall by summing the different hydrogen types to meet demand}
\usage{
H2_conversion(data, decision_variables)
}
\arguments{
\item{data}{list of tables generated by preprocessing the raw inputs from the
input spreadsheet using \code{comit_preprocess_inputs()}}

\item{decision_variables}{dataframe for the decision variables. This is the
first element of the list produced by \code{comit_objective_function}.}
}
\value{
list of constraints. One constraint for each time point where
there are hydrogen used capacity decision variables in the model. Each
element contains a nested list with 4 elements:
\itemize{
\item column indices
\item values
\item direction
\item rhs
}
}
\description{
To account for different hydrogen types in the model, each with their own
costs, this function ensures that there is enough hydrogen input overall to
meet the demand of all hydrogen consuming technologies.
}
\details{
All hydrogen consuming technologies consume hydrogen as a generic commodity
named 'INDMAINSHYG'. In the model coefficients INDMAINSHYG has no cost.
Instead the hydrogen_conversion variables take the separate types of hydrogen
'INDMAINSHYGG', 'INDMAINSHYGB' and 'INDMAINSHYGR' as inputs and produces
a generic generated hydrogen commodity 'HYGEN' as output. This constraint
ensures that as much of the hydrogen_conversion variables are used to ensure
as much of the separate hydrogen types (which are costed) are used to meet
demand for hydrogen consumption overall.

More briefly, here we are making the HYGEN produced by the hydrogen_conversion
variables equal to the hydrogen consumed in each year.
}
