% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_hydrogen.R
\name{H2_infrastructure_capacity}
\alias{H2_infrastructure_capacity}
\title{Constrain H2 infrastructure/H2 technology deployment to ensure enough H2 can
be transported to each site}
\usage{
H2_infrastructure_capacity(data, decision_variables)
}
\arguments{
\item{data}{list of tables generated by preprocessing the raw inputs from the
input spreadsheet using \code{comit_preprocess_inputs()}}

\item{decision_variables}{dataframe for the decision variables. This is the
first element of the list produced by \code{comit_objective_function}.}
}
\value{
list of constraints. One constraint for each site with the potential
to use hydrogen, in each time period. Each element contains a nested list
with 4 elements:
\itemize{
\item column indices
\item values
\item direction
\item rhs
}
}
\description{
Creates a constraint to ensure that there is at least as much H2 infrastructure
to a site as there is capacity to use H2 by all of the sites hydrogen
consuming technologies.
}
