% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_energy_tables.R
\name{H2_production_infrastructure}
\alias{H2_production_infrastructure}
\title{Generate table of H2 production infrastructure outputs. Only used in when
the model is run in H2_production mode}
\usage{
H2_production_infrastructure(solved)
}
\arguments{
\item{solved}{a table containing the model solution, decision variables and
PV_coefficients}
}
\value{
table of H2 production infrastructure outputs with 1 row per
variable_type for each site or cluster.
}
\description{
Generate table of H2 production infrastructure outputs. Only used in when
the model is run in H2_production mode
}
