% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_geography.R
\name{hav.dist}
\alias{hav.dist}
\title{Distance in kilometers between two points}
\usage{
hav.dist(long1, lat1, long2, lat2)
}
\arguments{
\item{long1}{numeric, longitude of point 1.}

\item{lat1}{numeric, latitude of point 1.}

\item{long2}{numeric, longitude of point 2.}

\item{lat2}{numeric, latitude of point 2.}
}
\value{
numeric, distance between the input coordinates in kilometers.
}
\description{
Uses the formula for the havershine distance to calculate the distance between
two points on the surface of earth. Earth's diameter is included as 12742km.
Note that the entries of arguments is backwards to the typical entries of
latitude and longitude, i.e. here longitude comes before latitude.
}
\details{
Future development could remove the need for this function by instead using
\code{geosphere::distHaversine()} - this should be looked at.
}
