% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_hydrogen.R
\name{hydrogen_production}
\alias{hydrogen_production}
\title{Formulates the hydrogen production constraint}
\usage{
hydrogen_production(data, decision_variables)
}
\description{
Formulates the hydrogen production constraint
}
