% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_process_sites.R
\name{impute_site_details_for_ratios}
\alias{impute_site_details_for_ratios}
\title{Assign values to aggregated non point source sites}
\usage{
impute_site_details_for_ratios(df, input_data)
}
\arguments{
\item{df}{dataframe produce inside of \code{\link[=get_non_point_site_values_for_ratios]{get_non_point_site_values_for_ratios()}}
function, that contains artificial sites for the non-point sources. One row
per sector and cluster combination.}

\item{input_data}{data read in from excel template, same as what is passed to
\code{\link[=comit_solver]{comit_solver()}}.}
}
\value{
dataframe with all columns required for the final sites dataframe,
including total_MtCO2 and pipe_dist. Returns twice as many rows as input
df when Two_nps_sites is set to TRUE, else returns same number of rows as
input df.
}
\description{
Gets the emissions and imputed locations of the artificial sites for the
non-point sources.
}
