% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_process_sites.R
\name{impute_site_location}
\alias{impute_site_location}
\title{Generate a new set of coordinates for the aggregate sites}
\usage{
impute_site_location(df, var)
}
\arguments{
\item{df}{dataframe with as many rows as aggregate sites to impute locations
for. Needs to include variables for the coordinates of the cluster center
(Lon_clust & Lat_clust) as well as a variable provided as the other parameter
for pipe distance.}

\item{var}{string for the name of the column which is to be used as the
distance from the cluster centers. Normally this will be the mean pipe
distance of the sites that formed the aggregate.}
}
\value{
a dataframe with as many rows as aggregate sites in the input data,
with a column each for the new longitudes (lon) and latitudes (lat).
}
\description{
Imputes the locations for the artificial sites, by taking the mean pipe
distance as input as a radius and finding a location randomly on the circle
formed by the radius around the coordinates of the cluster center.
}
