% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_capacity_transfer.R
\name{initiate_capacity_transfer_dt}
\alias{initiate_capacity_transfer_dt}
\title{Initiate the data.table for the capacity transfer data}
\usage{
initiate_capacity_transfer_dt(
  decision_variables,
  variable_types,
  min_ind,
  max_ind
)
}
\arguments{
\item{decision_variables}{dataframe for the decision variables. This is the
first element of the list produced by \code{comit_objective_function}.}

\item{min_ind}{integer, minimum of index values to include}

\item{max_ind}{integer, maximum of index values to include}
}
\value{
data.table containing the information required for a subset of the
decision variables (those indexed between the min and max indexes).
}
\description{
Creates a data.table of variables for the subset of indexes of avaialble
variables, in order to create the capacity transfer constraint. Decision
variables for new capacities are joined to the decision variables for available
capacities, so that previous 'new' capacities can later be used to determine
constraint on 'available' capacity.
}
