% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_production.R
\name{initiate_production_constraint_data}
\alias{initiate_production_constraint_data}
\title{Create an initial table containing the data required for the production
constraints}
\usage{
initiate_production_constraint_data(data, decision_variables)
}
\arguments{
\item{data}{list of tables generated by preprocessing the raw inputs from the
input spreadsheet using \code{comit_preprocess_inputs()}}

\item{decision_variables}{dataframe for the decision variables. This is the
first element of the list produced by \code{comit_objective_function}.}
}
\value{
dataframe with one row per used_capacity decision variable and the
following columns:
\itemize{
\item variable_index
\item year
\item site_ID
\item code
\item output_commodity
\item input_commodity
\item input_amount
}
}
\description{
Links technology data to the decision variables in to create the table required
to formulate the production constraints.
}
