% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_interpolate_data.R
\name{interpolate_all_columns}
\alias{interpolate_all_columns}
\title{Impute missing values of time series data}
\usage{
interpolate_all_columns(df)
}
\arguments{
\item{df}{dataframe for the table to be interpolated, with one row per year
for every year between the start and end years. The first column is the
year column and all following columns are interpolated. This df is created
inside the \code{interpolate_for_years()} function.}
}
\value{
dataframe as provided as df argument, but with all missing values
filled by interpolation.
}
\description{
Interpolate the missing values between known points of time series data. This
is used to provide datapoints for years between the years specified in the
inputs. For example, input data is often specified at 5 year intervals and
this function imputes the years between those specified to provide
yearly estimates.
}
\details{
All interpolation is linear. Only missing values are interpolated, if there
is no missing data then the input data will remain unchanged.
}
