% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_interpolate_data.R
\name{interpolate_data}
\alias{interpolate_data}
\title{Ammend time series data tables to include all modelled years}
\usage{
interpolate_data(data)
}
\arguments{
\item{data}{list of dataframes containing the input data for the model.}
}
\value{
data list of dataframes. Tables specified by \code{tables_to_interpolate()}
are interpolated whilst all other tables are left unchanged.
}
\description{
Reduce or expand specified timeseries input datatables to ensure values are
present for every modelled year. This allows for different timesteps to be
used in the model without having to specify different years in the input
spreadsheet.
}
