% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_interpolate_data.R
\name{interpolate_for_years}
\alias{interpolate_for_years}
\title{Interpolate a dataframe to get values for all years to be modelled}
\usage{
interpolate_for_years(df, all_years, modelled_years)
}
\arguments{
\item{df}{dataframe to be interpolated.}

\item{all_years}{dataframe with a single column containing every year between
the start and end years.}

\item{modelled_years}{dataframe with a single column containing only the years
to be modelled.}
}
\value{
dataframe, with the same columns as input dataframe but now with
one row per year to be modelled and interpolation carried out to fill all
missing values.
}
\description{
This function takes all years, makes values in between the years specified NA,
interpolates and then returns a dataframe with only the modelled years present.
}
