% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_interpolate_data.R
\name{interpolate_relevant_tabs}
\alias{interpolate_relevant_tabs}
\title{Perform interpolation on a specified table}
\usage{
interpolate_relevant_tabs(
  table,
  name,
  yearly_data_tables,
  fuel_constraints,
  all_years,
  modelled_years
)
}
\arguments{
\item{table}{dataframe, a specific table from the input data.}

\item{name}{character, the name of the table to be interpolated.}

\item{yearly_data_tables}{vector of characters}

\item{fuel_constraints}{vector of characters, the fuel constraints which
require additional processing due to containing an initial row with a
boolean setting.}

\item{all_years}{dataframe with a single column containing every year between
the start and end years.}

\item{modelled_years}{dataframe with a single column containing only the years
to be modelled.}
}
\value{
dataframe, the table with interpolation having been completed. All of
the original columns remain and there will be one row per modelled year.
}
\description{
Get the required values for each modeled year for a specific table from the
input data. This function calls \code{interpolate_for_years()} and allows for
the additional processing required for some of the fuel constraints.
}
