% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_fuel_share.R
\name{join_fuel_share_data}
\alias{join_fuel_share_data}
\title{Gets the relevant fuel share constraints data for the fuel consuming
decision variables}
\usage{
join_fuel_share_data(fuel_consuming_variables, max_fuel_share_df)
}
\arguments{
\item{fuel_consuming_variables}{dataframe for the fuel consuming
decision variables.}

\item{max_fuel_share_df}{dataframe for the fuel share constraints information,
produced by \code{get_raw_max_fuel_share_data()}.}
}
\value{
dataframe with same number of rows as input dataframe
fuel_consuming_variables, but with an additional column for share_limit
(numeric) and group (character for fuel category).
}
\description{
Joins the share limit information to the relevant decision variables to
allow for the formulation of the fuel share constraints.
}
