% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_production.R
\name{link_production_inputs_to_outputs}
\alias{link_production_inputs_to_outputs}
\title{Join input commodities to output commodities}
\usage{
link_production_inputs_to_outputs(production_df)
}
\arguments{
\item{production_df}{}
}
\value{
dataframe, with one row per combination of input and output commodities
for each of the relevant decision variables (used capacities). This means
that there will be more rows than there are decision variables. Contains
the following columns:
\itemize{
\item variable_index.output - int, index of the output producing variable
\item variable_index.input - int, index of the variable providing the input
\item input_amount.input - numeric, amount of input commodity produced by the technology
\item output_commodity - character, the commodity that is output by the variable
\item demand_technology.output - boolean, whether the output is an end demand commodity.
\item site_ID
\item year
}
}
\description{
Expand the production data so that all combinations of outputs and inputs
are accounted for, for each of the relevant decision variable (used_capacities).
}
\details{
Columns from the output side (left) are suffixed with '.output' whilst columns from
the input side (right) are suffixed with '.input'. Although note that determining
each side as 'output' and 'input' is somewhat arbitrary as the data is identical
\itemize{
\item the join is to itself. What is different is the column used for joining:
'output_commodity' for the left side and 'input_commodity' for the right.
}
}
