% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utilities.R
\name{log_model_timings}
\alias{log_model_timings}
\title{Create a txt file of the model parameters and timings}
\usage{
log_model_timings(data, file_location = "")
}
\arguments{
\item{file_location}{string for the directory to find the 'timing_logs' folder.
Default is an emtpy string ''. Make sure the filepath ends with '/'.}
}
\value{
NULL. Saves out a txt file into timing_logs folder
}
\description{
Used for recording timings of the model to track changes in speed during
development.
}
