% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-assay_summary.R
\name{makeCombinedDataframe}
\alias{makeCombinedDataframe}
\title{Make Combined Dataframe}
\usage{
makeCombinedDataframe(DB, HR = FALSE)
}
\arguments{
\item{DB}{File path to subset database file. In the case of making summary tables, we use the
3M version from CoMIT, returned from \code{pull_subset_DB()}}

\item{HR}{if TRUE, returns 3 prime end, i.e. high risk, version}
}
\value{
combined_df, a dataframe with counts and percentages both by assay and
by co-occurring assays
}
\description{
Make Combined Dataframe
}
