% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-co_occurring.R
\name{makeCoOccTables}
\alias{makeCoOccTables}
\title{Make Tables for Co-occurring mutations}
\usage{
makeCoOccTables(
  DB,
  variantFile,
  saveTable = FALSE,
  saveFolder = NULL,
  height = NULL,
  width = NULL,
  colWidth1 = NULL,
  colWidth2 = NULL
)
}
\arguments{
\item{DB}{File path to subset database file. In the case of making summary tables, we use the
3M version from CoMIT, returned from \code{pull_subset_DB()}}

\item{variantFile}{csv file containing "Displayed_Lineage_Order" column for ordering row names/row groups in the GT table}

\item{saveTable}{logical value indicating whether to save the tables to the specified saveFolder}

\item{saveFolder}{file path containing folder the finished tables will be saved in}

\item{height}{desired height of table .png images}

\item{width}{desired width of table .png images}

\item{colWidth1}{width of "Pangolin Lineage" and "WHO Label" columns in px; if NULL, widths are automatically assigned to try to fit all text}

\item{colWidth2}{width of all other columns in px; if NULL, widths are automatically assigned to try to fit all text}
}
\description{
Make 2 tables (one version with all mutated sequences and one version with sequences with high-risk mutations) showing mutated sequence counts categorized by number of assays affected on top and by variant on the left
}
