% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-co_occurring_condensed.R
\name{makeCoOccTablesCondensed}
\alias{makeCoOccTablesCondensed}
\title{Make Tables for Co-occurring mutations (Condensed Version)}
\usage{
makeCoOccTablesCondensed(
  DB,
  variantFile,
  startDate = NULL,
  endDate = NULL,
  archiveDBFile = NULL,
  saveTable = FALSE,
  saveFolder = NULL,
  height = NULL,
  width = NULL,
  colWidth1 = NULL,
  colWidth2 = NULL,
  colWidth3 = NULL,
  colWidth4 = NULL,
  colWidth5 = NULL
)
}
\arguments{
\item{DB}{File path to subset database file. In the case of making summary tables, we use the
3M version from CoMIT, returned from \code{pull_subset_DB()}}

\item{variantFile}{csv file containing "Displayed_Lineage_Order" column for ordering row names/row groups in the GT table}

\item{startDate}{start date for subset DB, if NULL then the .1 percent analysis won't be run}

\item{endDate}{start date for subset DB, if NULL then the .1 percent analysis won't be run}

\item{archiveDBFile}{File path to archive db (used to retrieve older subset of data for comparison; older subset uses same length of time as input subset db, e.g. start_date is "2022-5" and end_date is "2022-07", then older subset start date is "2022-02 and "2022-04)}

\item{saveTable}{logical value indicating whether to save the tables to the specified saveFolder}

\item{saveFolder}{file path containing folder the finished tables will be saved in}

\item{height}{desired height of table .png images}

\item{width}{desired width of table .png images; if the width is not large enough, text will wrap or table will be cut off}

\item{colWidth1}{width of "Pangolin Lineage" column in px; if NULL, widths are automatically assigned to try to fit all text}

\item{colWidth2}{width of "5" column in px; if NULL, widths are automatically assigned to try to fit all text}

\item{colWidth3}{width of ">=6" column in px; if NULL, widths are automatically assigned to try to fit all text}

\item{colWidth4}{width of "Sequences by Lineage" column in px; if NULL, widths are automatically assigned to try to fit all text}

\item{colWidth5}{width of all other columns in px; if NULL, widths are automatically assigned to try to fit all text}
}
\description{
Make 2 tables (one version with all mutated sequences and one version with sequences with high-risk mutations) showing mutated sequence counts categorized by number of assays affected on top and by variant on the left
}
