% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-database_breakdown_condensed.R
\name{makedbBreakdownTableCondensed}
\alias{makedbBreakdownTableCondensed}
\title{Make Database Breakdown Table (Condensed Version)}
\usage{
makedbBreakdownTableCondensed(
  DB,
  variantFile,
  interval,
  lastMonth,
  colWidth,
  extraNote = NULL,
  saveTable = FALSE,
  saveFolder = NULL,
  height = NULL,
  width = NULL
)
}
\arguments{
\item{DB}{File path to subset database file. In the case of making summary tables, we use the
3M version from CoMIT, returned from \code{pull_subset_DB()}}

\item{variantFile}{csv file containing "Displayed_Lineage_Order" column for ordering row names/row groups in the GT table}

\item{interval}{string describing the first period analyzed, which is a 3-month period, e.g. May - July 2022}

\item{lastMonth}{string describing the second period analyzed, which is a 1-month period, e.g. July 2022}

\item{colWidth}{desired column width; increase if text is wrapping (applies to all columns, for both printed and saved version)}

\item{extraNote}{string with extra source note to be displayed at the bottom of the table}

\item{saveTable}{logical value indicating whether to save the tables to the specified saveFolder}

\item{saveFolder}{file path containing folder the finished tables will be saved in}

\item{height}{desired height of table .png images (when saveTable is TRUE)}

\item{width}{desired width of table .png images (when saveTable is TRUE)}
}
\description{
Make a table showing sequence counts categorized by lineage, comparing a given 3-month period to the last month of the 3-month period
}
