% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-co_occurring.R
\name{makeDFcoOcc}
\alias{makeDFcoOcc}
\title{Make Co-occurring Mutations Dataframe}
\usage{
makeDFcoOcc(in_df)
}
\arguments{
\item{in_df}{input dataframe, taken from getCoOccCounts function}
}
\value{
out_df, a dataframe ready to be made into a GT table
}
\description{
Make Co-occurring Mutations Dataframe
}
