% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-co_occurring_condensed.R
\name{makeDFcoOcc_2}
\alias{makeDFcoOcc_2}
\title{Make Co-occurring Mutations Dataframe V2}
\usage{
makeDFcoOcc_2(in_df)
}
\arguments{
\item{in_df}{input dataframe, taken from getCoOccCounts function}
}
\value{
out_df, a dataframe ready to be made into a GT table
}
\description{
Make Co-occurring Mutations Dataframe V2
}
