% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-co_occurring_heatmap.R
\name{makeDFcoOccHeatmap}
\alias{makeDFcoOccHeatmap}
\title{Make Dataframe for Co-occurring Mutations Heatmap}
\usage{
makeDFcoOccHeatmap(
  DB,
  HR = FALSE,
  BASE_PAIR_LIMIT = 10,
  threeAssay = FALSE,
  startDate = NULL,
  endDate = NULL,
  archiveDBFile = NULL
)
}
\arguments{
\item{DB}{File path to subset database file. In the case of making summary tables, we use the
3M version from CoMIT, returned from \code{pull_subset_DB()}}

\item{HR}{if TRUE, returns 3 prime end, i.e. high risk, version}

\item{BASE_PAIR_LIMIT}{number of base pairs from the 3 prime end to be considered high risk}

\item{threeAssay}{if TRUE, makes three-assay version of dataframe (a, d, and e assays) for 3-assay version of Covid test}

\item{startDate}{start date for subset DB, if NULL then the .1 percent analysis won't be run}

\item{endDate}{end date for subset DB, if NULL then the .1 percent analysis won't be run}

\item{archiveDBFile}{File path to archive db (used to retrieve older subset of data for comparison; older subset uses same length of time as input subset db, e.g. start_date is "2022-5" and end_date is "2022-07", then older subset start date is "2022-02 and "2022-04)}
}
\value{
comboFreq, a dataframe with counts and percentages of each co-occurring mutation combo of assays
}
\description{
Create dataframe to be used by \code{makeCoOccHeatmaps()} in creating visulization of co-occurring mutations on different combinations of assays, with the sequences organized according to WHO designations/Pango lineages
}
