% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-database_breakdown.R
\name{makeDFdatabaseBreakdown}
\alias{makeDFdatabaseBreakdown}
\title{Make Database Breakdown Dataframe}
\usage{
makeDFdatabaseBreakdown(DB, interval, lastMonth, total_seqs, total_seqs_1m)
}
\arguments{
\item{DB}{File path to subset database file. In the case of making summary tables, we use the
3M version from CoMIT, returned from \code{pull_subset_DB()}}

\item{interval}{string describing the first period analyzed, which is a 3-month period, e.g. May - July 2022}

\item{lastMonth}{string describing the second period analyzed, which is a 1-month period, e.g. July 2022}

\item{total_seqs}{total number of sequences for the 3-month period}

\item{total_seqs_1m}{total number of sequences for the 1-month period}
}
\value{
combined_df, a dataframe with data for both time periods and all needed formatting applied for creation of the gt table
}
\description{
Make Database Breakdown Dataframe
}
