% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-primer_variant_visualization.R
\name{makePrimerVarFreqLinDFs}
\alias{makePrimerVarFreqLinDFs}
\title{Make primer variant frequency and lineage dataframes}
\usage{
makePrimerVarFreqLinDFs(dbFile)
}
\arguments{
\item{dbFile}{File path for subset db}
}
\value{
freqDF, which has frequencies for each significant var ID, and linDF, which has lineage information for each significant var ID (a single var ID can be marked as part of multiple lineages in the database)
}
\description{
Make primer variant frequency and lineage dataframes
}
