% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-primer_variant_visualization.R
\name{makePrimerVarMutationDF}
\alias{makePrimerVarMutationDF}
\title{Make Primer Variant Mutation Dataframe}
\usage{
makePrimerVarMutationDF(
  start_date,
  end_date,
  varIds = NULL,
  archiveDBFile,
  subsetDBFile = NULL,
  printableDoubles = FALSE
)
}
\arguments{
\item{start_date}{Start date for subset db}

\item{end_date}{End date for subset db}

\item{varIds}{Vector of variant IDs to be used in the visualization, provided if subsetDBFile is NULL (not using Sig_Var_IDs from the subsetDBFile)}

\item{archiveDBFile}{File path to archive db (used to retrieve older subset of data for comparison; older subset uses same length of time as input subset db, e.g. start_date is "2022-5" and end_date is "2022-07", then older subset start date is "2022-02 and "2022-04)}

\item{subsetDBFile}{File path to subset db, provided if using variant IDs from Sig_Var_IDs as those in the visualization (varIds parameter should be NULL if this is provided)}

\item{printableDoubles}{Boolean, default is FALSE, if TRUE will show double mutations for a single variant ID in a single column of the visualization}
}
\value{
unique_var_table, a table with mutations and trends for each significant variant, comparing this subset to an older subset
}
\description{
Make Primer Variant Mutation Dataframe
}
