% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-primer_variant_visualization.R
\name{makePrimerVarTable}
\alias{makePrimerVarTable}
\title{Make primer variant visualization}
\usage{
makePrimerVarTable(
  save_table,
  save_folder,
  start_date,
  end_date,
  archive_file,
  subset_file,
  who_var_file,
  color_palette,
  section_heights = NULL,
  figure_res = NULL,
  figure_width = NULL,
  figure_height = NULL
)
}
\arguments{
\item{save_table}{Whether or not to save the table}

\item{save_folder}{Folder name for saving table}

\item{start_date}{Start date for subset db}

\item{end_date}{End date for subset db}

\item{archive_file}{File path to archive db (used to retrieve older subset of data for comparison; older subset uses same length of time as input subset db, e.g. start_date is "2022-5" and end_date is "2022-07", then older subset start date is "2022-02 and "2022-04)}

\item{subset_file}{File path to subset db}

\item{who_var_file}{File path to WHO variant file (for getting Displayed Lineage order for legend in bottom stacked plot)}

\item{color_palette}{Letter value A - H corresponding to a viridis color palette; palettes are [A: magma, B: inferno, C: plasma, D: viridis, E: cividis, F: rocket, G: mako, H: turbo]; if color value doesn't match, it defaults to "viridis" palette}

\item{section_heights}{Optional vector containing the heights of the three sections of the visualization (top graph, middle table, and bottom graph). If no value is supplied, default is c(2,1.25,2).}

\item{figure_res}{Resolution of saved png figure}

\item{figure_width}{Width of saved png figure}

\item{figure_height}{Height of saved png figure}
}
\description{
Create a figure with three parts: a bar graph with variant (set of mutations at a primer binding region) counts organized by assay on top, a table with specific variant information in the middle, and a bar graph with lineage makeup per variant on the bottom
}
