% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_CO2.R
\name{max_CCS}
\alias{max_CCS}
\title{Constrain the amount of CCS that can be used in industry}
\usage{
max_CCS(data, decision_variables)
}
\arguments{
\item{data}{list of tables generated by preprocessing the raw inputs from the
input spreadsheet using \code{comit_preprocess_inputs()}}

\item{decision_variables}{dataframe for the decision variables. This is the
first element of the list produced by \code{comit_objective_function}.}
}
\value{
list of constraints. One constraint for each time period with a
constraint on the maximum amount of carbon that can be captured.
Each element contains a nested list with 4 elements:
\itemize{
\item column indices
\item values
\item direction
\item rhs
}
}
\description{
Limits the total amount of carbon that can be captured in a given year,
across all of industry.
}
