% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_fuel_share.R
\name{max_fuel_share}
\alias{max_fuel_share}
\title{Constrain fuel use to remain under a specified share of the total fuel used
to produce a commodity}
\usage{
max_fuel_share(data, decision_variables)
}
\arguments{
\item{data}{list of tables generated by preprocessing the raw inputs from the
input spreadsheet using \code{comit_preprocess_inputs()}}

\item{decision_variables}{dataframe for the decision variables. This is the
first element of the list produced by \code{comit_objective_function}.}
}
\value{
list of constraints. One constraint for each commodity specified to
be constrained in the input, in each time point and for each fuel type
that is to be constrained. Each element contains a nested list with 4 elements:
\itemize{
\item column indices
\item values
\item direction
\item rhs
}
}
\description{
Create constraints that set upper limits for the proportional amount of a fuel
that can be used to produce a given commodity, as specified in the input file.
}
\details{
Note that there are limits in the input dataframe for fuel categories for
commodities that don't actually have inputs from that category - these are
dropped when setting the constraint up.
}
