% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_fuel.R
\name{max_fuel_use}
\alias{max_fuel_use}
\title{Constrain fuel consumption to remain under a maximum absolute value}
\usage{
max_fuel_use(data, decision_variables)
}
\arguments{
\item{data}{list of tables generated by preprocessing the raw inputs from the
input spreadsheet using \code{comit_preprocess_inputs()}}

\item{decision_variables}{dataframe for the decision variables. This is the
first element of the list produced by \code{comit_objective_function}.}
}
\value{
list of constraints. One constraint for each fuel specified to be
constrainted in the input, in each time point. Each element contains a
nested list with 4 elements:
\itemize{
\item column indices
\item values
\item direction
\item rhs
}
}
\description{
Create constraints that set upper limits for the absolute amount of a fuel
that can be used, as specified in the input file. Fuels can be constrained
either at the individual commodity code level (e.g. "INDMAINSHYG") or at
the fuel category level (e.g. "Biomass and organic waste").
}
\details{
This function is really just a wrapper for \code{absolute_fuel_constraints()} as the
underlying code is essentially the same for \code{min_fuel_use()} - we just change
the capacity type from 'max' to 'min'.
}
