% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_capacities.R
\name{maximum_capacity}
\alias{maximum_capacity}
\title{Constrain technology rollout to a given maximum proportion of commodity production}
\usage{
maximum_capacity(data, decision_variables)
}
\arguments{
\item{data}{list of tables generated by preprocessing the raw inputs from the
input spreadsheet using \code{comit_preprocess_inputs()}}

\item{decision_variables}{dataframe for the decision variables. This is the
first element of the list produced by \code{comit_objective_function}.}
}
\value{
list of constraints. One constraint for each relevant technology, in
each time point.
Each element contains a nested list with 4 elements:
\itemize{
\item column indices
\item values
\item direction
\item rhs
}
}
\description{
Create constraints to set an upper limit for the proportion of outputs that
can be produced by a given technology. For example, we may wish to set a limit
in a given year for the proportion of all steel that can be produced by an
electric arc furnace.
}
\details{
This function is really just a wrapper for \code{capacity_constraints()} as the
underlying code is essentially the same for \code{minimum_capacity()} - we just change
the capacity type from 'max' to 'min'.
}
