% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F-2-classify_helpers.R
\name{mismatchGenerator}
\alias{mismatchGenerator}
\title{Generate Mismatches for a primer sequence}
\usage{
mismatchGenerator(seq, saveList = FALSE)
}
\arguments{
\item{seq}{Primer sequence to generate mismatches for}

\item{saveList}{Boolean, if this is set to true the list will be saved in cache$mismatchDict
for later reference. Note: If there are many primers, or they are particularly long, then
this can cause memory errors as lists can be long especially on second generation (2 mismatch lists)}
}
\value{
vector of all possible SNP mutations
}
\description{
Create a vector of sequences with all possible SNP mutations (single base pair mismatches) given a string sequence
}
\examples{
mismatchGenerator("AATTTGTGTCAA")

}
