% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_read_data.R
\name{mround}
\alias{mround}
\title{Helper function to round a number x to the nearest multiple of base}
\usage{
mround(x, base)
}
\description{
Helper function to round a number x to the nearest multiple of base
}
