% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_sites.R
\name{naei_pipe_to_spur_adjustment}
\alias{naei_pipe_to_spur_adjustment}
\title{Adjust the NAEI data's pipe_dist column to account for any cluster
spurs.}
\usage{
naei_pipe_to_spur_adjustment(data)
}
\arguments{
\item{data}{list of input dataframes after initial preprocessing steps.}
}
\value{
dataframe of the NAEI_data as provided in the data input list,
but with the pipe_dist for any sites specified in the input spreadsheet
'spur_sites' tab to be effected by spurs to be adjusted to the specified
values.
}
\description{
Adjust the NAEI data's pipe_dist column to account for any cluster
spurs.
}
