% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_emissions.R
\name{net_emissions}
\alias{net_emissions}
\title{This function returns net emissions outputed per used capacity of a technology in kilotonnes of CO2e}
\usage{
net_emissions(technology_code, gas = NULL, year = 2020, .data = data)
}
\description{
This function returns net emissions outputed per used capacity of a technology in kilotonnes of CO2e
}
