% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_no_building_in_start_year.R
\name{no_building_in_start_year}
\alias{no_building_in_start_year}
\title{Create a constraint to stop the building of new capacity in the first modeled
year}
\usage{
no_building_in_start_year(data, decision_variables)
}
\arguments{
\item{data}{list of tables generated by preprocessing the raw inputs from the
input spreadsheet using \code{comit_preprocess_inputs()}}

\item{decision_variables}{dataframe for the decision variables. This is the
first element of the list produced by \code{comit_objective_function}.}
}
\value{
list of constraints. One constraint per variable for each new
capacity variable in the start year. Each element contains a nested list
with 4 elements:
\itemize{
\item column indices
\item values
\item direction
\item rhs
}
}
\description{
Limit the building of new capacity for technology, hydrogen pipes and CO2
pipes to 0 for the first period in the model. This ensures only the preset
available capacities are used in the start year.
}
