% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_hydrogen.R
\name{non_industry_H2_demand}
\alias{non_industry_H2_demand}
\title{Constrain non-industry hydrogen variables to ensure non-industry demand of
hydrogen is met}
\usage{
non_industry_H2_demand(data, decision_variables)
}
\arguments{
\item{data}{list of tables generated by preprocessing the raw inputs from the
input spreadsheet using \code{comit_preprocess_inputs()}}

\item{decision_variables}{dataframe for the decision variables. This is the
first element of the list produced by \code{comit_objective_function}.}
}
\value{
list of constraints. One constraint for each cluster, in each time
point. Each element contains a nested list with 4 elements:
\itemize{
\item column indices
\item values
\item direction
\item rhs
}
}
\description{
Make sure enough non-industry hydrogen is included in the model to meet the
non-industry demand in each cluster and in each year.
}
