% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_process_sites.R
\name{non_point_sites_from_ratios}
\alias{non_point_sites_from_ratios}
\title{Produce a table for artificial sites that represent the non-point sources}
\usage{
non_point_sites_from_ratios(df, input_data)
}
\arguments{
\item{input_data}{data read in from excel template, same as what is passed to
\code{\link[=comit_solver]{comit_solver()}}.}
}
\value{
dataframe with one row per artificial site, representing the
non-point source sites for modelling purposes. There should be as many sites
as sectors * used clusters when Two_nps_sites is set to FALSE, and twice this
amount otherwise. Columns are:
\itemize{
\item site_name
\item IPM_sector
\item H2_point
\item total_MtCO2 (emissions)
\item pipe_dist
\item num_sites
\item Latitude
\item Longitude
\item traded_flag
}
}
\description{
Performs a series of calculations to impute the relevant details so that
the sites are represented as accurately as possible.
}
