% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F-4-GISAID_parsing.R
\name{parseFastaLine}
\alias{parseFastaLine}
\title{Parse GISAID FASTA line for metadata and check for sequences with no accession number,
the wrong host, and/or invalid collection dates}
\usage{
parseFastaLine(line)
}
\arguments{
\item{line}{A string containing the initial line of a FASTA item (line after >).
This must be pulled from the GISAID database}
}
\value{
List with keys: ("accNum", "location", "colDate", "continue") containing
        the Accession ID, location, collection date, and error status receptively.
        If a host other than human is found or an accession number is not identified
        the "continue" field will be marked false and a "Fail_Type" key will be added
        containing failure mode details
}
\description{
Parse GISAID FASTA line for metadata and check for sequences with no accession number,
the wrong host, and/or invalid collection dates
}
