% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_read_data.R
\name{pivot_specified_table}
\alias{pivot_specified_table}
\title{Pivot a specified dataframe}
\usage{
pivot_specified_table(data, pivot_details)
}
\arguments{
\item{data}{list of dataframes read in by \code{read_excel_data_template()}}

\item{pivot_details}{vector of length 3 containing the information required
for pivoting of input data. The first element is the name of the table to be
pivoted. The second element provides the name of the new naming column (as per
the 'names_to' argument in pivot longer) and the third element provides the
name of the new values column (as per the 'values_to' argument of pivot longer).}
}
\value{
data, the entire list of dataframes provided but with the named table
pivoted.
}
\description{
This function carries out common pivoting, whereby tables are made longer,
pivoting all columns apart from a year column.
}
