% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_attribution.R
\name{plot_attribution}
\alias{plot_attribution}
\title{Plot Emissions Attribution}
\usage{
plot_attribution(emissions_total, this_sector)
}
\arguments{
\item{emissions_total}{Data frame containing total emissions data.}

\item{this_sector}{Character string specifying the sector to filter by.
Use 'All' for no filter.}
}
\value{
ggplot object
}
\description{
This function creates a plot of emissions attribution for a specified sector
or all sectors and visualizing the data using a ribbon plot.
}
