% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_outputs.R
\name{plot_capture}
\alias{plot_capture}
\title{Plot CO2 captured by year}
\usage{
plot_capture(data, my_colours)
}
\arguments{
\item{data}{Output of \code{\link[=summarise_capture]{summarise_capture()}}}

\item{my_colours}{Named vector of colours keyed by run}
}
\value{
A ggplot object
}
\description{
Plot CO2 captured by year
}
