% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_outputs.R
\name{plot_cost_shares_by_type}
\alias{plot_cost_shares_by_type}
\title{Plot stacked cost shares by year (per run)}
\usage{
plot_cost_shares_by_type(data)
}
\arguments{
\item{data}{Output of \code{\link[=get_cost_shares]{get_cost_shares()}}}
}
\value{
A ggplot object
}
\description{
Plot stacked cost shares by year (per run)
}
