% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_outputs.R
\name{plot_costs_by_type}
\alias{plot_costs_by_type}
\title{Plot total costs by year, faceted by cost type}
\usage{
plot_costs_by_type(data, my_colours)
}
\arguments{
\item{data}{Output of \code{\link[=summarise_costs]{summarise_costs()}} with \code{total_cost}}

\item{my_colours}{Named vector of colours keyed by run}
}
\value{
A ggplot object
}
\description{
Plot total costs by year, faceted by cost type
}
