% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_outputs.R
\name{plot_deployment}
\alias{plot_deployment}
\title{Plot technology deployment by year, faceted by technology category}
\usage{
plot_deployment(data, my_colours)
}
\arguments{
\item{data}{Output of \code{\link[=summarise_deployment]{summarise_deployment()}}}

\item{my_colours}{Named vector of colours keyed by run}
}
\value{
A ggplot object
}
\description{
Plot technology deployment by year, faceted by technology category
}
