% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_outputs.R
\name{plot_emissions}
\alias{plot_emissions}
\title{Plot total direct emissions by year}
\usage{
plot_emissions(data, my_colours)
}
\arguments{
\item{data}{Output of \code{\link[=summarise_emissions]{summarise_emissions()}}}

\item{my_colours}{Named vector of colours keyed by run}
}
\value{
A ggplot object
}
\description{
Plot total direct emissions by year
}
