% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_outputs.R
\name{plot_fuel}
\alias{plot_fuel}
\title{Plot total fuel use by year, faceted by fuel category}
\usage{
plot_fuel(data, my_colours)
}
\arguments{
\item{data}{Output of \code{\link[=summarise_fuel]{summarise_fuel()}}}

\item{my_colours}{Named vector of colours keyed by run}
}
\value{
A ggplot object
}
\description{
Plot total fuel use by year, faceted by fuel category
}
