% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_outputs.R
\name{plot_fuel_shares}
\alias{plot_fuel_shares}
\title{Plot fuel shares over time, faceted by fuel}
\usage{
plot_fuel_shares(data, my_colours)
}
\arguments{
\item{data}{Output of \code{\link[=get_fuel_shares]{get_fuel_shares()}}}

\item{my_colours}{Named vector of colours keyed by run}
}
\value{
A ggplot object
}
\description{
Plot fuel shares over time, faceted by fuel
}
