% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_outputs.R
\name{plot_fuel_shares_stacked}
\alias{plot_fuel_shares_stacked}
\title{Plot stacked fuel shares by year (per run)}
\usage{
plot_fuel_shares_stacked(data)
}
\arguments{
\item{data}{Output with \code{share}, \code{ymin}, \code{ymax} from \code{\link[=get_fuel_shares]{get_fuel_shares()}}}
}
\value{
A ggplot object
}
\description{
Plot stacked fuel shares by year (per run)
}
