% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_finance.R
\name{PMT}
\alias{PMT}
\title{Calculates the payment for a loan based on constant payments and a constant
interest rate}
\usage{
PMT(principal, rate, n_payments)
}
\arguments{
\item{principal}{numeric, or vector of numeric values. The principal present
value.}

\item{rate}{numeric. The interest rate for the loan.}

\item{n_payments}{integer. The total number of payments for the loan. Usually
a lifetime variable.}
}
\value{
numeric, or vector of numeric values with same length as principle.
Required payment per loan period.
}
\description{
Estimates the value of periods to be made in each period. Similar to the
excel formula PMT.
}
\author{
Parameter descriptions from Microsoft Corporation 2021
}
