% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_finance.R
\name{present_value}
\alias{present_value}
\title{Function to calculate present value in given periods}
\usage{
present_value(FV, rate, start_period, n_periods = 1)
}
\arguments{
\item{FV}{numeric, or numeric column of a df.
This Future Value: the actual value at a given point in time.}

\item{rate}{numeric. The assumed interest rate. This should be provided as a
decimal value, e.g. 1\% interest is provided as 0.01.}

\item{start_period}{integer, or integer column of a df.
The period for which the Future Value is provided.
This value should be provided as the number of years from the model's start
year, where the start year is 0. It is not the actual year (i.e. do not
input the year as "2050").}

\item{n_periods}{integer. The number of periods after the period for which
the present value should be evaluated. This is usually the model's timestep
parameter. If it is set to 1, it is just a standard present value calculation,
otherwise it allows for the summing of values for years not modelled, when it
is appropriate, such as for in the calculation of opex or fuel values.}
}
\value{
vector of numeric values, of same length as FV input parameter. Each
value is the present value calculated for the respective future value provided.
In cases where n_periods is > 1, this value is the sum of the present values
for the start_period and the subsequent n_periods that follow.
}
\description{
Perform present value calculations on future values at a specified year. This
allows for values for things such as fuels and technologies to be considered
at the present value, to allow for more accurate decision making in the model.
Note that this function can also be used to aggregate the present values at
multiple years, which is helpful when using a model with a timestep, to account
for accrued costs (such as opex) in non-modelled years.
}
