% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_process_sites.R
\name{process_sites}
\alias{process_sites}
\title{Cleans site data and prepares for use in model}
\usage{
process_sites(input_data)
}
\arguments{
\item{input_data}{data read in from excel template, same as what is passed to
\code{\link[=comit_solver]{comit_solver()}}.}
}
\value{
list of data as provided as input_data, with the following updates:
\itemize{
\item A data frame NAEI_clean appended to be used as the source for site data
\item NAEI_df_clean_2021_revised has a row appended to flag traded status
}
}
\description{
Tidies and prepares the site data ready for use in the models, including
adding imputed aggregate sites for non-point sources and aggregating small
point sources to reduce computational load.
}
