% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_constraints_production.R
\name{production}
\alias{production}
\title{Formulate Production Constraints}
\usage{
production(data, decision_variables)
}
\arguments{
\item{data}{list of tables generated by preprocessing the raw inputs from the
input spreadsheet using \code{comit_preprocess_inputs()}}

\item{decision_variables}{dataframe for the decision variables. This is the
first element of the list produced by \code{comit_objective_function}.}
}
\value{
list of constraints. One constraint for each commodity, at each site,
at each time point.
Each element contains a nested list with 4 elements:
\itemize{
\item column indices
\item values
\item direction
\item rhs
}
}
\description{
The production constraint ensures that the required amount of commodities are
produced in each year and at each site. There are two types of constraints
created here: one for demand commodities (final products) and another for
process commodities (required as inputs in the process of creating final
commodities). An example of a demand commodity would be \emph{cement}, whilst an
example of a process commodity would be \emph{high temperature heat}.
}
\details{
For demand commodities:
∑used_capacity(t,s,tech) = demand(t,s)
for tech ∈ process\if{html}{\out{<sub>}}d\if{html}{\out{<sub>}}(s)

For process commodities:
∑used_capacity(t,s,tech\if{html}{\out{<sub>}}a\if{html}{\out{</sub>}}) = ∑(used_capacity(t,s,tech\if{html}{\out{<sub>}}b\if{html}{\out{</sub>}}) *
input_amount(tech\if{html}{\out{<sub>}}b\if{html}{\out{</sub>}})),
for any output commodity, for tech\if{html}{\out{<sub>}}a\if{html}{\out{</sub>}} ∈ process(s)
which is not process\if{html}{\out{<sub>}}d\if{html}{\out{</sub>}}(s)
}
