% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utilities.R
\name{progress_updater}
\alias{progress_updater}
\title{Update the rshiny progress bar when running the model through the app}
\usage{
progress_updater(finished_stage, in_app, data)
}
\arguments{
\item{finished_stage, }{the point in the model that has just finished. These
should be one of the values in \code{data$model_timings$stage}.}

\item{in_app}{boolean, TRUE if code is being run in the app,
in which case the function is executed, otherwise it is ignored.}
}
\value{
NULL. Progress bar is updated to the value associated with the
finished stage in \code{data$model_timings}
}
\description{
The timings are pulled from the input spreadsheet and are inferred from
previous runs to give a representative estimate of the progress made after
each stage of the process has been completed.
}
