% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_subset_DB.R
\name{pull_subset_db}
\alias{pull_subset_db}
\title{Pull subset DB}
\usage{
pull_subset_db(
  archiveFile,
  startDate,
  endDate,
  NewDBFile,
  variantFile,
  force = FALSE,
  silently = TRUE
)
}
\arguments{
\item{archiveFile}{string name of archive db file or db file to subset from, e.g. "databases/All_Assay_Archive.db"}

\item{startDate}{string start date of data to be included in subset db, e.g. "2022-05"}

\item{endDate}{string end date of data to be included in subset db, e.g. "2022-07"}

\item{NewDBFile}{string name of new subset db file, e.g. "DB_files/DB_3M/all_May-July_SD08-15_subgroups.db"}

\item{variantFile}{string name of .csv file containing breakdown of lineage/subgroup names of sequences, e.g. "who_lineages_subgroups.csv"}

\item{force}{logical value indicating whether or not to force override already existing subset db with same name}

\item{silently}{logical value indicating whether to print tables created in subset db}
}
\description{
Pull a subset database from a given "archive" database based on a given start date and end date and a variant file
}
