% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_finance.R
\name{PV_calculation}
\alias{PV_calculation}
\title{Calculate a single present value}
\usage{
PV_calculation(FV, rate, start_period, n_periods)
}
\arguments{
\item{FV, }{numeric. Future value.}

\item{rate}{numeric. Assumed interest rate.}

\item{start_period}{integer. The year of the respective future value. First
modelled year is 0, i.e. not 2020.}

\item{n_periods}{number of years after the year provided to aggregate, useful
if a timestep model is ran and costs need to be accrued.}
}
\value{
numeric. The present value or accrued present value.
}
\description{
Helper function to present_value. By specifying the calculation separately,
we can call it in within \code{present_value()} to allow for mapply to be used in
order to provide the function within 'mutate' and provide multiple columns
as arguments.
}
\details{
Present value is defined as PV = FV * (1 / (1+rate)^period)
}
