% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_pv_co2.R
\name{PV_CO2_national_transport}
\alias{PV_CO2_national_transport}
\title{Function which gets coefficient associated with CO2 transport from cluster to storage site}
\usage{
PV_CO2_national_transport(data, decision_variables)
}
\arguments{
\item{data}{list of data tables read in from excel data template}

\item{decision_variables}{dataframe of decision variables}
}
\value{
A two column dataframe. The first column specifies the index of the decision variable.
The second column gives the the coefficient associated with CO2 transport costs between cluster and storage site
}
\description{
Function which gets coefficient associated with CO2 transport from cluster to storage site
}
