% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_pv_fixed_opex.R
\name{PV_fixed_opex}
\alias{PV_fixed_opex}
\title{Function which gets the coefficient associated with fixed opex for each
decision variable}
\usage{
PV_fixed_opex(data, decision_variables)
}
\arguments{
\item{data}{list of tables generated by preprocessing the raw inputs from the
input spreadsheet using \code{comit_preprocess_inputs()}}

\item{decision_variables}{dataframe for the decision variables.}
}
\value{
A two column dataframe. The first column specifies the index of the
decision variable. The second column gives the the coefficient associated
with fixed opex.
}
\description{
Function which gets the coefficient associated with fixed opex for each
decision variable
}
