% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_pv_fuel.R
\name{PV_fuel_cost}
\alias{PV_fuel_cost}
\title{Function which gets coefficient associated with fuel use of each used technology}
\usage{
PV_fuel_cost(data, decision_variables)
}
\arguments{
\item{data}{list of data tables read in from excel data template}

\item{decision_variables}{dataframe of decision variables}
}
\value{
A two column dataframe. The first column specifies the index of the decision variable.
The second column gives the the coefficient associated with fuel use
}
\description{
Function which gets coefficient associated with fuel use of each used technology
}
