% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_pv_h2.R
\name{PV_H2_pipe_cluster_to_site}
\alias{PV_H2_pipe_cluster_to_site}
\title{Function which gets coefficient associated with opex and capex of H2 pipes from site to cluster}
\usage{
PV_H2_pipe_cluster_to_site(data, decision_variables)
}
\arguments{
\item{data}{list of data tables read in from excel data template}

\item{decision_variables}{dataframe of decision variables}
}
\value{
A two column dataframe. The first column specifies the index of the decision variable.
The second column gives the the coefficient associated with capex or opex costs of H2 pipes between
site and cluster
}
\description{
Function which gets coefficient associated with opex and capex of H2 pipes from site to cluster
}
