% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_pv_h2.R
\name{PV_H2_pipe_national}
\alias{PV_H2_pipe_national}
\title{Function which gets coefficient associated with H2 pipes between clusters}
\usage{
PV_H2_pipe_national(data, decision_variables)
}
\arguments{
\item{data}{list of data tables read in from excel data template}

\item{decision_variables}{dataframe of decision variables}
}
\value{
A two column dataframe. The first column specifies the index of the decision variable.
The second column gives the the coefficient associated with opex and capex of H2 pipes connecting clusters
}
\description{
Function which gets coefficient associated with H2 pipes between clusters
}
