% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_pv_technology_capex.R
\name{PV_technology_capex}
\alias{PV_technology_capex}
\title{Function which gets coefficient associated with capex of each technology}
\usage{
PV_technology_capex(data, decision_variables)
}
\arguments{
\item{data}{list of data tables read in from excel data template}

\item{decision_variables}{dataframe of decision variables}
}
\value{
A two column dataframe. The first column specifies the index of the decision variable.
The second column gives the the coefficient associated with capex costs
}
\description{
Function which gets coefficient associated with capex of each technology
}
