% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_read_data.R
\name{read_excel_data_template}
\alias{read_excel_data_template}
\title{Read in the input template workbook}
\usage{
read_excel_data_template(path)
}
\arguments{
\item{path}{character for the file location of the workbook}
}
\value{
list of dataframes. One dataframe per sheet from the input data
workbook. Note that only the sheets in the workbook between the section
headers (with titles ending '==>') are read in, all other sheets are
ignored.
}
\description{
Read in the relevant sheets of the excel spreadsheet that provides all
input assumptions, data and model parameters for a comit scenario to be
ran.
}
