% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_read_data.R
\name{read_input_sheet}
\alias{read_input_sheet}
\title{Read in an individual sheet from the input workbook}
\usage{
read_input_sheet(path, sheet)
}
\arguments{
\item{path}{character for the file location of the workbook}

\item{sheet}{string, the name of the sheet to bread}
}
\value{
dataframe version of the input table present in the sheet named from
the input workbook.
}
\description{
Reads in the table from the relevant sheet. The read in starts from the 7th
row and 1st column, and reads until the first completely empty row and first
completely empty column.
}
