% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_attribution.R
\name{read_outputs}
\alias{read_outputs}
\title{Read an output file and complete some basic preprocessing}
\usage{
read_outputs(file, type)
}
\arguments{
\item{file}{(character), file path for the output file to be read, or alternatively a wb object.}

\item{type}{(character), name of the tab to be read - either 'Emissions' or 'Energy'}
}
\value{
dataframe, the output file in longer format with years as a variable.
}
\description{
Read an output file and complete some basic preprocessing
}
