% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_read_data.R
\name{read_xlsx_with_warning_handler}
\alias{read_xlsx_with_warning_handler}
\title{Read in a sheet from the xlsx worksheet in the standard way, but suppress
a particular type of warning that does not need to be flagged.}
\usage{
read_xlsx_with_warning_handler(path, sheet)
}
\arguments{
\item{path}{character for the file location of the workbook}

\item{sheet}{string, the name of the sheet to bread}
}
\value{
dataframe version of the specified xlsx sheet.
}
\description{
Read in a sheet from the xlsx worksheet in the standard way, but suppress
a particular type of warning that does not need to be flagged.
}
