% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_emissions.R
\name{recode_fuel_CO2e}
\alias{recode_fuel_CO2e}
\title{Recode fuel_CO2e Variable}
\usage{
recode_fuel_CO2e(emissions_df, zero_emissions_from_biomass)
}
\arguments{
\item{emissions_df}{dataframe, created by \code{create_emissions_df()}}

\item{zero_emissions_from_biomass}{boolean, whether to treat biomass fuels
as having 0 emissions.}
}
\value{
dataframe, with same number of rows and columns as emissions_df
provided, but with fuel_CO2e recoded where relevant.
}
\description{
Process the fuel_CO2e variable, setting its missing values to 0 and setting
rows with the commodity category for "Biomass and organic waste" to 0 when
zero_emissions_from_biomass is TRUE.
}
