% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_geography.R
\name{region_lookup}
\alias{region_lookup}
\title{Classify region of sites/clusters}
\usage{
region_lookup(coord_data, link_to_cluster = FALSE)
}
\arguments{
\item{coord_data}{dataframe to classify, must include Longitude and Latitude.}

\item{link_to_cluster}{logical, use TRUE if linking to clusters rather than sites.}
}
\value{
A dataframe with classified regions, to be used as a lookup. This can
then be joined on site_ID (or 'Cluster' if link_to_cluster = TRUE).
}
\description{
Assigns the GOR9D region to either sites or clusters of a dataframe based on
its co-ordinates. This is done so that outputs can be dis-aggrageted by region.
}
