% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_process_sites.R
\name{regional_non_point_site_shares}
\alias{regional_non_point_site_shares}
\title{Helper function to calculate the amount of emissions from non-point-sites
for each sector in each region}
\usage{
regional_non_point_site_shares(region, df)
}
\description{
Helper function to calculate the amount of emissions from non-point-sites
for each sector in each region
}
