% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-primer_variant_visualization.R
\name{reorderVect}
\alias{reorderVect}
\title{Reorder vect - change the position of what in vect to where}
\usage{
reorderVect(vect, what, where)
}
\arguments{
\item{vect}{vector containing elements to reorder}

\item{what}{element of vector to move}

\item{where}{new position of what}
}
\value{
vect
}
\description{
Reorder vect - change the position of what in vect to where
}
