% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_read_data.R
\name{round_years}
\alias{round_years}
\title{Rounds year columns from the Technologies table to nearest multiple of the
timestep used (set in the model parameters)}
\usage{
round_years(data)
}
\arguments{
\item{data}{list of dataframes read in by \code{read_excel_data_template()}}
}
\value{
list of data tables as provided in inputs, with the changes outlined
made to the Technologies table.
}
\description{
Rounds lifetime and technology start years columns to the nearest timestep year
(e.g. nearest 5 year period when timestep is 5).
}
