% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_combining_constraints.R
\name{run_constraint_function}
\alias{run_constraint_function}
\title{Calls a specific constraint functions to generate a constraint list before
converting to a matrix}
\usage{
run_constraint_function(constraint_name, data, decision_variables, in_app)
}
\arguments{
\item{constraint_name, }{str name of function to be run, usually passed from
an item of \code{\link[=get_constraint_functions]{get_constraint_functions()}}}

\item{data}{list of dataframes - raw_data after initial processing from
\code{\link[=process_sites]{process_sites()}}}

\item{decision_variables}{dataframe of decision variables as produced by
\code{\link[=create_decision_variables]{create_decision_variables()}}}
}
\value{
sparseMatrix for the constraint named, see
\code{\link[=get_constraint_matrix]{get_constraint_matrix()}} for more details.
}
\description{
Calls the function supplied and gives data and decision_variables as the
arguments. The subsequent list returned from the function is then passed
to get_constraint_matrix to convert into matrix format.
}
