% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_process_sites.R
\name{select_cols_for_sites_data}
\alias{select_cols_for_sites_data}
\title{Helper function - avoids repeating the same column select for small and large
point sites.}
\usage{
select_cols_for_sites_data(df)
}
\description{
Helper function - avoids repeating the same column select for small and large
point sites.
}
