% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_comit_counterfactual_solver.R
\name{set_all_vals_to_0}
\alias{set_all_vals_to_0}
\title{Set all values apart from year column to 0}
\usage{
set_all_vals_to_0(df)
}
\arguments{
\item{df, }{a single dataframe containing a year column}
}
\value{
dataframe with all values set to 0 for all columns apart from 'year'.
}
\description{
Set all values apart from year column to 0
}
