% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_utilities.R
\name{show_spinners}
\alias{show_spinners}
\title{Show all plot spinners in the UI}
\usage{
show_spinners()
}
\value{
Invisibly, after calling UI spinners
}
\description{
Convenience wrapper for app's spinner function.
}
\keyword{internal}
