% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_sites.R
\name{site_demand}
\alias{site_demand}
\title{Calculate annual demand for each site}
\usage{
site_demand(data)
}
\arguments{
\item{data}{list of data tables as read in from excel template and having
gone through a series of early preprocessing steps as follows:

\if{html}{\out{<div class="sourceCode">}}\preformatted{data <- raw_data \%>\%
  process_sites()

data \%<>\%
  interpolate_data() \%>\%
  tidy() \%>\%
  round_years() \%>\%
  adjust_for_optimism() \%>\%
  apply_energy_efficiency() \%>\%
  adjust_existing_capacity() \%>\%
  add_site_ID()

# add region variable
data$NAEI_clean \%<>\% assign_site_region()
}\if{html}{\out{</div>}}}
}
\value{
dataframe for demand at each site. One row for each commodity
produced at each site in each year. Contains the following columns:
\itemize{
\item year
\item site_ID
\item output_commodity
\item scaling_factor_within_sector - factor used to multiply national demand
to produce site level demand
\item demand - site level demand of output commodity
}
}
\description{
Calculate annual demand for each site
}
