% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_run_model_for_all_inputs.R
\name{solved_check}
\alias{solved_check}
\title{Check if the model found a valid solution}
\usage{
solved_check(solved)
}
\arguments{
\item{solved, }{solved model object as returned by \code{comit_problem_solver()}.}
}
\value{
NULL, error is thrown if the model found no valid solution.
}
\description{
Check if the model found a valid solution
}
