% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_attribution.R
\name{sort_efficiencies}
\alias{sort_efficiencies}
\title{Tidy a resource or energy efficiency dataframe}
\usage{
sort_efficiencies(data, efficiency_type)
}
\arguments{
\item{data}{list of dataframes, generated from passing the filepath for an
input spreadsheet to \code{read_excel_data_template()} and then passing the output
to \code{tidy()}.}

\item{efficiency_type}{A string specifying the type of efficiency ('resource'
or 'energy').}
}
\value{
A dataframe with sorted efficiencies. One row for each sector and year
combination. The column 'efficiency' provides the mean efficiency for the
given sector and year.
}
\description{
This function sorts efficiencie dataframes based on the specified efficiency
type, providing the mean efficiency for each sector in each year.
}
