% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_read_data.R
\name{sort_retrofit}
\alias{sort_retrofit}
\title{Make adjustments to input data based on retrofit parameters}
\usage{
sort_retrofit(data)
}
\arguments{
\item{data}{list of dataframes read in by \code{read_excel_data_template()}}
}
\value{
data, list of input data tables as provided but with minor adjustments
made based on retrofit parameters.
}
\description{
Make adjustments to input data based on retrofit parameters
}
