% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comit_solver.R
\name{standard_data_read}
\alias{standard_data_read}
\title{Read in the standard version of the input data template}
\usage{
standard_data_read(testing = FALSE)
}
\arguments{
\item{testing}{boolean for whether the code is being ran as part of formal tests,
in order to account for filepath change.}
}
\value{
list of all input data tables
}
\description{
This allows for quick running whilst developing and isn't exported as part
of the package.
}
