% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_emissions.R
\name{sum_emission_types}
\alias{sum_emission_types}
\title{Aggregate each of the emissions type to the technology/year level}
\usage{
sum_emission_types(emissions_df)
}
\arguments{
\item{emissions_df}{dataframe, originally created by \code{create_emissions_df()}
and then having had all required types of emissions calculated by subsequent
functions: \code{calculate_CO2_direct_emissions},
\code{calculate_nonCO2_direct_emissions},
\code{calculate_CO2_indirect_emissions} and
\code{calculate_nonCO2_indirect_emissions}.}
}
\value{
dataframe, with 16 rows per technology and year (one for each possible
type of emissions). Columns include:
\itemize{
\item code
\item year
\item gas - type of emission: 'CO2' or 'nonCO2'
\item source - type of emission: 'process' or 'fuel'
\item capture - type of emission: boolean
\item location - type of emission: 'direct' or 'indirect'
\item CO2e - numeric, amount of emissions per unit of capacity in kilotonnes of
of CO2e.
}
}
\description{
Sums each type of emissions for each technology in each year. This is done
because until this point, emissions are produced at the commodity level and
there are often many commodities to a single technology.
}
