% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_process_sites.R
\name{sum_emissions_by_sector}
\alias{sum_emissions_by_sector}
\title{Get total traded and non-traded emissions in each sector}
\usage{
sum_emissions_by_sector(df)
}
\arguments{
\item{df}{dataframe containing columns for IPM_sector, traded_flag and
emissions_MtCO2.}
}
\value{
dataframe with one row per sector and traded status combination and
the following additional columns:
\itemize{
\item sector_emissions - total amount of emissions for each sector by traded and
non-traded status (emissions per row)
\item total_point_sites_emissions_in_sector -  the total amount of emissions for
the sector overall (not split  by traded and non-traded).
}
}
\description{
Helper function to \code{\link[=get_sector_emission_totals]{get_sector_emission_totals()}}
}
