% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_pv_sum_coefficients.R
\name{sum_PV_coefficients}
\alias{sum_PV_coefficients}
\title{Function to combine multiple PV dataframes and do a rowwise sum}
\usage{
sum_PV_coefficients(PV_dataframes, n_decision_variables)
}
\arguments{
\item{PV_dataframes}{A list of two column dataframes. The first column should be a variable index
The second column should be the coefficient}

\item{n_decision_varaibles}{The number of decision variables in the optimisation. This is usually
the number of rows of the decision variables object}
}
\description{
Function to combine multiple PV dataframes and do a rowwise sum
}
