% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_outputs.R
\name{summarise_capture}
\alias{summarise_capture}
\title{Summarise CO2 captured (MtCO2) by year}
\usage{
summarise_capture(data, filter_options)
}
\arguments{
\item{data}{Data frame of outputs}

\item{filter_options}{named list for the values to keep for different variables.
Should contain the following elements:
\itemize{
\item "sector" Character, selected sector or 'All'
\item "technology" Character, selected technology category or 'All'
\item "this_cluster" Character, selected cluster or 'All'
\item "cluster_category" Character vector of cluster radii/categories to keep
\item "models_to_present" Character vector of run names to include
}}
}
\value{
Tibble with \code{total_emissions} (captured)
}
\description{
As per \code{\link[=summarise_emissions]{summarise_emissions()}} but for the \code{Captured} emissions category.
}
