% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_attribution.R
\name{summarise_emissions_attr}
\alias{summarise_emissions_attr}
\title{Get Total Emissions by Year and Sector}
\usage{
summarise_emissions_attr(df, filter_type, new_name)
}
\arguments{
\item{df}{A data frame containing emissions data.}

\item{filter_type}{A value specifying the type of emissions to filter by.}

\item{new_name}{A string specifying the new column name for the summarised
emissions.}
}
\value{
A data frame with summarised emissions by year and sector.
}
\description{
Get Total Emissions by Year and Sector
}
