% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_process_sites.R
\name{summarise_sites_by}
\alias{summarise_sites_by}
\title{Get total number of emissions and total number of sites by a given grouping
variable. E.g. by passing the data for large point sites, and a sector
variable for grouping, the total emissions and a total number of sites by
sector are calculated.}
\usage{
summarise_sites_by(group, df, include_pipe_dist = FALSE)
}
\description{
Get total number of emissions and total number of sites by a given grouping
variable. E.g. by passing the data for large point sites, and a sector
variable for grouping, the total emissions and a total number of sites by
sector are calculated.
}
