% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_interpolate_data.R
\name{tables_to_interpolate}
\alias{tables_to_interpolate}
\title{Get names of time series tables to interpolate}
\usage{
tables_to_interpolate()
}
\value{
vector of characters, the names of tables that are to be interpolated.
}
\description{
If any timeseries tables are added to the input spreadsheet, add the tab name
to the list below to allow for additional interpolation.
}
