% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_cost_tables.R
\name{technology_cost_capex_adjustment}
\alias{technology_cost_capex_adjustment}
\title{Adjusts the capex costs for new technological capacity
to un-discounted yearly spread costs}
\usage{
technology_cost_capex_adjustment(cost_data, model_data)
}
\arguments{
\item{cost_data}{a table containing the technology costs}

\item{model_data}{list of data tables read in from excel data template}
}
\value{
table of capex_costs for each site, technology and year
combination (4 columns)
}
\description{
Adjusts the capex costs for new technological capacity
to un-discounted yearly spread costs
}
