% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_read_data.R
\name{tidy}
\alias{tidy}
\title{Tidy the input data into a more usable format}
\usage{
tidy(data)
}
\arguments{
\item{data}{list of dataframes read in by \code{read_excel_data_template()}}
}
\value{
list of dataframes as provided as input, but with basic tidying
completed on some dataframes.
}
\description{
Basic wrangling of the input dataframes to facilitate further processing of
the data later. This function pivots many of the dataframes into longer
tables, as well as doing some specific transformations such as tidying the
cluster connection tables and making required adjustments to retrofit data.
}
