% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_read_data.R
\name{tidy_cluster_connections}
\alias{tidy_cluster_connections}
\title{Modify the data table for cluster connection availability}
\usage{
tidy_cluster_connections(data)
}
\arguments{
\item{data}{list of dataframes read in by \code{read_excel_data_template()}}
}
\value{
dataframe with row per possible cluster connection. Columns for:
\itemize{
\item cluster_1 - character
\item cluster_2 - character
\item allowed_route - boolean
}
}
\description{
Modify the data table for cluster connection availability
}
