% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_read_data.R
\name{tidy_fuel_constraints}
\alias{tidy_fuel_constraints}
\title{Performing some additional specific tidying on fuel constraints}
\usage{
tidy_fuel_constraints(data, table, min_or_max)
}
\arguments{
\item{data}{list of dataframes read in by \code{read_excel_data_template()}}

\item{table, }{string for the name of the sheet that is to be tidied}

\item{min_or_max, }{string for whether the constraint uses 'min' or 'max' values}
}
\value{
dataframe, the tidied table for the fuel constraint, with one row per
year and fuel combination from the inputs. Columns for:
\itemize{
\item year (numeric)
\item fuel (character)
\item group (character)
\item apply_to_industry_only (boolean)
\item 'min' or 'max' depending on value of 'min_or_max' argument (numeric)
}
}
\description{
Performing some additional specific tidying on fuel constraints
}
