% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_process_sites.R
\name{tidy_non_point_sites}
\alias{tidy_non_point_sites}
\title{Helper function to get non-point sites data in the correct format}
\usage{
tidy_non_point_sites(df)
}
\description{
Helper function to get non-point sites data in the correct format
}
