% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_read_data.R
\name{tidy_parameters}
\alias{tidy_parameters}
\title{Get a single wide dataframe for main model parameters}
\usage{
tidy_parameters(data)
}
\value{
list of dataframes, as input but with a single parameter tab called
'model_parameters' instead of the previous two tabs.
}
\description{
If two model parameter tables exist (model_parameters_a and
model_parameters_b), combine them into a single wide dataframe for easier
calling of parameters.
}
