% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_process_sites.R
\name{tidy_sites_data}
\alias{tidy_sites_data}
\title{Final tidying of the site data}
\usage{
tidy_sites_data(df, input_data)
}
\arguments{
\item{df}{dataframe containing all sites to be modeled (usually the combination
of large point source, small point source and non-point source sites), with
the following columns:
\itemize{
\item site_name
\item IPM_sector
\item H2_point
\item total_MtCO2 (emissions)
\item pipe_dist
\item num_sites
\item Latitude
\item Longitude
\item traded_flag
}}

\item{input_data}{data read in from excel template, same as what is passed to
\code{\link[=comit_solver]{comit_solver()}}.}
}
\value{
reordered input dataframe with four new columns:
\itemize{
\item H2_first_year
\item CCS_first_year
\item in_cluster_H2 (whether site is within the H2 cluster radius)
\item in_cluster_CCS (whether site is within the CCS cluster radius)
}
}
\description{
Adds variables for whether a site is within the H2 or CCS clusters, and
arranges the sites in alphabetical order and then by Latitude.
}
