% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F-2-classify_helpers.R
\name{translateMutationForward}
\alias{translateMutationForward}
\title{Translate a Seq Change String to a Primer Template Mismatch for Forward primers}
\usage{
translateMutationForward(mutString)
}
\arguments{
\item{mutString}{String describing a sequence change in the format Original->New i.e. "A->G"}
}
\value{
Primer Template mismatch in the format Primer/Template i.e. "A/C"
}
\description{
Given a mutation string, return a mutation string with the template nucleotide translated (because forward primer)
}
\examples{
translateMutationForward("A->G")

}
