% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F-2-classify_helpers.R
\name{translateMutationReverse}
\alias{translateMutationReverse}
\title{Translate a Seq Change String to a Primer Template Mismatch for Reverse primers}
\usage{
translateMutationReverse(mutString)
}
\arguments{
\item{mutString}{String describing a sequence change in the format Original->New i.e. "A->G"}
}
\value{
Primer Template mismatch in the format Primer/Template i.e. "T/G"
}
\description{
Given a mutation string, return a mutation string with the primer nucleotide translated (because reverse primer)
}
\examples{
translateMutationReverse("A->G")

}
