% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_cost_tables.R
\name{undiscounted_yearly_cost}
\alias{undiscounted_yearly_cost}
\title{Un-discounts opex / carbon cost / fuel cost from 5 year aggregate into yearly
costs}
\usage{
undiscounted_yearly_cost(cost_data, model_data, new_capacity_type)
}
\arguments{
\item{cost_data}{a table containing the cost data to be undiscounted}

\item{model_data}{list of data tables read in from excel data template}

\item{new_capacity_type}{the new_capacity variable type to filter (pipe or tech)}
}
\value{
table of cost data with the following two columns added:
\itemize{
\item year_cost: the capex cost split into yearly payments over the loan period
\item end_year_loan: the final loan year
}
}
\description{
Un-discounts opex / carbon cost / fuel cost from 5 year aggregate into yearly
costs
}
