% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_cost_tables.R
\name{yearly_cost_calc}
\alias{yearly_cost_calc}
\title{Converts technology capex costs into yearly payments to be made over the loan
period (technology lifetime) using the PMT function}
\usage{
yearly_cost_calc(cost_data, model_data)
}
\arguments{
\item{cost_data}{a table containing the capex costs to calculate the yearly
payments for}

\item{model_data}{list of data tables read in from excel data template}
}
\value{
table of cost data with the following two columns added:
\itemize{
\item year_cost: the capex cost split into yearly payments over the loan period
\item end_year_loan: the final loan year
}
}
\description{
Converts technology capex costs into yearly payments to be made over the loan
period (technology lifetime) using the PMT function
}
