% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_comit_counterfactual_solver.R
\name{yearly_used_capacity}
\alias{yearly_used_capacity}
\title{Calculate Used Capacity For Existing Technologies For a Given Year}
\usage{
yearly_used_capacity(year, data)
}
\arguments{
\item{year, }{numeric}

\item{data}{list of dataframes containing input data, produced by
\code{comit_preprocess_inputs()}.}
}
\value{
dataframe with the required capacity for each technology.
}
\description{
Uses EEP rollout to get the required yearly capacities for existing technologies.
}
